### 2023-09-10 #5827 @editorialbot
[PRE REVIEW]: ParticleTracking: A GUI and library for particle tracking on stereo camera images
**Submitting author:** @a-niem (<a href="http://orcid.org/0009-0008-2025-1946">Adrian Niemann</a>)
**Repository:** https://github.com/ANP-Granular/ParticleTracking
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v0.6.0
