### 2023-09-10 #5827 @arfon
@a-niem – thanks for your submission to JOSS. It looks like there's an issue with the formatting of your paper, see this error:

> Problem with affiliations for Dmitry Puzyrev, perhaps the affiliations index need quoting? (Theoj::Error)
