### 2023-09-11 f33629f
Fix rod dataset correction
Adds the capability to 'correct' rods when none are present in the loaded
dataset for the current frame.
Fixes #73.
