+++ RodTracker/src/RodTracker/ui/mainwindow_layout.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'mainwindow_layout.ui'
 #
-# Created by: PyQt5 UI code generator 5.15.7
+# Created by: PyQt5 UI code generator 5.15.9
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic5 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -16,7 +16,11 @@ class Ui_MainWindow(object):
         MainWindow.setObjectName("MainWindow")
         MainWindow.resize(1190, 607)
         icon = QtGui.QIcon()
-        icon.addPixmap(QtGui.QPixmap("resources/icon_main.ico"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
+        icon.addPixmap(
+            QtGui.QPixmap("resources/icon_main.ico"),
+            QtGui.QIcon.Normal,
+            QtGui.QIcon.Off,
+        )
         MainWindow.setWindowIcon(icon)
         self.centralized = QtWidgets.QWidget(MainWindow)
         self.centralized.setObjectName("centralized")
@@ -30,9 +34,15 @@ class Ui_MainWindow(object):
         self.scrollArea.setWidgetResizable(True)
         self.scrollArea.setObjectName("scrollArea")
         self.scrollAreaWidgetContents_all = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_all.setGeometry(QtCore.QRect(0, 0, 1176, 622))
-        self.scrollAreaWidgetContents_all.setObjectName("scrollAreaWidgetContents_all")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_all)
+        self.scrollAreaWidgetContents_all.setGeometry(
+            QtCore.QRect(0, 0, 1176, 622)
+        )
+        self.scrollAreaWidgetContents_all.setObjectName(
+            "scrollAreaWidgetContents_all"
+        )
+        self.verticalLayout = QtWidgets.QVBoxLayout(
+            self.scrollAreaWidgetContents_all
+        )
         self.verticalLayout.setContentsMargins(0, 5, 0, 0)
         self.verticalLayout.setSpacing(0)
         self.verticalLayout.setObjectName("verticalLayout")
@@ -46,58 +56,92 @@ class Ui_MainWindow(object):
         self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
         self.line_2.setObjectName("line_2")
         self.grid_loads.addWidget(self.line_2, 1, 0, 1, 2)
-        self.pb_load_images = QtWidgets.QPushButton(self.scrollAreaWidgetContents_all)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
+        self.pb_load_images = QtWidgets.QPushButton(
+            self.scrollAreaWidgetContents_all
+        )
+        sizePolicy = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed
+        )
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pb_load_images.sizePolicy().hasHeightForWidth())
+        sizePolicy.setHeightForWidth(
+            self.pb_load_images.sizePolicy().hasHeightForWidth()
+        )
[…]
