### 2023-09-11 4afa074
Fix affiliation and add missing DOIs
