+++ paper/paper.bib
@@ -69,6 +69,7 @@
   year      = {1999},
   volume    = {83},
   pages     = {440--443},
+  doi       = {10.1103/physrevlett.83.440},
   publisher = {American Physical Society},
 }
 
@@ -79,6 +80,7 @@
   year    = {2006},
   pages   = {830-836},
   volume  = {74},
+	doi     = {10.1209/epl/i2005-10589-8},
 }
 
 @Article{Noirhomme2018,
@@ -96,7 +98,6 @@
   url       = {https://doi.org/10.1209%2F0295-5075%2F123%2F14003},
 }
 
-
 @Article{Harth2013,
   author    = {Harth, K. and Kornek, U. and Trittel, T. and Strachauer, U. and H\"ome, S. and Will, K. and Stannarius, R.},
   title     = {Granular Gases of Rod-Shaped Grains in Microgravity},
@@ -105,7 +106,8 @@
   volume    = {110},
   pages     = {144102},
   publisher = {American Physical Society},
-  }
+	doi       = {10.1103/PhysRevLett.110.144102},
+}
 
 @article{Harth2018,
   title = {Free Cooling of a Granular Gas of Rodlike Particles in Microgravity},
@@ -145,12 +147,13 @@
 }
 
 @article{Yu2020,
-Author = {P. Yu and M. Schr\"oter and M. Sperl},
-Title = {Velocity Distribution of a Homogeneously Cooling Granular Gas},
-Journal = {Phys. Rev. Lett.},
-Year = {2020}, 
-Volume = {124},
-Pages = {208007}
+  Author = {P. Yu and M. Schr\"oter and M. Sperl},
+  Title = {Velocity Distribution of a Homogeneously Cooling Granular Gas},
+  Journal = {Phys. Rev. Lett.},
+  Year = {2020},
+  Volume = {124},
+  Pages = {208007},
+  doi = {10.1103/PhysRevLett.124.208007},
 }
 
 @Article{Pitikaris2022,
@@ -198,4 +201,3 @@ url={https://doi.org/10.1007/s12217-020-09800-4}
   title = {Continuously heated granular gas of elongated particles},
   journal = {{EPJ} Web of Conferences}
 }
-
