+++ paper/paper.md
@@ -11,7 +11,7 @@ authors:
     affiliation: 1
   - name: Dmitry Puzyrev
     orcid: 0000-0002-9269-3224
-    affiliation: 1,2
+    affiliation: "1, 2"
   - name: Ralf Stannarius
     orcid: 0000-0002-6836-3663
     affiliation: 2
@@ -39,7 +39,7 @@ The main features of this software are given below with a more extensive feature
 - automated tracking of particles over multiple stereo camera frames, i.e. the course of an experiment
 - providing a GUI for applying manual corrections to the automatically generated data with a typical workflow shown in \autoref{fig:workflow}
 
-The main focus of this software is currently on elongated (rod-shaped) particles, but it is extensible with new particle geometries. 
+The main focus of this software is currently on elongated (rod-shaped) particles, but it is extensible with new particle geometries.
 The software can also be modified for inclusion of additional camera views for more accurate 3D tracking, or for 1-view 2D particle tracking.
 The `RodTracker` software is currently employed for data extraction in the German Aerospace Center (DLR) projects EVA (50WM2048), VICKI (50WM2252), and CORDYGA (50WM2242). Several publications that use this library for data extraction are currently in preparation.
 The prototype software for particle detection and tracking was described in [@Puzyrev2020].
@@ -48,20 +48,19 @@ The prototype software for particle detection and tracking was described in [@Pu
 
 # Statement of need
 
-Many natural and industrial processes deal with granular gases, i.e. dilute ensembles of macroscopic particles floating and colliding in space. One of the defining features of such systems is inelasticity of the collision, i.e. dissipation of particle kinetic energy. This leads to fascinating phenomena such as spontaneous clustering, absence of energy equipartition and non-Gaussian velocity distributions. 
-While most of 2D experiments can be performed in normal gravity, 3D experiments with granular gases require microgravity conditions. Starting from the pioneering results on cluster formation [@Falcon1999], 3D experiments have been reported for spherical grains [@Falcon2006; @Yu2020], ellipsoids [@Pitikaris2022] and rods [@Harth2013; @Harth2018]. 
+Many natural and industrial processes deal with granular gases, i.e. dilute ensembles of macroscopic particles floating and colliding in space. One of the defining features of such systems is inelasticity of the collision, i.e. dissipation of particle kinetic energy. This leads to fascinating phenomena such as spontaneous clustering, absence of energy equipartition and non-Gaussian velocity distributions.
+While most of 2D experiments can be performed in normal gravity, 3D experiments with granular gases require microgravity conditions. Starting from the pioneering results on cluster formation [@Falcon1999], 3D experiments have been reported for spherical grains [@Falcon2006; @Yu2020], ellipsoids [@Pitikaris2022] and rods [@Harth2013; @Harth2018].
 
-<!-- FIXME: The following paragraph might not be necessary. Instead focus on the intended experiments? -->
 In typical microgravity experiments, ensembles of particles are placed in the container, excited mechanically or magnetically and observed with a stereo-camera setup. Many experiments were performed in the VIP-Gran instrument by the Space Grains ESA Topical team (spacegrains.org) during parabolic flight campaigns. In the majority of VIP-Gran experiments, particle density does not allow for tracking individual grains.
 
 Another possibility is to perform the experiment with dilute ensembles, where most particles can be directly observed on video footage [@Harth2018; @Puzyrev2020]. In this case, the focus has been on experiments with elongated particles, due to the fact that collision rates for such particles are much higher than for spheres for the same packing fraction. Thus, even if particles overlap on the camera views, usually their endpoints still can be observed and their 3D positions and orientations can be reconstructed. In addition, study of elongated particles allows to observe the evolution of their orientations and to find the kinetic energy associated with the rotational degrees of freedom. Experiments with other particle types are planned as well.
 
[…]
