### 2023-09-15 53acaa2
Add dynamic versioning to RodTracker
