+++ RodTracker/pyproject.toml
@@ -1,12 +1,12 @@
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
+build-backend = "poetry_dynamic_versioning.backend"
 
 [tool.poetry]
 name = "RodTracker"
 description = "GUI to track rod-like particles on multiple cameras"
 readme = "README.md"
-version = "0.6.0"
+version = "0.0.0"
 authors = [
     "Adrian Niemann <adrian.niemann@gmx.de>",
     "Dmitry Puzyrev <dmitry.puzyrev@ovgu.de>"
@@ -53,13 +53,14 @@ docutils = {version = "==0.16", optional = true}
 Sphinx = {version = "==5.3", optional = true}
 sphinx_rtd_theme = {version = "^0.5.1", optional = true}
 myst_parser = {version = "^0.18.1", optional = true}
+dunamai = {version=">=1.18.0", optional = true}
 
 [tool.poetry.extras]
 dev = ["qt5-applications", "PyQt5-stubs", "flake8", "pytest", "pytest-qt",
        "pytest-cov", "pyinstaller", "Sphinx", "docutils", "sphinx_rtd_theme",
-       "myst_parser"]
+       "myst_parser", "dunamai"]
 build = ["pyinstaller", "Sphinx", "docutils", "sphinx_rtd_theme",
-         "myst_parser"]
+         "myst_parser", "dunamai"]
 test = ["pytest", "pytest-qt", "pytest-cov"]
 docs = ["Sphinx", "docutils", "sphinx_rtd_theme", "myst_parser"]
 
@@ -71,3 +72,9 @@ addopts = "--cov=src --cov-report=html"
 testpaths = [
     "tests/",
 ]
+
+[tool.poetry-dynamic-versioning]
+enable = true
+vcs = "git"
+style = "pep440"
+bump = true
