### 2023-09-16 ba90c77
Add support for Mac to ParticleDetection
