### 2023-09-18 cf6faf1
Add dynamic versioning to Win & Linux installers
Adds dynamic versioning for the Windows & Linux installer generation.
Additionally, adds file version info to the windows executable.
