### 2023-09-20 6646c67
Update icon/logo
Updates the used icon and splash screen.
