+++ RodTracker/build-files/create_logo.py
@@ -1,0 +1,201 @@
+from pathlib import Path
+import numpy as np
+import matplotlib.pyplot as plt
+import matplotlib.colors as mpl_colors
+from matplotlib import patches
+from matplotlib.transforms import Bbox
+from mpl_toolkits.mplot3d import art3d
+from PIL import Image
+
+
+if __name__ == "__main__":
+    output_path = Path(__file__).parent / "../src/RodTracker/resources"
+    my_colors = ["black", "blue", "green", "purple", "red", "yellow"]
+    my_colors = [mpl_colors.to_rgba(c, alpha=1.0) for c in my_colors]
+    my_colors_proj = [mpl_colors.to_rgba(c, alpha=0.5) for c in my_colors]
+    num_rod = 4
+    c_list = []
+    c_list_proj = []
+    for c in my_colors:
+        c_list.extend(num_rod * [c])
+    for c in my_colors_proj:
+        c_list_proj.extend(num_rod * [c])
+
+    # np.random.seed(1)
+    np.random.seed(5)
+    with_projection = True
+
+    rod_len = 0.2
+    x = np.random.rand(num_rod * len(my_colors), 2)
+    y = np.random.rand(num_rod * len(my_colors), 2)
+    z = np.random.rand(num_rod * len(my_colors), 2)
+    dx = np.diff(x, axis=1)
+    dy = np.diff(y, axis=1)
+    dz = np.diff(z, axis=1)
+
+    vec_lens = np.sqrt(dx**2 + dy**2 + dz**2)
+    # scale_fs = np.reshape(rod_len/vec_lens, 2*len(my_colors))
+    scale_fs = rod_len / vec_lens
+    x[:, 1] = x[:, 0] + np.reshape(dx * scale_fs, num_rod * len(my_colors))
+    y[:, 1] = y[:, 0] + np.reshape(dy * scale_fs, num_rod * len(my_colors))
+    z[:, 1] = z[:, 0] + np.reshape(dz * scale_fs, num_rod * len(my_colors))
+
+    fig = plt.figure()
+    ax = fig.add_subplot(projection="3d", zorder=1)
+    if with_projection:
+        a_off = 0.2
+        # Bottom plane
+        verts_bot = [
+            (-0.1, -0.1, -a_off),
+            (-0.1, 1.1, -a_off),
+            (1.1, 1.1, -a_off),
+            (1.1, -0.1, -a_off),
+        ]
+        # Top plane
+        verts_top = [
+            (-0.1, -0.1, 1.0 + a_off),
+            (-0.1, 1.1, 1.0 + a_off),
+            (1.1, 1.1, 1.0 + a_off),
+            (1.1, -0.1, 1.0 + a_off),
+        ]
+        # Left plane
+        verts_left = [
+            (-a_off, -0.1, 1.1),
+            (-a_off, 1.1, 1.1),
+            (-a_off, 1.1, -0.1),
+            (-a_off, -0.1, -0.1),
+        ]
+
+        ax.add_collection3d(
+            art3d.Poly3DCollection(
+                [verts_bot, verts_left],
+                color="gray",
+                alpha=0.15,
+                edgecolor="None",
+            )
+        )
+
+    # frame in background
+    ax.plot([0, 0], [0, 1], [0, 0], color="black")
+    ax.plot([0, 1], [1, 1], [0, 0], color="black")
+    ax.plot([0, 0], [1, 1], [0, 1], color="black")
+    # "rods"
+    for i in range(x.shape[0]):
+        ax.plot(x[i, :], y[i, :], z[i, :], linewidth=3, color=c_list[i])
+
+        if with_projection:
+            # projections
+            ax.plot(
[…]
