### 2023-09-21 88e3b8f
Add dynamic versioning to documentation
