+++ docs/source/developer_notes.md
@@ -19,7 +19,11 @@ Refer to the [Sphinx documentation](https://www.sphinx-doc.org/) for further opt
 ```{warning}
 On Windows make sure that no other process, e.g. Dropbox, attempts to access files necessary for the build during this. Otherwise an `OSError: [WinError 110]` might occur and break the process.
 
-On Linux this problem does not seem to occur.
+On Linux/macOS this problem does not seem to occur.
+```
+
+```{note}
+On macOS: Make sure that XCode is installed.
 ```
 
 ## RodTracker
@@ -28,7 +32,7 @@ The different tasks described below require additional libraries compared to the
 
 ### PyQt UI files
 Please make sure that the dependencies from the `DEV` extra are installed.
-The main GUI-layout is modeled in [mainwindow_layout.ui](../../RodTracker/src/RodTracker/ui/mainwindow_layout.ui) and can be changed using QtDesigner. 
+The main GUI-layout is modeled in [mainwindow_layout.ui](../../RodTracker/src/RodTracker/ui/mainwindow_layout.ui) and can be changed using QtDesigner.
 Generate the [Python file](../../RodTracker/src/RodTracker/ui/mainwindow_layout.py) after changing the [UI-File](../../RodTracker/src/RodTracker/ui/mainwindow_layout.ui) using:
 ```shell
 pyuic5 -o path/to/mainwindow_layout.py path/to/mainwindow_layout.ui
@@ -48,32 +52,13 @@ YOUR/REPO/PATH/RodTracker$ pytest
 Some tests require a display (see [pytest-qt](https://pytest-qt.readthedocs.io/en/latest/troubleshooting.html#tox-invocationerror-without-further-information))! Make sure the machine either has a screen connected or use [pytest-xvfb](https://pypi.org/project/pytest-xvfb/).
 ```
 
-### Building the docs
-Please make sure that the dependencies of the `DOCS`, `BUILD`, or `DEV` extra are installed.
-1. Clean the build directory:
-   ```shell
-   YOUR/REPO/PATH/docs$ make clean
-   ```
-2. (Re-)Build the documentation:
-   ```shell
-   YOUR/REPO/PATH/docs$ make html
-   ```
-
-Refer to the [Sphinx documentation](https://www.sphinx-doc.org/) for further options.
-
-```{warning}
-On Windows make sure that no other process, e.g. Dropbox, attempts to access files necessary for the build during this step. Otherwise an `OSError: [WinError 110]` might occur and break the process.
-
-On Linux this problem does not seem to occur.
-```
-
 ### Bundling the app
 
 Please make sure that the dependencies of the `BUILD`, or `DEV` extra are installed. On Linux there are additional requirements to be able to run the bundling script. Refer to the [PyInstaller documentation](https://pyinstaller.org/en/stable/requirements.html#gnu-linux) for that.
 
 There is a script to generate an executable for both Windows and Linux. Prior to bundling it will also (re-)build the documentation because this is supposed to be contained in the bundled app.
 
-**On Windows**: 
+**On Windows**:
 ```shell
 YOUR/REPO/PATH/RodTracker/build-files> build_app.bat
 ```
@@ -82,16 +67,16 @@ YOUR/REPO/PATH/RodTracker/build-files> build_app.bat
 YOUR/REPO/PATH/RodTracker/build-files$ build_app.sh
 ```
 
-There are two command line options `-onedir`(default)/`-onefile`. 
-Both will generate a `.\RodTracker\build` and a `.\RodTracker\dist` folder. 
+There are two command line options `-onedir`(default)/`-onefile`.
+Both will generate a `.\RodTracker\build` and a `.\RodTracker\dist` folder.
 The generated executables are located in the `.\RodTracker\dist` folder.
 
[…]
