### 2023-09-21 8834260
Remove support for 'onefile' bundling
