+++ RodTracker/build-files/build_app.sh
@@ -23,24 +23,5 @@ fi
 # Bundle the application
 cd $BASEDIR
 cd ..
-if [ $# -eq 0 ]; then
-    # build as one directory
-    pyinstaller --distpath ./dist/unix --workpath ./build/unix \
-      ./build-files/build_onedir.spec
-  else
-    if [ $1 == "-onefile" ]
-      then
-        # build as a single executable file
-        pyinstaller ./build-files/build_onefile.spec
-    fi
-    if [ $1 == "-onedir" ]
-      then
-        # build as one directory
-        pyinstaller --distpath ./dist/unix --workpath ./build/unix \
-          ./build-files/build_onedir.spec
-    fi
-    if [[ $1 != "-onedir" && $1 != "-onefile" ]]
-      then
-        echo "Unknown parameter...Use -onefile or -onedir"
-    fi
-fi
+pyinstaller --distpath ./dist/unix --workpath ./build/unix \
+  ./build-files/build_onedir.spec
