### 2023-09-21 4d74802
Add choice to open local/online documentation
