+++ RodTracker/src/RodTracker/ui/mainwindow_layout.py
@@ -16,11 +16,7 @@ class Ui_MainWindow(object):
         MainWindow.setObjectName("MainWindow")
         MainWindow.resize(1190, 607)
         icon = QtGui.QIcon()
-        icon.addPixmap(
-            QtGui.QPixmap("resources/icon_windows.ico"),
-            QtGui.QIcon.Normal,
-            QtGui.QIcon.Off,
-        )
+        icon.addPixmap(QtGui.QPixmap("resources/icon_windows.ico"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
         MainWindow.setWindowIcon(icon)
         self.centralized = QtWidgets.QWidget(MainWindow)
         self.centralized.setObjectName("centralized")
@@ -34,15 +30,9 @@ class Ui_MainWindow(object):
         self.scrollArea.setWidgetResizable(True)
         self.scrollArea.setObjectName("scrollArea")
         self.scrollAreaWidgetContents_all = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_all.setGeometry(
-            QtCore.QRect(0, 0, 1176, 622)
-        )
-        self.scrollAreaWidgetContents_all.setObjectName(
-            "scrollAreaWidgetContents_all"
-        )
-        self.verticalLayout = QtWidgets.QVBoxLayout(
-            self.scrollAreaWidgetContents_all
-        )
+        self.scrollAreaWidgetContents_all.setGeometry(QtCore.QRect(0, 0, 1190, 586))
+        self.scrollAreaWidgetContents_all.setObjectName("scrollAreaWidgetContents_all")
+        self.verticalLayout = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_all)
         self.verticalLayout.setContentsMargins(0, 5, 0, 0)
         self.verticalLayout.setSpacing(0)
         self.verticalLayout.setObjectName("verticalLayout")
@@ -56,92 +46,58 @@ class Ui_MainWindow(object):
         self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
         self.line_2.setObjectName("line_2")
         self.grid_loads.addWidget(self.line_2, 1, 0, 1, 2)
-        self.pb_load_images = QtWidgets.QPushButton(
-            self.scrollAreaWidgetContents_all
-        )
-        sizePolicy = QtWidgets.QSizePolicy(
-            QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed
-        )
+        self.pb_load_images = QtWidgets.QPushButton(self.scrollAreaWidgetContents_all)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(
-            self.pb_load_images.sizePolicy().hasHeightForWidth()
-        )
+        sizePolicy.setHeightForWidth(self.pb_load_images.sizePolicy().hasHeightForWidth())
         self.pb_load_images.setSizePolicy(sizePolicy)
         self.pb_load_images.setMaximumSize(QtCore.QSize(16777215, 16777215))
         self.pb_load_images.setObjectName("pb_load_images")
         self.grid_loads.addWidget(self.pb_load_images, 0, 0, 1, 1)
-        self.pb_load_rods = QtWidgets.QPushButton(
-            self.scrollAreaWidgetContents_all
-        )
[…]
