### 2023-09-29 c76f3c7
Update dependency constraints and project info
