### 2023-10-02 963ebc2
Fix warnings in tests
