+++ RodTracker/src/RodTracker/ui/view3d.py
@@ -23,13 +23,19 @@ import pandas as pd
 import matplotlib.colors as mpl_colors
 from PyQt5 import Qt3DCore, QtCore, QtWidgets, QtGui
 from PyQt5.Qt3DRender import QDirectionalLight, QCamera
-from PyQt5.Qt3DExtras import QPhongMaterial, \
-    QCylinderMesh, Qt3DWindow, QOrbitCameraController, QCuboidMesh, \
-    QExtrudedTextMesh, QPhongAlphaMaterial
+from PyQt5.Qt3DExtras import (
+    QPhongMaterial,
+    QCylinderMesh,
+    Qt3DWindow,
+    QOrbitCameraController,
+    QCuboidMesh,
+    QExtrudedTextMesh,
+    QPhongAlphaMaterial,
+)
 
-BOX_WIDTH = 112.
-BOX_HEIGHT = 80.
-BOX_DEPTH = 80.
+BOX_WIDTH = 112.0
+BOX_HEIGHT = 80.0
+BOX_DEPTH = 80.0
 ROD_RADIUS = 0.5
 _logger = logging.getLogger(__name__)
 
@@ -58,6 +64,7 @@ class View3D(QtWidgets.QWidget):
         Root entity for the 3D scene.
 
     """
+
     rods: List[Qt3DCore.QEntity] = []
     """List[QEntity] : Entities that resemble a rod each.
 
@@ -76,8 +83,13 @@ class View3D(QtWidgets.QWidget):
 
         # Create 3D scene
         self.scene = Qt3DCore.QEntity()
-        self.box, self.top, self.top_t, self.front, self.front_t = \
-            self.create_box(self.scene)
+        (
+            self.box,
+            self.top,
+            self.top_t,
+            self.front,
+            self.front_t,
+        ) = self.create_box(self.scene)
         self.fences = self.create_fences(self.scene)
 
         front_light = QDirectionalLight()
@@ -106,27 +118,15 @@ class View3D(QtWidgets.QWidget):
 
     def show_front(self):
         """Move the camera to display the front for the experiment box."""
-        self.camera.setViewCenter(
-            QtGui.QVector3D(0.0, 0.0, 0.0)
-        )
-        self.camera.setPosition(
-            QtGui.QVector3D(0.0, 0.0, 1.5 * BOX_WIDTH)
-        )
-        self.camera.setUpVector(
-            QtGui.QVector3D(0.0, 1.0, 0.0)
-        )
+        self.camera.setViewCenter(QtGui.QVector3D(0.0, 0.0, 0.0))
+        self.camera.setPosition(QtGui.QVector3D(0.0, 0.0, 1.5 * BOX_WIDTH))
+        self.camera.setUpVector(QtGui.QVector3D(0.0, 1.0, 0.0))
 
     def show_top(self):
         """Move the camera to display the top for the experiment box."""
-        self.camera.setViewCenter(
-            QtGui.QVector3D(0.0, 0.0, 0.0)
-        )
-        self.camera.setPosition(
-            QtGui.QVector3D(0.0, 1.5 * BOX_WIDTH, 0.0)
-        )
-        self.camera.setUpVector(
-            QtGui.QVector3D(0.0, 0.0, -1.0)
-        )
+        self.camera.setViewCenter(QtGui.QVector3D(0.0, 0.0, 0.0))
+        self.camera.setPosition(QtGui.QVector3D(0.0, 1.5 * BOX_WIDTH, 0.0))
+        self.camera.setUpVector(QtGui.QVector3D(0.0, 0.0, -1.0))
 
     @QtCore.pyqtSlot(pd.DataFrame)
     def update_rods(self, data: pd.DataFrame) -> None:
[…]
