### 2023-10-02 da5c440
Fix documentation build
Updates dependencies for building the documentation.
Adds a mocked import to fix broken API documentation.
