+++ RodTracker/pyproject.toml
@@ -52,10 +52,10 @@ flake8 = {version = ">=5", optional = true}
 pytest = {version = ">=7.1.2", optional = true}
 pytest-qt = {version = ">=4.0.0", optional = true}
 pytest-cov = {version = ">=3.0.0", optional = true}
-docutils = {version = "==0.16", optional = true}
-Sphinx = {version = "==5.3", optional = true}
-sphinx_rtd_theme = {version = "^0.5.1", optional = true}
-myst_parser = {version = "^0.18.1", optional = true}
+docutils = {version = ">=0.18", optional = true}
+Sphinx = {version = ">=7.2", optional = true}
+sphinx_rtd_theme = {version = "^1.3", optional = true}
+myst_parser = {version = "^2", optional = true}
 dunamai = {version=">=1.18.0", optional = true}
 Pillow = {version=">=10", optional = true}
 
