### 2023-10-06 3b87967
Update installation process & dependencies
