+++ ParticleDetection/src/ParticleDetection/utils/datasets.py
@@ -27,6 +27,7 @@ import logging
 from typing import List, Set, Dict, TypedDict, Tuple
 from pathlib import Path
 from dataclasses import dataclass
+import warnings
 import cv2
 import numpy as np
 import pandas as pd
@@ -36,28 +37,59 @@ _logger = logging.getLogger(__name__)
 
 
 DEFAULT_CLASSES = {
-    0: 'blue', 1: 'green', 2: 'orange', 3: 'purple', 4: 'red',
-    5: 'yellow', 6: 'black', 7: 'lilac', 8: 'brown'
+    0: "blue",
+    1: "green",
+    2: "orange",
+    3: "purple",
+    4: "red",
+    5: "yellow",
+    6: "black",
+    7: "lilac",
+    8: "brown",
 }
 """Class-color correspondences most commonly used by the trained networks."""
 
-DEFAULT_COLUMNS = ['x1', 'y1', 'z1', 'x2', 'y2', 'z2', 'x', 'y', 'z', 'l',
-                   'x1_{id1:s}', 'y1_{id1:s}', 'x2_{id1:s}', 'y2_{id1:s}',
-                   'x1_{id2:s}', 'y1_{id2:s}', 'x2_{id2:s}', 'y2_{id2:s}',
-                   'frame', 'seen_{id1:s}', 'seen_{id2:s}', 'color']
+DEFAULT_COLUMNS = [
+    "x1",
+    "y1",
+    "z1",
+    "x2",
+    "y2",
+    "z2",
+    "x",
+    "y",
+    "z",
+    "l",
+    "x1_{id1:s}",
+    "y1_{id1:s}",
+    "x2_{id1:s}",
+    "y2_{id1:s}",
+    "x1_{id2:s}",
+    "y1_{id2:s}",
+    "x2_{id2:s}",
+    "y2_{id2:s}",
+    "frame",
+    "seen_{id1:s}",
+    "seen_{id2:s}",
+    "color",
+]
 """Columns of rod position datasets used, e.g. in the RodTracker app."""
 
 RNG_SEED = 1
 """Seed to allow reproducibility of results, that are dependent on the
 generation of random numbers."""
 
-DetectionResult = TypedDict("DetectionResult",
-                            {'pred_boxes': torch.Tensor,
-                             'pred_classes': torch.Tensor,
-                             'pred_masks': torch.Tensor,
-                             'scored': torch.Tensor,
-                             'input_size': List[int],
-                             }, total=False)
+DetectionResult = TypedDict(
+    "DetectionResult",
+    {
+        "pred_boxes": torch.Tensor,
+        "pred_classes": torch.Tensor,
+        "pred_masks": torch.Tensor,
+        "scored": torch.Tensor,
+        "input_size": List[int],
+    },
+    total=False,
+)
 """Results of detecting particles in an image file.
 
 See also
@@ -69,6 +101,7 @@ See also
 
 class DataSet:
     """Representation of a dataset for training a network."""
+
     folder: str
     annotation: str
     name: str
@@ -82,8 +115,10 @@ class DataSet:
         with open(self.annotation) as metadata:
             annotations = json.load(metadata)
         return iter(
-            os.path.join(self.folder, anno["filename"]) for
-            anno in annotations.values() if anno["regions"])
+            os.path.join(self.folder, anno["filename"])
+            for anno in annotations.values()
+            if anno["regions"]
+        )
 
     def __len__(self):
         return get_dataset_size(self)
@@ -92,6 +127,7 @@ class DataSet:
 @dataclass
 class DataGroup:
     """Collection of training and test set for training a network."""
+
     train: DataSet
     val: DataSet
 
[…]
