+++ ParticleDetection/pyproject.toml
@@ -29,67 +29,146 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-scipy = ">=1.9.3"
-numpy = ">=1.23.4"
-matplotlib = ">=3.6.1"
-opencv-python-headless = ">=4.6.0.66"
+python = ">=3.8,<3.12"
+matplotlib = [
+    {version = ">=3.8", python = ">=3.9"},
+    {version = ">=3.6.1,<3.8", python = "<3.9"}
+]
+networkx = ">=3.1"
+numpy = [
+    {version = ">=1.23.4", python = ">=3.9"},
+    {version = ">=1.21,<=1.25", python = "<3.9"}
+]
+opencv-python-headless = ">=4.8.1.78"
+pandas= [
+    {version = ">=2.1.1", python = ">=3.9"},
+    {version = "<2.1", python = "<3.9"}
+]
+pulp = ">=2.7.0"
 tqdm = ">=4.64.1"
-networkx = ">=2.8.7"
-pandas = ">=2.1.1"
-pulp = ">=2.6.0"
-scikit-learn = ">=1.1.3"
-scikit-image = ">=0.21.0"
-trackpy = ">=0.5.0"
-tensorboard =  {version = "*", optional = true}
-protobuf = {version = "==3.20.1", optional = true}
+scikit-learn = ">=1.3.1"
+scikit-image = [
+    {version = ">=0.22.0", python = ">=3.9"},
+    {version = ">=0.21.0,<0.22", python = "<3.9"},
+]
+scipy = [
+    {version=">=1.9.3", python = ">=3.9"},
+    {version="<1.11.0", python = "<3.9"}
+]
 torch = [
-    {version = ">=2", markers = "(sys_platform == 'win32' or sys_platform == 'darwin') and (extra != 'detectron' and extra != 'gpu')"},   # default
-    {version = "*", source = "gpu-win", markers = "sys_platform == 'win32' and (extra == 'gpu' or extra == 'detectron')"},
-    {version = ">=1.8", source = "cpu-linux", markers = "sys_platform == 'linux' and (extra != 'detectron' and extra != 'gpu')"},    # default
-    {version = ">=1.8", markers = "sys_platform == 'linux' and (extra == 'detectron' or extra == 'gpu')"},
+    # Linux
+    {version = ">=2", platform = "linux", markers = "extra != 'cpu'"},
+    # The commented version below is the intended one, that does not work with
+    # pip. The used version specifies the URLs manually which is not desirable.
+    # {version = ">=2", platform = "linux", source = "pytorch_cpu", markers = "extra == 'cpu'"},
+    {python = "==3.8", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp38-cp38-linux_x86_64.whl"},
+    {python = "==3.9", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp39-cp39-linux_x86_64.whl"},
+    {python = "==3.10", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl"},
+    {python = "==3.11", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp311-cp311-linux_x86_64.whl"},
[…]
