### 2023-10-06 590d14b
Update dependencies
All versions of python now require the PyPI version of
importlib-resources.
The new version of PyInstaller required changes in the build-relevant
files.
