+++ RodTracker/src/RodTracker/RodTracker.py
@@ -14,28 +14,11 @@
 #  You should have received a copy of the GNU General Public License
 #  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
 
+import inspect
 import pathlib
 import sys
-import inspect
 
-if sys.version_info < (3, 9):
-    # importlib.resources either doesn't exist or lacks the files()
-    # function, so use the PyPI version:
-    import importlib_resources
-
-    importlib_resources.path = lambda module, file: importlib_resources.files(
-        module
-    ).joinpath(file)
-else:
-    # importlib.resources has files(), so use that:
-    import importlib.resources as importlib_resources
-
-    if sys.version_info >= (3, 11):
-        importlib_resources.path = (
-            lambda module, file: importlib_resources.files(module).joinpath(
-                file
-            )
-        )
+import importlib_resources
 from PyQt5 import QtWidgets, QtGui, QtCore
 
 
@@ -48,7 +31,11 @@ def main():
 
     app = QtWidgets.QApplication(sys.argv)
     pixmap = QtGui.QPixmap(
-        str(importlib_resources.path("RodTracker.resources", "splash.png"))
+        str(
+            importlib_resources.files("RodTracker.resources").joinpath(
+                "splash.png"
+            )
+        )
     )
     align = QtCore.Qt.AlignBottom | QtCore.Qt.AlignHCenter
     color = QtGui.QColorConstants.White
