### 2023-10-25 #5827 @aquilesC
I would like to review this submission.
In the meantime, I would like to ask @a-niem if there is data (and a pre-trained model) available in order to test the software (didn't have time to go through the docs yet, so disregard if this is obvious)?
