### 2023-10-26 #5986 @jgostick
Hi @mhubii and @aquilesC, the review is now underway. Thanks to you both for donating your time and expertise to this endeavor.  (I often wonder how much the peer review system would cost if journals actually paid reviewers as consultants...)

The JOSS review process, if you're not familiar, is based around filling out a checklist (the instructions for generating your own are given at the top of this thread).  If you have changes to ask of the author(s) you can use the issue tracker on their repo, which is on github and is public.

Hopefully you can complete your reviewers in 2-3 week time frame, as this this submission has been sitting for a while already.

Happy coding!
