+++ RodTracker/src/RodTracker/ui/view3d.py
@@ -1,560 +1,0 @@
-#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
-#
-#  This file is part of RodTracker.
-#  RodTracker is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
-#
-#  RodTracker is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
-#
-#  You should have received a copy of the GNU General Public License
-#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
-
-"""**TBD**"""
-
-import logging
-from typing import List, Tuple
-import numpy as np
-import pandas as pd
-import matplotlib.colors as mpl_colors
-from PyQt5 import Qt3DCore, QtCore, QtWidgets, QtGui
-from PyQt5.Qt3DRender import QDirectionalLight, QCamera
-from PyQt5.Qt3DExtras import (
-    QPhongMaterial,
-    QCylinderMesh,
-    Qt3DWindow,
-    QOrbitCameraController,
-    QCuboidMesh,
-    QExtrudedTextMesh,
-    QPhongAlphaMaterial,
-)
-
-BOX_WIDTH = 112.0
-BOX_HEIGHT = 80.0
-BOX_DEPTH = 80.0
-ROD_RADIUS = 0.5
-_logger = logging.getLogger(__name__)
-
-
-class View3D(QtWidgets.QWidget):
-    """A custom ``QWidget`` for display of 3D rod data.
-
-    Parameters
-    ----------
-    *args : iterable
-        Positional arguments for the ``QWidget`` superclass.
-    **kwargs : dict
-        Keyword arguments for the ``QWidget`` superclass.
-
-
-    .. admonition:: Slots
-
-        - :meth:`update_rods`
-        - :meth:`update_settings`
-
-    Attributes
-    ----------
-    view : Qt3DWindow
-    camera : QOrbitCameraController
-    scene : QEntity
-        Root entity for the 3D scene.
-
-    """
-
-    rods: List[Qt3DCore.QEntity] = []
-    """List[QEntity] : Entities that resemble a rod each.
-
-    Default is ``[]``.
-    """
-    _components: List[List[Qt3DCore.QEntity]] = []
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.view = Qt3DWindow()
-        self.container = self.createWindowContainer(self.view)
-
-        vboxlayout = QtWidgets.QHBoxLayout()
-        vboxlayout.addWidget(self.container)
-        self.setLayout(vboxlayout)
-
-        # Create 3D scene
-        self.scene = Qt3DCore.QEntity()
-        (
-            self.box,
-            self.top,
-            self.top_t,
-            self.front,
-            self.front_t,
-        ) = self.create_box(self.scene)
-        self.fences = self.create_fences(self.scene)
-
-        front_light = QDirectionalLight()
-        front_light.setWorldDirection(QtGui.QVector3D(0, 0, -1))
-        top_light = QDirectionalLight()
-        top_light.setWorldDirection(QtGui.QVector3D(0, -1, 0))
-        right_light = QDirectionalLight()
-        right_light.setWorldDirection(QtGui.QVector3D(1, 0, 0))
-        left_light = QDirectionalLight()
-        left_light.setWorldDirection(QtGui.QVector3D(-1, 0, 0))
[…]
