+++ RodTracker/src/RodTracker/ui/rod_tree.py
@@ -1,244 +1,0 @@
-# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev, and others
-#
-# This file is part of RodTracker.
-# RodTracker is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# RodTracker is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
-
-"""**TBD**"""
-from PyQt5 import QtCore, QtWidgets
-
-
-class RodTree(QtWidgets.QTreeWidget):
-    """A custom ``QTreeWidget`` to display all loaded rods as a tree.
-
-    Parameters
-    ----------
-    *args : iterable
-        Positional arguments for the ``QTreeWidget`` superclass.
-    **kwargs : dict
-        Keyword arguments for the ``QTreeWidget`` superclass.
-
-    Attributes
-    ----------
-    rod_info : None | Dict[Dict[Dict[list]]]
-        Holds the loaded information of the rod dataset about a rod being
-        ``'seen'`` or ``'unseen'``.\n
-        Dimensions: ``(frame, color, particle, camera)``
-
-    """
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.rod_info = None
-
-    @QtCore.pyqtSlot(dict, list)
-    def setup_tree(self, rod_info: dict, columns: list):
-        """Handles the setup of the treeview from extracted rod data.
-
-        Accepts a new ``rod_info`` dataset and (re-)sets the tree display of
-        it.
-
-        Parameters
-        ----------
-        rod_info : Dict[Dict[Dict[list]]]
-            Information of the rod dataset about a rod being ``'seen'`` or
-            ``'unseen'``.\n
-            Dimensions: ``(frame, color, particle, camera)``
-        columns : list
-            List of *camera* IDs on which a rod can be ``'seen'``/``'unseen'``.
-        """
-        self.rod_info = rod_info
-        self.clear()
-        self.setColumnCount(len(columns) + 1)
-        headers = [self.headerItem().text(0), *columns]
-        self.setHeaderLabels(headers)
-        self.generate_tree()
-
-    def generate_tree(self):
-        """(Re)generates the tree for display of loaded rod data."""
-        self.clear()
-        for frame in self.rod_info.keys():
-            current_frame = QtWidgets.QTreeWidgetItem(self)
-            current_frame.setText(0, f"Frame: {frame}")
-            for color in self.rod_info[frame].keys():
-                current_color = QtWidgets.QTreeWidgetItem(current_frame)
-                current_color.setText(0, color)
-                for particle in self.rod_info[frame][color].keys():
-                    current_particle = QtWidgets.QTreeWidgetItem(current_color)
-                    current_particle.setText(0, f"Rod{particle:3d}: ")
-                    for idx, gp in enumerate(
-                        self.rod_info[frame][color][particle]
-                    ):
-                        current_particle.setText(idx + 1, gp)
-                current_color.sortChildren(0, QtCore.Qt.AscendingOrder)
-            current_frame.sortChildren(0, QtCore.Qt.AscendingOrder)
-
-    @QtCore.pyqtSlot(dict)
-    def update_tree(self, new_data: dict):
[…]
