+++ RodTracker/src/RodTracker/ui/detection.py
@@ -1,708 +1,0 @@
-# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev, and others
-#
-# This file is part of RodTracker.
-# RodTracker is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# RodTracker is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
-
-"""**TBD**"""
-
-import logging
-import os
-from typing import Dict, List
-
-import pandas as pd
-import ParticleDetection.utils.datasets as ds
-import torch
-from PyQt5 import QtCore, QtGui, QtWidgets
-
-import RodTracker.backend.logger as lg
-import RodTracker.backend.parallelism as pl
-import RodTracker.ui.mainwindow_layout as mw_l
-from RodTracker import APPNAME, LOG_DIR, exception_logger
-from RodTracker.backend import detection
-from RodTracker.backend.detection import Detector, RodDetection
-from RodTracker.backend.img_data import ImageData
-
-# Don't remove the following imports, see GitHub issue as reference
-# https://github.com/pytorch/pytorch/issues/48932#issuecomment-803957396
-# isort: off
-import cv2  # noqa: F401
-import torchvision  # noqa: F401
-import ParticleDetection  # noqa: F401
-
-# isort: on
-
-_logger = logging.getLogger(__name__)
-
-
-def init_detection(ui: mw_l.Ui_MainWindow, image_managers: List[ImageData]):
-    """Initialize the functionality of detecting particles.
-
-    Parameters
-    ----------
-    ui : Ui_MainWindow
-        UI object of the main window of the application, i.e. also containing
-        the UI tab/objects for detection tasks.
-    image_managers : List[ImageData]
-        List of (relevant) image data management objects providing access to
-        loaded image datasets. Only images from these objects will be available
-        for particle detection.
-
-    Returns
-    -------
-    None | DetectorUI
-        Returns ``None``, if the system requirements for particle detections
-        are not met. Otherwise the ``DetectorUI`` object handling particle
-        detections is returned.
-    """
-    return DetectorUI(ui.tab_detection, image_managers)
-
-
-class DetectorUI(QtWidgets.QWidget):
-    """A custom ``QWidget`` to interface with a neural network for rod
-    detection.
-
-    Parameters
-    ----------
-    ui : QWidget
-        Widget containing the tab that is the GUI for the detection
-        functionality.
-    image_managers: List[ImageData]
-        List of (relevant) image data management objects providing access to
-        loaded image datasets.
-    *args : Iterable
-        Positional arguments for the ``QWidget`` superclass.
-    **kwargs : dict
-        Keyword arguments for the ``QWidget`` superclass.
-
-    Attributes
-    ----------
-    model : ScriptModule
-        Neural network model that is used for detection.\n
-        Default is ``None``.
-
-
-    .. admonition:: Signals
-
-        - :attr:`detected_data`
-
-    .. admonition:: Slots
-
-        - :meth:`images_loaded`
-        - :meth:`update_settings`
-    """
-
-    used_colors: List[str] = []
-    """List[str] : Colors of rods that are supposed to be detected.
-
-    Default is ``[]``.
-    """
-
-    number_rods: int = 1
-    """int : Expected number of rods per color in one frame.
-
-    Default is ``1``.
-    """
-
[…]

[…]
