+++ RodTracker/src/extensions/example_extension/requirements.txt
@@ -1,0 +1,1 @@
+# Python Requirements for this extension
