+++ RodTracker/src/extensions/example_extension/README.md
@@ -1,0 +1,7 @@
+This is an example on how to build an extension. It showcases the (major) 
+features of the extension mechanism of the RodTracker.
+
+Delete the `DEACTIVATED` file to let the RodTracker load this extension during
+its startup phase.
+(Re-)Create an empty file called `DEACTIVATED` to deactivate this or any other
+extension.
