### 2023-11-22 8675edc
Add an example extension
Adds an example how to build an extension. This also showcases upcoming
mechanisms of the extension system not yet implemented like the use of
example data and requirements that shall be installed automatically.
