+++ RodTracker/src/extensions/example_extension/example_data/README.md
@@ -1,0 +1,2 @@
+Put example data used for/by the extension in this folder. They will be used
+by upcoming iterations of the extension mechanism.
