### 2023-11-22 4f07288
Add missing data provider connection
Adds a missing signal connection for image interaction tabs to their
respective data provider objects.
