### 2023-11-22 608ffc5
Add extension for 'rods'
Adds the majority of the 'rods' extension.
WARNING: it is not yet usable
