+++ RodTracker/src/extensions/rods/__init__.py
@@ -1,0 +1,149 @@
+import logging
+from PyQt5 import QtWidgets, QtGui
+from RodTracker.ui.mainwindow import RodTrackWindow
+from RodTracker.ui.settings import IntSetting, ColorSetting, FloatSetting
+from RodTracker.backend.settings import Settings
+
+_logger = logging.getLogger(__name__)
+
+
+def setup(
+    _: QtWidgets.QSplashScreen = None,
+    main_window: RodTrackWindow = None,
+    *args,
+    **kwargs,
+):
+    try:
+        # load all required modules
+        from . import display_tab, data, rod_tree, rods
+    except ModuleNotFoundError as e:
+        _logger.error(f"Extension is missing a module: {e}")
+
+    # Insert image interaction tabs
+    front_view = display_tab.RodImageTab()
+    front_view.ID = "front_view"
+    main_window.add_image_interaction_tab(front_view, "Front View")
+
+    top_view = display_tab.RodImageTab()
+    top_view.ID = "top_view"
+    main_window.add_image_interaction_tab(top_view, "Top View")
+
+    # Add a position data provider
+    rod_data = data.RodData()
+    main_window.register_position_data(rod_data, [front_view, top_view])
+
+    # Insert utility tabs
+    tree_display = rod_tree.RodTree("Rods")
+    rod_data.data_update.connect(tree_display.update_tree)
+    rod_data.batch_update.connect(tree_display.batch_update_tree)
+    rod_data.seen_loaded.connect(tree_display.setup_tree)
+    tree_display.itemClicked.connect(
+        lambda: _logger.info(
+            "'RodTree.itemClicked()'-connection is not (yet) implemented."
+        )
+    )
+
+    # TODO: verify the selected class thing works, as this might be only run at
+    #   creation time of the outer lambda function
+    top_view.image_manager.next_img.connect(
+        lambda frame, _: tree_display.update_tree_folding(
+            frame, lambda: main_window.get_selected_class()
+        )
+    )
+    main_window.add_utility_tab(tree_display, "Rods")
+
+    # Insert settings
+    main_window.add_setting(
+        FloatSetting(
+            "Rods.rod_increment", 1, "Rod length in-/decrement [px]: "
+        )
+    )
+    main_window.add_setting(
+        IntSetting("Rods.rod_thickness", 3, "Rod Thickness [px]: ", 0, 15)
+    )
+    main_window.add_setting(
+        IntSetting("Rods.number_offset", 15, "Number Offset [px]: ", 0, 50)
+    )
+    main_window.add_setting(
+        IntSetting("Rods.number_size", 11, "Number Size [pt]: ", 0, 30)
+    )
+    main_window.add_setting(
+        FloatSetting("Rods.position_scaling", 1, "Position Scaling: ")
+    )
+    main_window.add_setting(
+        ColorSetting("Rods.rod_color", (0, 255, 255), "Rod Color: ")
+    )
+    main_window.add_setting(
+        ColorSetting("Rods.number_color", (224, 27, 36), "Number Color: ")
+    )
+    # Connect objects to settings
+    settings_obj = Settings()
+    settings_obj.setting_signals.setting_changed.connect(
+        front_view.update_settings
+    )
+    settings_obj.setting_signals.setting_changed.connect(
+        top_view.update_settings
+    )
+    settings_obj.setting_signals.setting_changed.connect(
+        tree_display.update_settings
+    )
+    settings_obj.setting_signals.setting_changed.connect(
+        rod_data.update_settings
+    )
[…]
