+++ RodTracker/src/extensions/rods/rod_tree.py
@@ -1,0 +1,252 @@
+#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+#
+#  This file is part of RodTracker.
+#  RodTracker is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  RodTracker is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+
+"""**TBD**"""
+from typing import Any
+from PyQt5 import QtCore, QtWidgets, QtGui
+from RodTracker.ui import tabs
+import RodTracker.backend.logger as lg
+
+
+class RodTree(QtWidgets.QTreeWidget, tabs.UtilityTab):
+    def __init__(self, id: str = "unknown", *args, **kwargs):
+        super().__init__(id=id, *args, **kwargs)
+        # Setup UI:
+        header_font = QtGui.QFont()
+        header_font.setBold(True)
+        header_font.setWeight(75)
+        self.headerItem().setFont(0, header_font)
+        self.headerItem().setText(0, "Loaded Rods:")
+        self.headerItem().setTextAlignment(0, QtCore.Qt.AlignCenter)
+        self.header().setDefaultSectionSize(150)
+        self.header().setMinimumSectionSize(125)
+
+        # Initialization
+        self.rod_info = None
+
+        # TODO: Connect signals
+
+    @QtCore.pyqtSlot(lg.Action)
+    def undo_action(self, _: lg.Action):
+        # TODO: evaluate, if something is necessary here
+        pass
+
+    @QtCore.pyqtSlot(dict, list)
+    def setup_tree(self, rod_info: dict, columns: list):
+        """Handles the setup of the treeview from extracted rod data.
+
+        Accepts a new ``rod_info`` dataset and (re-)sets the tree display of
+        it.
+
+        Parameters
+        ----------
+        rod_info : Dict[Dict[Dict[list]]]
+            Information of the rod dataset about a rod being ``'seen'`` or
+            ``'unseen'``.\n
+            Dimensions: ``(frame, color, particle, camera)``
+        columns : list
+            List of *camera* IDs on which a rod can be ``'seen'``/``'unseen'``.
+        """
+        self.rod_info = rod_info
+        self.clear()
+        self.setColumnCount(len(columns) + 1)
+        headers = [self.headerItem().text(0), *columns]
+        self.setHeaderLabels(headers)
+        self.generate_tree()
+
+    def generate_tree(self):
+        """(Re)generates the tree for display of loaded rod data."""
+        self.clear()
+        for frame in self.rod_info.keys():
+            current_frame = QtWidgets.QTreeWidgetItem(self)
+            current_frame.setText(0, f"Frame: {frame}")
+            for color in self.rod_info[frame].keys():
+                current_color = QtWidgets.QTreeWidgetItem(current_frame)
+                current_color.setText(0, color)
+                for particle in self.rod_info[frame][color].keys():
+                    current_particle = QtWidgets.QTreeWidgetItem(current_color)
+                    current_particle.setText(0, f"Rod{particle:3d}: ")
+                    for idx, gp in enumerate(
+                        self.rod_info[frame][color][particle]
+                    ):
+                        current_particle.setText(idx + 1, gp)
[…]
