+++ RodTracker/src/extensions/rods/rods.py
@@ -1,0 +1,538 @@
+#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+#
+#  This file is part of RodTracker.
+#  RodTracker is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  RodTracker is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+
+"""**TBD**"""
+
+from enum import Enum
+from typing import List
+import numpy as np
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class RodStyle(str, Enum):
+    """Styles for rod numbers."""
+
+    GENERAL = (
+        "background-color: transparent;"
+        "color: rgb({},{},{}); "
+        "font: {}px; font-weight: bold;"
+    )
+    SELECTED = (
+        "background-color: transparent;"
+        "color: white; "
+        "font: {}px; font-weight: bold;"
+    )
+    CONFLICT = (
+        "background-color: transparent;"
+        "color: red; "
+        "font: {}px; font-weight: bold;"
+    )
+    CHANGED = (
+        "background-color: transparent;"
+        "color: green; "
+        "font: {}px; font-weight: bold;"
+    )
+
+
+class RodState(Enum):
+    """States of a rod."""
+
+    NORMAL = 0
+    SELECTED = 1
+    EDITING = 2
+    CHANGED = 3
+    CONFLICT = 4
+
+
+class RodStateError(ValueError):
+    """Custom error that is raised when an unknown :class:`RodState` is
+    encountered."""
+
+    def __init__(self):
+        self.message = (
+            "The assigned RodState is invalid. Please assign a "
+            "known RodState."
+        )
+        super().__init__(self.message)
+
+
+class RodNumber(QtWidgets.QLineEdit):
+    """A custom ``QLineEdit`` to display rod numbers and have associated rods.
+
+    Parameters
+    ----------
+    color : str
+        The color of the rod that this widget represents.
+    parent : QWidget, optional
+        The widgets parent widget. Default is ``None``.
+    text : str, optional
+        The text displayed by the widget. Default is ``""``.
+    pos : QPoint, optional
+        The position of the widget (relative to its parent widget).
+        Default is ``QPoint(0, 0)``.
+
+
+    .. admonition:: Signals
+
+        - :attr:`activated`
+        - :attr:`dropped`
+        - :attr:`id_changed`
+        - :attr:`request_delete`
+
+    .. admonition:: Slots
+
+        - :meth:`update_settings`
+
+    Attributes
+    ----------
+    rod_points : List[int]
+        The starting and ending points of the rod in **UNSCALED** form:\n
+        ``[x1, y1, x2, y2]``
+    rod_center : np.array[float]
+        Center of the rod in **SCALED** form, i.e. adjusted to zoom levels and
+        offsets.
+    color : str
+        The color of the rod being represented.
+    """
+
+    activated = QtCore.pyqtSignal(int, name="activated")
+    """pyqtSignal(int) : Notifies, that this rod has been activated."""
+
+    dropped = QtCore.pyqtSignal(QtCore.QPoint, name="dropped")
[…]
