+++ RodTracker/src/extensions/rods/actions.py
@@ -1,0 +1,792 @@
+#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+#
+#  This file is part of RodTracker.
+#  RodTracker is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  RodTracker is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+
+"""**TBD**"""
+
+
+from enum import Enum, auto
+import logging
+from typing import Optional, Iterable, Union, List
+
+import numpy as np
+
+from RodTracker.backend.logger import Action
+from . import rods
+
+_logger = logging.getLogger(__name__)
+
+
+class NumberChangeActions(Enum):
+    """Helper class holding valid kinds of rod number changes.
+
+    Attributes
+    ----------
+    ALL : int
+        Indicates a switch of rod numbers in all cameras from the current frame
+        to the last frame of the dataset.
+    ALL_ONE_CAM : int
+        Indicates a switch of rod numbers in the currently displayed camera
+        from the current frame to the last frame of the dataset.
+    ONE_BOTH_CAMS : int
+        Indicates a switch of rod numbers in all cameras for the current frame
+        only.
+    CURRENT : int
+        Indicates a switch of rod numbers in the current camera only and the
+        current frame only.
+    """
+
+    ALL = auto()
+    ALL_ONE_CAM = auto()
+    ONE_BOTH_CAMS = auto()
+    CURRENT = auto()
+
+
+class ChangedRodNumberAction(Action):
+    """Class to represent a change of the rod number as a loggable action.
+
+    Parameters
+    ----------
+    rod : RodNumberWidget
+        A copy of the rod whose number is changed.
+    new_id : int
+        The new rod number of the changed rod.
+    coupled_action : Action, optional
+        The instance of an :class:`Action` that is performed at the same time
+        with this and must be reverted as well, if this :class:`Action` is
+        reverted. For example when the numbers of two rods are exchanged.
+        (Default is None)
+    *args : iterable
+        Positional arguments for the ``QListWidgetItem`` superclass.
+    **kwargs : dict
+        Keyword arguments for the ``QListWidgetItem`` superclass.
+
+    Attributes
+    ----------
+    rod : RodNumberWidget
+        A copy of the rod whose number is changed.
+    new_id : int
+        The new rod number of the changed rod.
+    action : str
+        Description of what kind of action was performed.
+        (Default is "Changed rod")
+    coupled_action : Union[Action, None]
+        The instance of an :class:`Action` that is performed at the same time
+        with this and must be reverted as well, if this :class:`Action` is
+        reverted.
+    """
+
+    def __init__(
+        self,
+        old_rod: rods.RodNumber,
+        new_id: int,
+        coupled_action: Action = None,
+        *args,
+        **kwargs,
+    ):
+        self.rod = old_rod
+        self.new_id = new_id
+        self.action = "Changed rod"
+        self.coupled_action = coupled_action
+        super().__init__(str(self), *args, **kwargs)
+
+    @property
+    def inverted(self):
+        rod = self.rod.copy()
+        rod.rod_id = self.new_id
+        inverted = ChangedRodNumberAction(rod, self.rod.rod_id)
+        inverted.parent_id = self.parent_id
+        inverted.frame = self.frame
+        return inverted
+
+    def __str__(self):
[…]

[…]
