### 2023-11-23 ANP-Granular/ParticleTracking#74 @aquilesC
[JOSS Review] Install missing dependencies and error hidden in logs
**Describe the bug**
Following the installation steps in the documentation, the program does not run because some dependencies are missing.
The first one I encountered was PyTorch, but there are others missing.

This happened after install RodTracker, and persisted after installing ParticleDetection.

The error is logged but hidden, so it took me a while to find out what was going on. It would be convinient, perhaps, to find another way of dealing with these type of exceptions.

JOSS Issue: https://github.com/openjournals/joss-reviews/issues/5986
