### 2023-11-23 #5986 @aquilesC
Some issues along the way:

- [ ]: The install misses some dependencies ([link](https://github.com/openjournals/joss-reviews/issues/5986))
