### 2023-11-23 ANP-Granular/ParticleTracking#75 @aquilesC
[JOSS Review] Couldn't find the trained RCNN for the example data
**Describe the bug**
The documentation states that "With this app comes a trained RCNN for detecting rods from the example dataset" but couldn't find it.

A potential improvement: Instead of leaving the example data within the software, move it outside, to the root folder, or perhaps make it avilable independently on Zenodo.

JOSS Issue: https://github.com/openjournals/joss-reviews/issues/5986
