### 2023-11-28 ANP-Granular/ParticleTracking#74 @dmitrypuzyrev
Dear @aquilesC,
I have notified our main developer, @a-niem , about the issues. He will be available from this Thursday, 30.11, and reply you as soon as possible. Meanwhile, could you please provide the additional information:

1. From which branch did you install the packages
2. What system/OS do you have
3. It would be important that you provide the logfile which you have obtained

Many thanks!
