### 2023-12-04 #5986 @jgostick
Shutting down for the northern hemisphere winter solstice to hibernate with friends and family is my favorite time of the year, I don't blame them. Let's see if they drop in here with a response to my question about timeline.
