### 2023-12-08 ANP-Granular/ParticleTracking#74 @dmitrypuzyrev
Dear @aquilesC,

the installation problems are already resolved in the `develop` branch.
There, the installation process should go without problems, if it is done from the fresh virtual environment created with `venv`.
We will update the `joss-paper` branch and software documentation accordingly in the beginning of the next week.

Thank you for pointing our the bug.
