### 2023-12-14 9ea2286
Update pyproject.toml
Fixed the ParticleDetection "rev" version for correct torchvision linux installation
