### 2023-12-14 ANP-Granular/ParticleTracking#74 @dmitrypuzyrev
Dear @aquilesC,

the `master` as well as `joss-paper` branches are updated with the latest program version.
The installation problems you've encountered should have been resolved and the documentation updated.
Please let us know if there are any other issues.
