### 2024-01-23 ANP-Granular/ParticleTracking#79 @mhubii
[JOSS Review] Install instructions missing
Install instruction missing at https://particletracking.readthedocs.io/en/latest/installation/rodtracker.html#installation-options

Should be installable as in https://github.com/ANP-Granular/ParticleTracking/tree/master/ParticleDetection readme.

Consider linking this readme to doc anyways, refer markdown sphinx https://www.sphinx-doc.org/en/master/usage/markdown.html

Also refers to #78
