### 2024-01-23 ANP-Granular/ParticleTracking#80 @mhubii
[JOSS Review] Sort imports re-factor
Hope this is not too OCD, but consider sorting imports, using e.g. https://pycqa.github.io/isort/, which has e.g. a visual studio code plugin https://marketplace.visualstudio.com/items?itemName=ms-python.isort

See e.g. this file https://github.com/ANP-Granular/ParticleTracking/blob/c946deafaffc8283fe4468d54604eb98765b9e8d/RodTracker/src/RodTracker/ui/reconstruction.py#L18
