### 2024-01-23 ANP-Granular/ParticleTracking#83 @mhubii
[JOSS Review] master -> main
Re-name inappropriate `master` branch to `main` for cultural reasons https://www.theserverside.com/feature/Why-GitHub-renamed-its-master-branch-to-main
