### 2024-01-23 ANP-Granular/ParticleTracking#84 @mhubii
[JOSS Review] Run tests / publish docs / scheduled run in workflow
Again, please advise if wrong, are the tests run in the workflow https://github.com/ANP-Granular/ParticleTracking/blob/master/.github/workflows/ci.yml ? I can only find linting https://github.com/ANP-Granular/ParticleTracking/actions/runs/7207660157

Also, are the docs published on PR?

Please further consider a scheduled run, say weekly or similar, refer https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule
