### 2024-01-24 ANP-Granular/ParticleTracking#75 @mhubii
hi both, just seeing this issue. It should not be necessary to download manually. Please use torch hub, refer #85. Please further consider publishing as asset in release (although this is not necessary)
