### 2024-01-25 ANP-Granular/ParticleTracking#84 @a-niem
### Tests
> Again, please advise if wrong, are the tests run in the workflow https://github.com/ANP-Granular/ParticleTracking/blob/master/.github/workflows/ci.yml ? I can only find linting https://github.com/ANP-Granular/ParticleTracking/actions/runs/7207660157
> [...]
> Please further consider a scheduled run, say weekly or similar, refer https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#schedule

Currently, it is intended to only run linting in the mentioned workflow. This was decided because the tests take relatively long and on machines without a display require a proper setup of xvfb which has not been done yet.
We can add the running of tests to the workflow, as well as a time trigger.

### Docs
> Also, are the docs published on PR?

The respective versions of docs are automatically updated upon commits to the `master` and `develop` and when a new version tag is added.
