### 2024-01-25 ANP-Granular/ParticleTracking#84 @mhubii
hm okay I see that this might be complicated, but I think running the tests via workflows would be necessary.
