+++ RodTracker/src/RodTracker/backend/logger.py
@@ -1,34 +1,36 @@
-#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev, and others
 #
-#  This file is part of RodTracker.
-#  RodTracker is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
+# This file is part of RodTracker.
+# RodTracker is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 #
-#  RodTracker is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
+# RodTracker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 #
-#  You should have received a copy of the GNU General Public License
-#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+# You should have received a copy of the GNU General Public License
+# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
 """**TBD**"""
 
-import os
 import logging
+import os
 import pathlib
-import sys
 import subprocess
+import sys
 from abc import abstractmethod
 from enum import Enum, auto
-from typing import Optional, Iterable, Union, List
+from typing import Iterable, List, Optional, Union
+
 import numpy as np
-from PyQt5.QtWidgets import QListWidgetItem
 from PyQt5 import QtCore
-from RodTracker import LOG_FILE
+from PyQt5.QtWidgets import QListWidgetItem
+
 import RodTracker.ui.rodnumberwidget as rn
+from RodTracker import LOG_FILE
 
 _logger = logging.getLogger(__name__)
 
@@ -38,12 +40,15 @@ def exception_logger(e_type, e_value, e_tb):
     _logger.exception("Uncaught exception:", exc_info=(e_type, e_value, e_tb))
 
 
-def qt_error_handler(mode: QtCore.QtMsgType,
-                     context: QtCore.QMessageLogContext, msg: str):
+def qt_error_handler(
+    mode: QtCore.QtMsgType, context: QtCore.QMessageLogContext, msg: str
+):
     """Handler for logging uncaught Qt exceptions during the program flow."""
-    context_info = (f"category: {context.category}\n"
-                    f"function: {context.function}, line: {context.line}\n"
-                    f"file: {context.file}\n")
+    context_info = (
+        f"category: {context.category}\n"
+        f"function: {context.function}, line: {context.line}\n"
+        f"file: {context.file}\n"
+    )
     if mode == QtCore.QtInfoMsg:
         _logger.info(context_info + f"{msg}")
     elif mode == QtCore.QtWarningMsg:
@@ -114,6 +119,7 @@ class NumberChangeActions(Enum):
         Indicates a switch of rod numbers in the current camera only and the
         current frame only.
     """
+
     ALL = auto()
     ALL_ONE_CAM = auto()
     ONE_BOTH_CAMS = auto()
@@ -122,6 +128,7 @@ class NumberChangeActions(Enum):
 
 class NotInvertableError(Exception):
     """Raised when a not invertable action is attempted to be inverted."""
+
     pass
 
 
@@ -220,9 +227,16 @@ class FileAction(Action):
 
[…]
