+++ RodTracker/tests/test_backend/test_rod_data.py
@@ -1,33 +1,34 @@
-#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev, and others
 #
-#  This file is part of RodTracker.
-#  RodTracker is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
+# This file is part of RodTracker.
+# RodTracker is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 #
-#  RodTracker is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
+# RodTracker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
 #
-#  You should have received a copy of the GNU General Public License
-#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+# You should have received a copy of the GNU General Public License
+# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
 import logging
-from pathlib import Path
 import random
 import shutil
+from pathlib import Path
 
 import importlib_resources
 import pandas as pd
-from PyQt5 import QtWidgets
 import pytest
+from conftest import load_rod_data
+from PyQt5 import QtWidgets
 from pytest import MonkeyPatch
 from pytestqt.qtbot import QtBot
 
-from conftest import load_rod_data
-from RodTracker.backend import rod_data, logger as lg
+from RodTracker.backend import logger as lg
+from RodTracker.backend import rod_data
 from RodTracker.backend.rod_data import RodData
 
 _logger = logging.getLogger()
