### 2024-01-30 3ac99fb
Sort imports in ParticleDetection
Sort the imports (see #80) and apply formatting with black in the ParticleDetection files.
