+++ ParticleDetection/tests/test_reconstruct_3D/test_calibrate_cameras.py
@@ -1,29 +1,29 @@
-#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev, and others
 #
-#  This file is part of ParticleDetection.
-#  ParticleDetection is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
+# This file is part of ParticleDetection.
+# ParticleDetection is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 #
-#  ParticleDetection is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
+# ParticleDetection is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
 #
-#  You should have received a copy of the GNU General Public License
-#  along with ParticleDetection.  If not, see <http://www.gnu.org/licenses/>.
+# You should have received a copy of the GNU General Public License
+# along with ParticleDetection. If not, see <http://www.gnu.org/licenses/>.
 
 import numpy as np
 import pandas as pd
 import pytest
-import ParticleDetection.reconstruct_3D.calibrate_cameras as cc
-import ParticleDetection.utils.data_loading as dl
 from conftest import EXAMPLES
 
+import ParticleDetection.reconstruct_3D.calibrate_cameras as cc
+import ParticleDetection.utils.data_loading as dl
+
 calibration = dl.load_camera_calibration(EXAMPLES / "gp34.json")
-transformation = dl.load_world_transformation(
-    EXAMPLES / "transformation.json")
+transformation = dl.load_world_transformation(EXAMPLES / "transformation.json")
 
 
 @pytest.mark.parametrize("transformation", [transformation, None])
