+++ ParticleDetection/src/ParticleDetection/reconstruct_3D/visualization.py
@@ -1,18 +1,18 @@
-#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+# Copyright (c) 2023-24 Adrian Niemann, Dmitry Puzyrev
 #
-#  This file is part of ParticleDetection.
-#  ParticleDetection is free software: you can redistribute it and/or modify
-#  it under the terms of the GNU General Public License as published by
-#  the Free Software Foundation, either version 3 of the License, or
-#  (at your option) any later version.
+# This file is part of ParticleDetection.
+# ParticleDetection is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 #
-#  ParticleDetection is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#  GNU General Public License for more details.
+# ParticleDetection is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
 #
-#  You should have received a copy of the GNU General Public License
-#  along with ParticleDetection.  If not, see <http://www.gnu.org/licenses/>.
+# You should have received a copy of the GNU General Public License
+# along with ParticleDetection. If not, see <http://www.gnu.org/licenses/>.
 
 """
 Collection of plotting functions to evaluate automatic 3D rod position
@@ -23,14 +23,15 @@ reconstruction from images of a stereocamera system.
 
 """
 import logging
-from typing import Iterable, List, Union, Tuple
-import numpy as np
-import networkx as nx
+from typing import Iterable, List, Tuple, Union
+
+import matplotlib.animation as animation
 import matplotlib.pyplot as plt
+import networkx as nx
+import numpy as np
 from matplotlib.figure import Figure
 from matplotlib.widgets import Slider
 from mpl_toolkits.mplot3d.art3d import Line3D
-import matplotlib.animation as animation
 
 _logger = logging.getLogger(__name__)
 

[…]
