### 2024-01-30 ee9ddf3
Adjust workflows
Replace previous linting workflow with updated version, that uses isort and black. (see #80)
