+++ .github/workflows/code_style.yml
@@ -1,0 +1,42 @@
+name: Code Style
+on:
+  push:
+    branches: 
+      - master
+      - develop
+  pull_request:
+    branches:
+      - master
+      - develop
+  workflow_dispatch:
+
+jobs:
+  Code_Style_Adherence: 
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python_version: ['3.11']
+    steps:
+        - uses: actions/checkout@v3
+        - name: Set up Python ${{ matrix.python_version }}
+          uses: actions/setup-python@v4
+          with:
+            python-version: ${{ matrix.python_version }}
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            pip install flake8
+            pip install isort
+            pip install black
+        - name: Lint
+          run: |
+            flake8 --append-config ./RodTracker/setup.cfg ./RodTracker
+            flake8 ./ParticleDetection
+        - name: Import Sorting
+          run: |
+            isort --settings-file ./RodTracker/pyproject.toml --check ./RodTracker
+            isort --settings-file ./ParticleDetection/pyproject.toml --check ./ParticleDetection
+        - name: Style Adherence
+          run: |
+            black --config ./RodTracker/pyproject.toml --check ./RodTracker
+            black --config ./ParticleDetection/pyproject.toml --check ./ParticleDetection
