+++ .github/workflows/ci.yml
@@ -1,35 +1,0 @@
-name: RodTracker package
-on:
-  push:
-    branches: 
-      - master
-      - develop
-  pull_request:
-    branches:
-      - master
-      - develop
-  workflow_dispatch:
-
-jobs:
-  Linting:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python_version: ['3.8', '3.9', '3.10']
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python_version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python_version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install ./RodTracker[DEV]
-        pip install ./ParticleDetection[TEST]
-    - name: Lint with flake8
-      run: |
-        flake8 --config ./RodTracker/setup.cfg ./RodTracker
-        flake8 ./ParticleDetection
-      continue-on-error: false
