+++ RodTracker/setup.cfg
@@ -6,4 +6,4 @@ exclude =
     __pycache__,
     build,
     dist,
-    src/RodTracker/ui/mainwindow_layout.py
+    *src/RodTracker/ui/mainwindow_layout.py
