+++ ParticleDetection/pyproject.toml
@@ -176,3 +176,11 @@ addopts = "--cov=src --cov-report=html"
 testpaths = [
     "tests/",
 ]
+
+[tool.black]
+line-length = 79
+target-version = ['py38', 'py39', 'py310', 'py311']
+
+[tool.isort]
+profile = "black"
+line_length = 79
