### 2024-01-30 cd99118
Add flake8 configuration
Adds flake8 configuration to ParticleDetection.
