+++ ParticleDetection/setup.cfg
@@ -1,0 +1,6 @@
+[flake8]
+ignore = W503, W504, E203
+exclude =
+    __pycache__,
+    build,
+    dist
