### 2024-01-31 5a21a40
Add workflow for ParticleDetection testing
Adds a workflow to (re-)run tests for the ParticleDetection package (see #84).
