+++ ParticleDetection/pyproject.toml
@@ -172,7 +172,11 @@ DETECTRON = ["imgaug", "Pillow", "protobuf", "tensorboard"]
 TEST = ["pytest", "importlib_resources", "pytest-cov"]
 
 [tool.pytest.ini_options]
-addopts = "--cov=src --cov-report=html"
+addopts = [
+    "--cov=ParticleDetection",
+    "--cov-report=term",
+    "--cov-fail-under=50"
+]
 testpaths = [
     "tests/",
 ]
