+++ .github/workflows/test_particledetection.yml
@@ -1,0 +1,37 @@
+name: Test ParticleDetection
+on:
+  push:
+    branches: 
+      - master
+      - develop
+    paths:
+      - 'ParticleDetection/**'
+  pull_request:
+    branches: 
+      - master
+      - develop
+    paths:
+      - 'ParticleDetection/**'
+  workflow_dispatch:
+
+jobs:
+    tests: 
+        strategy:
+          matrix:
+            os: [ubuntu-latest, macos-latest, windows-latest]
+            python_version: ['3.8', '3.9', '3.10', '3.11']
+        runs-on: ${{ matrix.os }}
+        steps:
+            - uses: actions/checkout@v4
+            - name: Set up Python ${{ matrix.python_version }}
+              uses: actions/setup-python@v5
+              with:
+                python-version: ${{ matrix.python_version }}
+            - name: Install dependencies
+              run: |
+                python -m pip install --upgrade pip
+                pip install ./ParticleDetection[test]
+            - name: Run tests
+              run: |
+                cd ./ParticleDetection
+                pytest
