+++ .github/workflows/code_style.yml
@@ -17,9 +17,9 @@ jobs:
       matrix:
         python_version: ['3.11']
     steps:
-        - uses: actions/checkout@v3
+        - uses: actions/checkout@v4
         - name: Set up Python ${{ matrix.python_version }}
-          uses: actions/setup-python@v4
+          uses: actions/setup-python@v5
           with:
             python-version: ${{ matrix.python_version }}
         - name: Install dependencies
