### 2024-02-03 a4b5e5b
Add workflow for RodTracker testing
Adds a workflow to (re-)run tests for the RodTracker package (see #84).
