+++ .github/workflows/test_rodtracker.yml
@@ -1,0 +1,40 @@
+name: Test RodTracker
+on:
+  push:
+    branches: 
+      - master
+      - develop
+    paths:
+      - 'RodTracker/**'
+  pull_request:
+    branches: 
+      - master
+      - develop
+    paths:
+      - 'RodTracker/**'
+  workflow_dispatch:
+
+jobs:
+    tests: 
+        runs-on: ubuntu-latest
+        strategy:
+          matrix:
+            python_version: ['3.8', '3.9', '3.10', '3.11']
+        env:
+          DISPLAY: ':99.0'
+        steps:
+            - uses: actions/checkout@v4
+            - name: Set up Python ${{ matrix.python_version }}
+              uses: actions/setup-python@v5
+              with:
+                python-version: ${{ matrix.python_version }}
+            - uses: tlambert03/setup-qt-libs@v1
+            - name: Install dependencies
+              run: |
+                python -m pip install --upgrade pip
+                pip install ./RodTracker[test]
+                pip install pytest-xvfb
+            - name: Run tests
+              run: |
+                cd ./RodTracker
+                pytest
