### 2024-02-03 5c3674c
Continue test workflows during parital error
Avoid aborting complete test workflow during an error in one/few running configuration(s).
