+++ .github/workflows/test_particledetection.yml
@@ -15,7 +15,8 @@ on:
   workflow_dispatch:
 
 jobs:
-    tests: 
+    tests:
+        continue-on-error: true 
         strategy:
           matrix:
             os: [ubuntu-latest, macos-latest, windows-latest]
