+++ .github/workflows/test_rodtracker.yml
@@ -17,6 +17,7 @@ on:
 jobs:
     tests: 
         runs-on: ubuntu-latest
+        continue-on-error: true
         strategy:
           matrix:
             python_version: ['3.8', '3.9', '3.10', '3.11']
