### 2024-03-06 ANP-Granular/ParticleTracking#86 @mhubii
[JOSS Review] Link broken
Link to https://particletracking.readthedocs.io/en/latest in

https://github.com/ANP-Granular/ParticleTracking/blob/joss_paper/paper/paper.md

broken.

Ref https://github.com/openjournals/joss-reviews/issues/5986
