+++ .github/workflows/test_particledetection.yml
@@ -2,13 +2,13 @@ name: Test ParticleDetection
 on:
   push:
     branches: 
-      - master
+      - main
       - develop
     paths:
       - 'ParticleDetection/**'
   pull_request:
     branches: 
-      - master
+      - main
       - develop
     paths:
       - 'ParticleDetection/**'
