+++ .github/workflows/code_style.yml
@@ -2,11 +2,11 @@ name: Code Style
 on:
   push:
     branches: 
-      - master
+      - main
       - develop
   pull_request:
     branches:
-      - master
+      - main
       - develop
   workflow_dispatch:
 
