### 2024-05-04 4784de3
Mitigate Tkinter issue
Mitigates a python issue where Tkinter is unable to find a usable init.tcl.
See https://github.com/python/cpython/issues/111754 for further information.
