+++ ParticleDetection/src/ParticleDetection/reconstruct_3D/visualization.py
@@ -22,7 +22,11 @@ reconstruction from images of a stereocamera system.
 **Date:**       01.11.2022
 
 """
+import glob
 import logging
+import os
+import sys
+from pathlib import Path
 from typing import Iterable, List, Tuple, Union
 
 import matplotlib.animation as animation
@@ -34,6 +38,37 @@ from matplotlib.widgets import Slider
 from mpl_toolkits.mplot3d.art3d import Line3D
 
 _logger = logging.getLogger(__name__)
+
+
+def set_tk_tcl_paths() -> None:
+    """Mitigate issue with not found Tkinter.
+
+    Uses the workaround shown in the python issue describing Tkinter being
+    unable to find a usable init.tcl. For further information see the original
+    issue:
+    https://github.com/python/cpython/issues/111754
+    """
+    os.environ["TCL_LIBRARY"] = str(
+        Path(
+            glob.glob(
+                os.path.join(sys.base_prefix, "tcl", "tcl*", "init.tcl")
+            )[0]
+        ).parent
+    )
+    os.environ["TK_LIBRARY"] = str(
+        Path(
+            glob.glob(
+                os.path.join(sys.base_prefix, "tcl", "tk*", "pkgIndex.tcl")
+            )[0]
+        ).parent
+    )
+    os.environ["TIX_LIBRARY"] = str(
+        Path(
+            glob.glob(
+                os.path.join(sys.base_prefix, "tcl", "tix*", "pkgIndex.tcl")
+            )[0]
+        ).parent
+    )
 
 
 def matching_results(
@@ -59,6 +94,7 @@ def matching_results(
         [0]: reprojection errors histogram\n
         [1]: rod lengths histogram
     """
+    set_tk_tcl_paths()
     fig1 = reprojection_errors_hist(reprojection_errors)
     fig2 = length_hist(rod_lengths)
 
@@ -66,6 +102,9 @@ def matching_results(
         return fig1, fig2
     plt.show()
     return
+
+
+set_tk_tcl_paths()
 
 
 def length_hist(rod_lengths: np.ndarray) -> Figure:
