### 2024-05-05 d87501d
Replacing/removing URL dependencies
Removes or replaces URL dependencies to comply with requirement from PyPI (see #78).
This also required removing the CPU/GPU extras because the choice of an extra-index is not yet working correctly with pip (see https://github.com/python-poetry/poetry/issues/6409 for more details).
