+++ ParticleDetection/pyproject.toml
@@ -52,102 +52,17 @@ scikit-image = [
     {version = ">=0.21.0,<0.22", python = "<3.9"},
 ]
 scipy = [
-    {version=">=1.9.3", python = ">=3.9"},
+    {version=">=1.9.3,!=1.13.0", python = ">=3.9"},
     {version="<1.11.0", python = "<3.9"}
 ]
-torch = [
-    # Linux
-    {version = ">=2", platform = "linux", markers = "extra != 'cpu'"},
-    # The commented version below is the intended one, that does not work with
-    # pip. The used version specifies the URLs manually which is not desirable.
-    # {version = ">=2", platform = "linux", source = "pytorch_cpu", markers = "extra == 'cpu'"},
-    {python = "==3.8", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp38-cp38-linux_x86_64.whl"},
-    {python = "==3.9", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp39-cp39-linux_x86_64.whl"},
-    {python = "==3.10", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl"},
-    {python = "==3.11", platform = "linux", markers = "extra == 'cpu' and platform_machine == 'x86_64'", url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp311-cp311-linux_x86_64.whl"},
-    # Fall-back version for architecture!=x86, although it probably won't work
-    # (see explanation above)
-    {version = ">=2", platform = "linux", source = "pytorch_cpu", markers = "extra == 'cpu' and platform_machine != 'x86_64'"},
-
-    # Windows
-    {version = ">=2", platform = "win32", markers = "extra != 'gpu' and extra != 'detectron'"},
-    # The commented version below is the intended one, that does not work with
-    # pip. The used version specifies the URLs manually which is not desirable.
-    # {version = ">=2", platform = "win32", source = "pytorch_cu118", markers = "extra == 'gpu' or extra == 'detectron'"},
-    {python = "==3.8", platform = "win32", markers = "(extra == 'gpu' or extra == 'detectron') and platform_machine == 'AMD64'", url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp38-cp38-win_amd64.whl"},
-    {python = "==3.9", platform = "win32", markers = "(extra == 'gpu' or extra == 'detectron') and platform_machine == 'AMD64'", url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp39-cp39-win_amd64.whl"},
-    {python = "==3.10", platform = "win32", markers = "(extra == 'gpu' or extra == 'detectron') and platform_machine == 'AMD64'", url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp310-cp310-win_amd64.whl"},
-    {python = "==3.11", platform = "win32", markers = "(extra == 'gpu' or extra == 'detectron') and platform_machine == 'AMD64'", url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-win_amd64.whl"},
-    # Fall-back version for architecture!=AMD64, although it probably won't
-    # work (see explanation above)
[…]
