### 2024-05-05 1fa1a12
Update ParticleDetection documentation
Update information about the installation process.
Fixes #79.
