### 2024-05-05 03ea0aa
Bump ParticleDetection version and update its changelog
