### 2024-05-06 b62db76
Remove interference of Tkinter fix on Linux and macOS
