### 2024-05-06 3f23644
Add dynamic versioning to ParticleDetection
