### 2024-05-06 cdd6e6e
Add automatic publishing of ParticleDetection to PyPI
