+++ .github/workflows/publish_particledetection.yml
@@ -1,0 +1,35 @@
+name: Publish ParticleDetection
+on:
+  push:
+    branches:
+      - main
+      - develop
+    paths:
+      - 'ParticleDetection/**'
+  workflow_dispatch:
+
+jobs:
+  publish:
+    runs-on: ubuntu-latest
+    strategy:
+        matrix:
+          python_version: ['3.11']
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python ${{ matrix.python_version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python_version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install "poetry>=1.8.2"
+          poetry self add "poetry-dynamic-versioning[plugin]"
+      - name: Build package
+        run: |
+          cd ./ParticleDetection
+          poetry build
+      - name: Publish package
+        run: |
+          poetry config pypi-token.pypi ${{ secrets.PYPI_PARTICLEDETECTION_API_TOKEN }}
+          poetry publish
