### 2024-05-06 155ef05
Fix ParticleDetection's publishing action
