### 2024-05-06 07b6758
Modify poetry's timeout
Extends poetry's timeout to avoid the publishing to PyPI workflow breaking during 'longer' upload periods.
