### 2024-05-06 ce0f27d
Add scheduled running of tests
Weekly runs are added for the workflows testing ParticleDetection and RodTracker.
See #84.
