+++ .github/workflows/test_particledetection.yml
@@ -12,6 +12,8 @@ on:
       - develop
     paths:
       - 'ParticleDetection/**'
+  schedule:
+    - cron: '0 16 * * 5'
   workflow_dispatch:
 
 jobs:
