### 2024-05-09 eb3af87
Replacing/removing URL dependencies in RodTracker
Removes or replaces URL dependencies to comply with requirement from PyPI (see #78).
This also lead to the removal of the "GPU" installation option.
