+++ docs/source/installation/rodtracker.md
@@ -1,26 +1,46 @@
 # Installation of the RodTracker
-As the RodTracker is a Python package usually you will be installing it from source or from a `*.whl` file, but there is also the possibility, that you are provided with a bundled standalone version. This is either done as an executable installer (`RodTracker-Setup.exe` - Windows, `RodTracker-Setup.deb` - Linux, `RodTracker-Setup.dmg` - MacOS) or as folder from which the `RodTrackerApp` executable runs the RodTracker.
+As the RodTracker is a Python package usually you will be installing it from source or from a `*.whl` file, but there is also the possibility, that you are provided with a bundled standalone version. Get the executable installer (`RodTracker-Setup.exe` - Windows, `RodTracker-Setup.deb` - Linux, `RodTracker-Setup.dmg` - macOS) from the [repository's releases](https://github.com/ANP-Granular/ParticleTracking/releases).
 
-The following installation instructions are only concerning installing the RodTracker from source.
+The following installation instructions are only concerning installing the RodTracker as a python package.
 
-## Installation from source
+### Installation from PyPI
+**Requirements:**
+- Python `>=3.8`
+- pip
+
+Install the default version using pip:
+```shell
+pip install RodTracker
+```
+Or use one of the options described in the [documentation](https://particletracking.readthedocs.io/en/stable/installation/rodtracker.html#installation-options).
+```shell
+pip install RodTracker[OPTION]
+```
+
+### Installation from source
 
 **Requirements:**
 - Python `>=3.8` is installed
 - pip `!=22` is installed (see [this issue](https://github.com/pypa/pip/issues/10851) for explanation)
 - Git is installed
-- It is strongly recommended to install RodTracker from a fresh virtual environment (created with `venv`)
+- It is strongly recommended installing RodTracker from a fresh virtual environment (created with `venv`)
 ---
 
-0. Upgrade your version of `pip`.
-  ```shell
-  python -m pip install --upgrade pip
-  ```
-1. Clone the [repository](https://github.com/ANP-Granular/ParticleTracking) containing the RodTracker. Do **NOT** just copy the `RodTracker` folder. This will lead to a missing dependency during the installation.
+1. Cloning the [repository](https://github.com/ANP-Granular/ParticleTracking) containing the RodTracker. Do **NOT** just copy the `RodTracker` folder. This will lead to a missing dependency during the installation.
 2. Install it using `pip`.
-  ```shell
-  YOUR/REPO/PATH/RodTracker$ pip install .
-  ```
+   ```shell
+   YOUR/REPO/PATH/RodTracker$ pip install .
+   ```
+
+It is also possible to install it directly from GitHub (requires `Git` to be installed):
+```shell
+pip install 'git+https://github.com/ANP-Granular/ParticleTracking.git#egg=rodtracker&subdirectory=RodTracker'
+```
+```shell
+pip install 'rodtracker[DOCS] @ git+https://github.com/ANP-Granular/ParticleTracking.git#egg=RodTracker&s
+ubdirectory=RodTracker'
+```
+
 ### Installation in conda environments
 1. Create a fresh conda environment with Python `>=3.8`
   ```shell
@@ -48,9 +68,6 @@ pip install -e .[BUILD]
 ```
 Please remember to install in *editable* mode when attempting to change any functionality of the RodTracker.
 
-- **`GPU`:**
-  - installs dependencies necessary to run particle detections using the GPU
-
 - **`DEV`:**
   - intended to install all dependencies necessary or useful for the development of RodTracker
   - this will install all dependencies from the other extras as well
@@ -77,8 +94,8 @@ Refer to its [installation instructions](particledetection.md) for that.
 On Linux the default installation allows running particle detections on a GPU, if an appropriate detection model is selected and no further actions are required.
 
[…]
