### 2024-05-09 b343f93
Add automatic publishing of RodTracker to PyPI
