### 2024-05-16 410938b
Remove appnames from multiple locations
