### 2024-05-17 7513fcf
Move miscellaneous functions to separate file
