+++ RodTracker/src/RodTracker/backend/logger.py
@@ -17,10 +17,7 @@
 """**TBD**"""
 
 import logging
-import os
 import pathlib
-import subprocess
-import sys
 from abc import abstractmethod
 from enum import Enum, auto
 from typing import Iterable, List, Optional, Union
@@ -30,7 +27,6 @@ from PyQt5 import QtCore
 from PyQt5.QtWidgets import QListWidgetItem
 
 import RodTracker.ui.rodnumberwidget as rn
-from RodTracker import LOG_FILE
 
 _logger = logging.getLogger(__name__)
 
@@ -62,15 +58,6 @@ def qt_error_handler(
 
 
 QtCore.qInstallMessageHandler(qt_error_handler)
-
-
-def open_logs():
-    """Opens the log file."""
-    if sys.platform == "win32":
-        os.startfile(LOG_FILE)
-    else:
-        opener = "open" if sys.platform == "darwin" else "xdg-open"
-        subprocess.run([opener, LOG_FILE])
 
 
 class FileActions(Enum):
