+++ RodTracker/src/RodTracker/ui/mainwindow.py
@@ -19,7 +19,7 @@
 import platform
 from functools import partial
 from pathlib import Path
-from typing import Callable, List
+from typing import List
 
 from PyQt5 import QtCore, QtGui, QtWidgets
 from PyQt5.QtGui import QWheelEvent
@@ -33,6 +33,7 @@ from PyQt5.QtWidgets import (
 import RodTracker.backend.file_locations as fl
 import RodTracker.backend.img_data as img_data
 import RodTracker.backend.logger as lg
+import RodTracker.backend.miscellaneous as misc
 import RodTracker.backend.rod_data as r_data
 import RodTracker.backend.settings as se
 import RodTracker.ui.mainwindow_layout as mw_l
@@ -197,7 +198,7 @@ class RodTrackWindow(QtWidgets.QMainWindow):
             )
 
         # Tab icons for 'busy' indication
-        default_icon = blank_icon()
+        default_icon = misc.blank_icon()
         self.ui.right_tabs.setIconSize(QtCore.QSize(7, 16))
         for tab in range(self.ui.right_tabs.count()):
             self.ui.right_tabs.setTabIcon(tab, default_icon)
@@ -407,21 +408,9 @@ class RodTrackWindow(QtWidgets.QMainWindow):
         self.ui.action_about_qt.triggered.connect(
             lambda: QMessageBox.aboutQt(self, APPNAME)
         )
-        self.ui.action_logs.triggered.connect(lg.open_logs)
-        self.ui.action_bug_report.triggered.connect(
-            lambda: QtGui.QDesktopServices.openUrl(
-                QtCore.QUrl(
-                    "https://github.com/ANP-Granular/ParticleTracking/issues/new?labels=bug&projects=&template=bug_report.md&title="  # noqa: E501
-                )
-            )
-        )
-        self.ui.action_feature_request.triggered.connect(
-            lambda: QtGui.QDesktopServices.openUrl(
-                QtCore.QUrl(
-                    "https://github.com/ANP-Granular/ParticleTracking/issues/new?labels=enhancement&projects=&template=feature_request.md&title="  # noqa: E501
-                )
-            )
-        )
+        self.ui.action_logs.triggered.connect(misc.open_logs)
+        self.ui.action_bug_report.triggered.connect(misc.report_issue)
+        self.ui.action_feature_request.triggered.connect(misc.request_feature)
 
     @QtCore.pyqtSlot(QTreeWidgetItem, int)
     def tree_selection(self, item: QTreeWidgetItem, col: int):
@@ -838,34 +827,34 @@ class RodTrackWindow(QtWidgets.QMainWindow):
         manager = self.image_managers[new_idx]
         cam = self.cameras[new_idx]
 
-        reconnect(self.rod_data.data_2d, cam.extract_rods)
+        misc.reconnect(self.rod_data.data_2d, cam.extract_rods)
 
-        reconnect(
+        misc.reconnect(
             self.ui.action_shorten_displayed.triggered,
             lambda: cam.adjust_rod_length(-self._rod_incr, False),
         )
-        reconnect(
+        misc.reconnect(
             self.ui.action_lengthen_displayed.triggered,
             lambda: cam.adjust_rod_length(self._rod_incr, False),
         )
-        reconnect(
+        misc.reconnect(
             self.ui.action_shorten_selected.triggered,
             lambda: cam.adjust_rod_length(-self._rod_incr, True),
         )
-        reconnect(
+        misc.reconnect(
             self.ui.action_lengthen_selected.triggered,
             lambda: cam.adjust_rod_length(self._rod_incr, True),
         )
 
-        reconnect(
+        misc.reconnect(
[…]
