### 2024-05-21 da3c7ed
Add a popup window when unhandled exceptions occur
Adds a popup error dialog such that users are made aware of unhandled exceptions. (See #74)
