### 2024-05-21 9bc2f6d
Adopt new ErrorDialog in all error handlers
Adopts the new error handling to all asynchronously executed functionalities.
