+++ RodTracker/src/RodTracker/backend/detection.py
@@ -22,12 +22,13 @@ from typing import Dict, List
 
 import pandas as pd
 import torch
+from PyQt5 import QtCore
+
 from ParticleDetection.utils import datasets as ds
 from ParticleDetection.utils import detection
 from ParticleDetection.utils import helper_funcs as hf
-from PyQt5 import QtCore
-
 from RodTracker.backend.logger import Action, NotInvertableError
+from RodTracker.backend.parallelism import error_handler
 
 _logger = logging.getLogger(__name__)
 abort_requested: bool = False
@@ -230,6 +231,7 @@ class Detector(QtCore.QRunnable):
             threshold = 0.0
         self.threshold = threshold
 
+    @error_handler
     def run(self):
         """Run the detection of rods with the parameters set in this
         :class:`Detector` object.
