+++ RodTracker/src/RodTracker/ui/mainwindow_layout.py
@@ -2,7 +2,7 @@
 
 # Form implementation generated from reading ui file 'mainwindow_layout.ui'
 #
-# Created by: PyQt5 UI code generator 5.15.9
+# Created by: PyQt5 UI code generator 5.15.10
 #
 # WARNING: Any manual changes made to this file will be lost when pyuic5 is
 # run again.  Do not edit this file unless you know what you are doing.
@@ -355,7 +355,7 @@ class Ui_MainWindow(object):
         self.sa_camera_1.setWidgetResizable(True)
         self.sa_camera_1.setObjectName("sa_camera_1")
         self.scrollAreaWidgetContents_8 = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_8.setGeometry(QtCore.QRect(0, 0, 679, 442))
+        self.scrollAreaWidgetContents_8.setGeometry(QtCore.QRect(0, 0, 100, 30))
         self.scrollAreaWidgetContents_8.setObjectName("scrollAreaWidgetContents_8")
         self.verticalLayout_18 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_8)
         self.verticalLayout_18.setContentsMargins(0, 0, 0, 0)
@@ -455,7 +455,7 @@ class Ui_MainWindow(object):
         self.scrollArea_2.setWidgetResizable(True)
         self.scrollArea_2.setObjectName("scrollArea_2")
         self.scrollAreaWidgetContents_2 = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_2.setGeometry(QtCore.QRect(0, 0, 431, 463))
+        self.scrollAreaWidgetContents_2.setGeometry(QtCore.QRect(0, 0, 234, 463))
         self.scrollAreaWidgetContents_2.setObjectName("scrollAreaWidgetContents_2")
         self.verticalLayout_9 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_2)
         self.verticalLayout_9.setObjectName("verticalLayout_9")
@@ -684,6 +684,13 @@ class Ui_MainWindow(object):
         self.tb_model = QtWidgets.QToolButton(self.tab_detection)
         self.tb_model.setObjectName("tb_model")
         self.horizontalLayout_13.addWidget(self.tb_model)
+        self.pb_use_example = QtWidgets.QPushButton(self.tab_detection)
+        font = QtGui.QFont()
+        font.setBold(True)
+        font.setWeight(75)
+        self.pb_use_example.setFont(font)
+        self.pb_use_example.setObjectName("pb_use_example")
+        self.horizontalLayout_13.addWidget(self.pb_use_example)
         self.verticalLayout_13.addLayout(self.horizontalLayout_13)
         self.horizontalLayout_12 = QtWidgets.QHBoxLayout()
         self.horizontalLayout_12.setObjectName("horizontalLayout_12")
@@ -1115,6 +1122,7 @@ class Ui_MainWindow(object):
         self.right_tabs.setTabText(self.right_tabs.indexOf(self.tab_3d), _translate("MainWindow", "3D-View"))
         self.lbl_model.setText(_translate("MainWindow", "Model: "))
         self.tb_model.setText(_translate("MainWindow", "..."))
+        self.pb_use_example.setText(_translate("MainWindow", "Use Example Model"))
         self.lbl_start_frame_detection.setText(_translate("MainWindow", "Start Frame:"))
         self.lbl_end_frame_detection.setText(_translate("MainWindow", "End Frame:"))
         self.lbl_expected_particles_default.setText(_translate("MainWindow", "Expected particles:"))
