### 2024-05-22 3258488
Don't stop style checks after first failure
