+++ RodTracker/src/RodTracker/backend/reconstruction.py
@@ -24,14 +24,14 @@ from typing import List
 import numpy as np
 import pandas as pd
 from matplotlib.figure import Figure
-from PyQt5 import QtCore
-from scipy.spatial.transform import Rotation as R
-
 from ParticleDetection.reconstruct_3D import calibrate_cameras as cc
 from ParticleDetection.reconstruct_3D import matchND
 from ParticleDetection.reconstruct_3D import visualization as vis
 from ParticleDetection.reconstruct_3D.match2D import match_frame
 from ParticleDetection.utils import data_loading as dl
+from PyQt5 import QtCore
+from scipy.spatial.transform import Rotation as R
+
 from RodTracker.backend.parallelism import error_handler
 
 _logger = logging.getLogger(__name__)
