### 2024-05-23 c592255
Add pre-commit configuration
