+++ .pre-commit-config.yaml
@@ -1,0 +1,60 @@
+default_language_version:
+  python: python3.11
+exclude: '.*/example_.*/.*|.*/mainwindow_layout.py'
+repos:
+  - repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.6.0
+    hooks:
+    - id: trailing-whitespace
+    - id: end-of-file-fixer
+    - id: check-yaml
+    - id: check-added-large-files
+  - repo: https://github.com/pycqa/flake8
+    rev: 7.0.0
+    hooks:
+      - id: flake8
+        name: flake8 (RodTracker)
+        files: ^RodTracker
+        args:
+          - "--append-config"
+          - "./RodTracker/setup.cfg"
+  - repo: https://github.com/pycqa/flake8
+    rev: 7.0.0
+    hooks:
+      - id: flake8
+        name: flake8 (ParticleDetection)
+        files: ^ParticleDetection
+        args:
+          - "--append-config"
+          - "./ParticleDetection/setup.cfg"
+  - repo: https://github.com/psf/black
+    rev: 24.4.2
+    hooks:
+      - id: black
+        name: black (RodTracker)
+        files: ^RodTracker
+        exclude: '.*/mainwindow_layout.py'
+        args:
+          - "--config"
+          - "./RodTracker/pyproject.toml"
+  - repo: https://github.com/psf/black
+    rev: 24.4.2
+    hooks:
+      - id: black
+        name: black (ParticleDetection)
+        files: ^ParticleDetection
+        args:
+          - "--config"
+          - "./ParticleDetection/pyproject.toml"
+  - repo: https://github.com/PyCQA/isort
+    rev: 5.13.2
+    hooks:
+      - id: isort
+        name: isort (RodTracker)
+        files: ^RodTracker
+  - repo: https://github.com/PyCQA/isort
+    rev: 5.13.2
+    hooks:
+      - id: isort
+        name: isort (ParticleDetection)
+        files: ^ParticleDetection
