### 2024-05-23 ffe934c
Apply initial fixes from pre-commit hooks
