+++ docs/source/RodTracker-api/RodTracker.backend.rst
@@ -15,5 +15,5 @@ RodTracker.backend
    backend/logger
    backend/parallelism
    backend/reconstruction
-   backend/rod_data 
+   backend/rod_data
    backend/settings
