+++ .github/workflows/test_rodtracker.yml
@@ -1,13 +1,13 @@
 name: Test RodTracker
 on:
   push:
-    branches: 
+    branches:
       - main
       - develop
     paths:
       - 'RodTracker/**'
   pull_request:
-    branches: 
+    branches:
       - main
       - develop
     paths:
@@ -17,7 +17,7 @@ on:
   workflow_dispatch:
 
 jobs:
-    tests: 
+    tests:
         runs-on: ubuntu-latest
         continue-on-error: true
         strategy:
