### 2024-05-25 1c13938
Update RodTracker docs
Adds information to the installation process on Linux.
Adds missing API documentation.
Adds a link to the source code (fixes #81).
Adds information on how to use the example detection model (fixes #85, #75).
