+++ docs/source/RodTracker-api/backend/miscellaneous.rst
@@ -1,0 +1,7 @@
+RodTracker.backend.miscellaneous
+--------------------------------
+
+.. automodule:: RodTracker.backend.miscellaneous
+   :members:
+   :undoc-members:
+   :show-inheritance:
