+++ docs/source/RodTracker-api/RodTracker.backend.rst
@@ -13,6 +13,7 @@ RodTracker.backend
    backend/file_locations
    backend/img_data
    backend/logger
+   backend/miscellaneous
    backend/parallelism
    backend/reconstruction
    backend/rod_data
