### 2024-05-26 6fd5351
Update app bundling
Updates ParticleDetection version.
Updates binary collection on Linux.
Fix link in changelog.
