+++ RodTracker/CHANGELOG.md
@@ -167,7 +167,7 @@ This especially improves performance and responsiveness of the `RodTracker` when
 - a versioning system
 
 [Unreleased]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.6.2...HEAD
-[v0.6.1]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.6.1...v0.6.2
+[v0.6.2]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.6.1...v0.6.2
 [v0.6.1]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.6.0...v0.6.1
 [v0.6.0]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.5.8...v0.6.0
 [v0.5.8]: https://github.com/ANP-Granular/ParticleTracking/compare/v0.5.7...v0.5.8
