### 2024-05-26 ANP-Granular/ParticleTracking#75 @a-niem
A button has been added to the `Detection` tab that downloads the example model from Zenodo.
The other example data won't be moved for now.
