### 2024-05-26 ANP-Granular/ParticleTracking#74 @a-niem
A popup window has been added that warns users when unhandled exceptions occur during program execution. It will also be shown when `ImportErrors` occur during startup of the application, e.g. when libraries are missing as described above.
