### 2024-05-26 71ea07a
Fix test workflow (RodTracker)
Taken from https://github.com/slaclab/pydm/blob/master/.github/workflows/run-tests.yml
