+++ .github/workflows/test_rodtracker.yml
@@ -36,7 +36,13 @@ jobs:
               run: |
                 python -m pip install --upgrade pip
                 pip install ./RodTracker[test]
-                pip install pytest-xvfb
+                sudo apt install xvfb herbstluftwm libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xfixes0 x11-utils
+            - name: Run Xvfb & window manager
+              run: |
+                sudo /sbin/start-stop-daemon --start --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -screen 0 1024x768x24 -ac +extension GLX +render -noreset
+                sleep 3
+                sudo /sbin/start-stop-daemon --start --pidfile /tmp/custom_herbstluftwm_99.pid --make-pidfile --background --exec /usr/bin/herbstluftwm
+                sleep 1
             - name: Run tests
               run: |
                 cd ./RodTracker
