### 2024-05-26 6a883a1
Fix problem CreateRod action
Adds a rod deactivation step before attempting to create a new rod.
