### 2024-05-26 f9630c0
Add additional waiting times and initial color selection to GUI tests
