+++ RodTracker/tests/gui_actions.py
@@ -101,6 +101,7 @@ class CreateRod:
                     cam, QtCore.Qt.MouseButton.LeftButton, pos=self.end
                 )
         if self.assertions:
+            qtbot.wait(1500)
             aa.post_create(main_window, self.new_id, self.start, self.end)
         return main_window
 
@@ -139,6 +140,7 @@ class DeleteRod:
                 qtbot.keyClick(rod, QtCore.Qt.Key_Enter)
                 qtbot.wait_until(increase_count, timeout=2000)
                 if self.assertions:
+                    qtbot.wait(500)
                     aa.post_delete(main_window, self.rod_id)
                 return main_window
         warn(f"The rod #{self.rod_id} was not found. No rod has been deleted.")
@@ -191,6 +193,7 @@ class ChangeRodPosition:
                 qtbot.keyClick(rod, QtCore.Qt.Key_Enter)
                 qtbot.wait_until(increased_count, timeout=2000)
                 if self.assertions:
+                    qtbot.wait(500)
                     aa.post_pos_change(
                         main_window, self.rod_id, self.start, self.end
                     )
@@ -265,7 +268,7 @@ class SwitchRodNumber:
                     if self.mode is not None:
                         qtbot.wait_until(increased_count, timeout=2000)
                     if self.assertions:
-                        qtbot.wait(150)
+                        qtbot.wait(500)
                         aa.post_number_switch(
                             main_window,
                             self.rod_id,
@@ -314,6 +317,7 @@ class SaveChanges:
             )
         qtbot.wait_until(contents_written, timeout=2000)
         if self.assertions:
+            qtbot.wait(500)
             aa.post_save(main_window, tmp_path, state)
         return main_window
 
@@ -351,6 +355,7 @@ class Undo:
         qtbot.wait_until(decrease_count, timeout=2000)
         qtbot.wait(250)
         if self.assertions:
+            qtbot.wait(500)
             aa.post_undo(main_window, state)
         return main_window
 
@@ -391,6 +396,7 @@ class Redo:
         qtbot.wait(250)
 
         if self.assertions:
+            qtbot.wait(500)
             aa.post_undo(main_window, state)
         return main_window
 
@@ -421,6 +427,7 @@ class SwitchFrame:
             qtbot.keyClick(main_window, direction_key)
             qtbot.wait(150)
         if self.assertions:
+            qtbot.wait(500)
             aa.post_switch_frame(main_window, self.direction, state)
         return main_window
 
@@ -458,6 +465,7 @@ class SwitchColor:
         qtbot.mouseClick(to_press, QtCore.Qt.MouseButton.LeftButton)
         qtbot.wait(150)
         if self.assertions:
+            qtbot.wait(500)
             aa.post_switch_color(main_window, self.color, state)
         return main_window
 
@@ -487,6 +495,7 @@ class SwitchCamera:
             )
 
         if self.assertions:
+            qtbot.wait(500)
             aa.post_switch_cam(main_window, state)
 
         return main_window
@@ -532,6 +541,7 @@ class LengthAdjustment:
 
         qtbot.keyClick(main_window, self.method)
         if self.assertions:
+            qtbot.wait(500)
             aa.post_length_adjustment(main_window, state)
         return main_window
 
@@ -583,7 +593,7 @@ class OpenImage:
                 main_window.ui.pb_load_images, QtCore.Qt.MouseButton.LeftButton
             )
         main_window.original_size()
-        qtbot.wait(50)
+        qtbot.wait(100)
         return main_window
 
 
