### 2024-05-27 #5986 @a-niem
@Kevin-Mattheus-Moerman I had provided updates, by fixing and closing the issues mentioned in this thread.
Specifically [#83](https://github.com/ANP-Granular/ParticleTracking/issues/83), [#80](https://github.com/ANP-Granular/ParticleTracking/issues/80), and [#79](https://github.com/ANP-Granular/ParticleTracking/issues/79). Since yesterday, all issues are fixed, with the exception of [#84](https://github.com/ANP-Granular/ParticleTracking/issues/84) not being closed yet.

I have put all my currently available strength in bringing this to a positive end since my last post here. And I thought that my work mentioned above would justify to continue this here.

I completely understand, if that's not the case for you. But would nevertheless kindly ask to keep the submission running and finish it.

Kind regards
Adrian
