### 2024-05-28 3594712
Apply formatting to all files
