+++ RodTracker/src/extensions/rods/dialogs.py
@@ -18,6 +18,7 @@
 
 import pandas as pd
 from PyQt5 import QtCore, QtGui, QtWidgets
+
 import RodTracker.backend.file_locations as fl
 
 
