+++ RodTracker/src/RodTracker/backend/data.py
@@ -16,8 +16,8 @@
 
 """**TBD**"""
 
-import os
 import logging
+import os
 from pathlib import Path
 from typing import Any, List, Tuple, Union
 
