### 2024-05-28 101b7ea
Add support for Python 3.12 to ParticleDetection
