+++ .github/workflows/test_particledetection.yml
@@ -22,7 +22,7 @@ jobs:
         strategy:
           matrix:
             os: [ubuntu-latest, macos-latest, windows-latest]
-            python_version: ['3.8', '3.9', '3.10', '3.11']
+            python_version: ['3.8', '3.9', '3.10', '3.11', '3.12']
         runs-on: ${{ matrix.os }}
         steps:
             - uses: actions/checkout@v4
