+++ ParticleDetection/pyproject.toml
@@ -29,7 +29,7 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.12"
+python = ">=3.8,<3.13"
 matplotlib = [
     {version = ">=3.8", python = ">=3.9"},
     {version = ">=3.6.1,<3.8", python = "<3.9"}
