+++ .github/workflows/test_rodtracker.yml
@@ -22,7 +22,7 @@ jobs:
         continue-on-error: true
         strategy:
           matrix:
-            python_version: ['3.8', '3.9', '3.10', '3.11']
+            python_version: ['3.8', '3.9', '3.10', '3.11', '3.12']
         env:
           DISPLAY: ':99.0'
         steps:
