### 2024-05-28 e43e784
Add Python 3.12 classifier to ParticleDetection metadata
