+++ docs/source/RodTracker-api/ui/loggerwidget.rst
@@ -1,8 +1,0 @@
-
-RodTracker.ui.loggerwidget
---------------------------
-
-.. automodule:: RodTracker.ui.loggerwidget
-   :members:
-   :undoc-members:
-   :show-inheritance:
