+++ docs/source/RodTracker-api/backend/rod_data.rst
@@ -1,7 +1,0 @@
-RodTracker.backend.rod\_data
-----------------------------
-
-.. automodule:: RodTracker.backend.rod_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
