+++ docs/source/RodTracker-api/ui/settings.rst
@@ -1,0 +1,7 @@
+RodTracker.ui.settings
+----------------------
+
+.. automodule:: RodTracker.ui.settings
+   :members:
+   :undoc-members:
+   :show-inheritance:
