+++ docs/source/RodTracker-api/backend/detection.rst
@@ -1,7 +1,0 @@
-RodTracker.backend.detection
-----------------------------
-
-.. automodule:: RodTracker.backend.detection
-   :members:
-   :undoc-members:
-   :show-inheritance:
