+++ docs/source/RodTracker-api/extensions/rods/display_tab.rst
@@ -1,0 +1,7 @@
+rods.display\_tab
+-----------------
+
+.. automodule:: extensions.rods.display_tab
+    :members:
+    :undoc-members:
+    :show-inheritance:
