+++ docs/source/RodTracker-api/backend/img_data.rst
@@ -1,7 +1,0 @@
-RodTracker.backend.img\_data
-----------------------------
-
-.. automodule:: RodTracker.backend.img_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
