+++ docs/source/RodTracker-api/extensions/rods/rods.rst
@@ -1,0 +1,7 @@
+rods.rods
+---------
+
+.. automodule:: extensions.rods.rods
+    :members:
+    :undoc-members:
+    :show-inheritance:
