+++ docs/source/RodTracker-api/extensions/rods/actions.rst
@@ -1,0 +1,7 @@
+rods.actions
+------------
+
+.. automodule:: extensions.rods.actions
+    :members:
+    :undoc-members:
+    :show-inheritance:
