+++ docs/source/RodTracker-api/ui/detection.rst
@@ -1,8 +1,0 @@
-RodTracker.ui.detection
------------------------
-
-.. automodule:: RodTracker.ui.detection
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members: _progress_update
