### 2024-05-31 875ca07
Connect slider and frame number
Slider and frame number now change according to image changes.
