### 2024-05-31 4f77010
Propagate loaded settings during app start
