### 2024-05-31 b252c00
Move Qt-specific error logging
