+++ docs/source/RodTracker/RodTracker.md
@@ -5,13 +5,13 @@ If you were given a bundled version, i.e. an executable, you can start right awa
 
 ## Starting the RodTracker
 Run the **RodTracker** GUI using one of the possibilities:
-  - Run `RodTracker.py` manually:
-    ```shell
-    YOUR/REPO/PATH/RodTracker/src/RodTracker$ python RodTracker.py
-    ```
   - Use the registered command:
     ```shell
     ARBITRARY/PATH$ RodTracker
+    ```
+  - Run `main.py` manually:
+    ```shell
+    YOUR/REPO/PATH/RodTracker/src/RodTracker$ python main.py
     ```
   - Execute a stand-alone executable bundled for your system.
 
