### 2024-06-03 dfceeb1
Update error handling in "rods" extension
