### 2024-06-04 62b281e
Connect Action logging to data loading
Trigger data changes/loading when an Action acting on data is logged.
