### 2024-06-04 151ed61
Initial commit of 'detection' extension
Adds the UI for the 'detection' extension.
WARNING: it is not yet usable
