+++ RodTracker/src/extensions/detection/detection.py
@@ -1,0 +1,283 @@
+#  Copyright (c) 2023 Adrian Niemann Dmitry Puzyrev
+#
+#  This file is part of RodTracker.
+#  RodTracker is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  RodTracker is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with RodTracker.  If not, see <http://www.gnu.org/licenses/>.
+
+"""**TBD**"""
+
+import logging
+from pathlib import Path
+from typing import Dict, List
+
+import pandas as pd
+import torch
+from ParticleDetection.utils import datasets as ds
+from ParticleDetection.utils import detection
+from ParticleDetection.utils import helper_funcs as hf
+from PyQt5 import QtCore
+
+from RodTracker.backend.logger import Action, NotInvertableError
+from RodTracker.ui.tabs import UtilityTab
+
+from .detection_ui import Ui_detector_ui
+
+_logger = logging.getLogger(__name__)
+abort_requested: bool = False
+lock = QtCore.QReadWriteLock(QtCore.QReadWriteLock.NonRecursive)
+
+
+class RodDetection(Action):
+    """Representation of the detection of rods on one frame as a loggable
+    action.
+
+    Parameters
+    ----------
+    frame : int
+        Frame the rods have been detected on.
+    cam_id : str
+        ID of the camera the image was produced by.
+    num_detected: int
+        Total number of rods that have been detected, i.e. all colors combined.
+    *args : Iterable
+        Positional arguments after ``text`` of the :class:`.Action`
+        superclass.
+    **kwargs : dict
+        Keyword arguments of the :class:`.Action` superclass.
+
+    Attributes
+    ----------
+    num_detected : int
+        Total number of rods that have been detected, i.e. all colors combined.
+    """
+
+    cam_id: str
+    """str : ID of the camera the image was produced by."""
+
+    def __init__(
+        self, frame: int, cam_id: str, num_detected: int, *args, **kwargs
+    ):
+        self.cam_id = cam_id
+        self.num_detected = num_detected
+        self._frame = frame
+        super().__init__(str(self), *args, **kwargs)
+
+    def __str__(self):
+        return (
+            f"({self.cam_id}, {self._frame}) Detected {self.num_detected} "
+            f"rods."
+        )
+
+    def undo(self, _):
+        """
+        Raises
+        ------
+        NotInvertableError
+            This action is generally not invertable.
+        """
+        raise NotInvertableError
+
+
+class DetectorSignals(QtCore.QObject):
+    """Helper object to provide :class:`Detector` access to ``pyqtSignal``."""
+
+    error = QtCore.pyqtSignal(tuple, name="error")
+    """pyqtSignal(tuple) : Signal for propagating errors occuring in the
+    :class:`Worker`'s thread.\n
+    | The transferred tuple should contain the following values:
+    | [0]: Exception type
+    | [1]: Exception value
+    | [2]: Exception traceback
+
+    See Also
+    --------
+    `sys.exc_info()`_
+
+    :py:obj:`sys.exc_info`
+
+    .. _sys.exc_info():
+        https://docs.python.org/3/library/sys.html#sys.exc_info
+    """
+
[…]
