+++ RodTracker/src/extensions/detection/detection_ui.py
@@ -1,0 +1,177 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '.\detection_ui.ui'
+#
+# Created by: PyQt5 UI code generator 5.15.10
+#
+# WARNING: Any manual changes made to this file will be lost when pyuic5 is
+# run again.  Do not edit this file unless you know what you are doing.
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class Ui_detector_ui(object):
+    def setupUi(self, detector_ui):
+        detector_ui.setObjectName("detector_ui")
+        detector_ui.resize(668, 533)
+        self.verticalLayout = QtWidgets.QVBoxLayout(detector_ui)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.horizontalLayout_13 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_13.setObjectName("horizontalLayout_13")
+        self.lbl_model = QtWidgets.QLabel(detector_ui)
+        self.lbl_model.setObjectName("lbl_model")
+        self.horizontalLayout_13.addWidget(self.lbl_model)
+        self.le_model = QtWidgets.QLineEdit(detector_ui)
+        self.le_model.setObjectName("le_model")
+        self.horizontalLayout_13.addWidget(self.le_model)
+        self.tb_model = QtWidgets.QToolButton(detector_ui)
+        self.tb_model.setObjectName("tb_model")
+        self.horizontalLayout_13.addWidget(self.tb_model)
+        self.verticalLayout.addLayout(self.horizontalLayout_13)
+        self.horizontalLayout_12 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_12.setObjectName("horizontalLayout_12")
+        self.lbl_start_frame_detection = QtWidgets.QLabel(detector_ui)
+        self.lbl_start_frame_detection.setObjectName("lbl_start_frame_detection")
+        self.horizontalLayout_12.addWidget(self.lbl_start_frame_detection)
+        self.start_frame_detection = QtWidgets.QSpinBox(detector_ui)
+        self.start_frame_detection.setObjectName("start_frame_detection")
+        self.horizontalLayout_12.addWidget(self.start_frame_detection)
+        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
+        self.horizontalLayout_12.addItem(spacerItem)
+        self.lbl_end_frame_detection = QtWidgets.QLabel(detector_ui)
+        self.lbl_end_frame_detection.setObjectName("lbl_end_frame_detection")
+        self.horizontalLayout_12.addWidget(self.lbl_end_frame_detection)
+        self.end_frame_detection = QtWidgets.QSpinBox(detector_ui)
+        self.end_frame_detection.setObjectName("end_frame_detection")
+        self.horizontalLayout_12.addWidget(self.end_frame_detection)
+        self.verticalLayout.addLayout(self.horizontalLayout_12)
+        self.horizontalLayout_14 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_14.setObjectName("horizontalLayout_14")
+        self.lbl_expected_particles_default = QtWidgets.QLabel(detector_ui)
+        self.lbl_expected_particles_default.setObjectName("lbl_expected_particles_default")
+        self.horizontalLayout_14.addWidget(self.lbl_expected_particles_default)
+        self.expected_particles_default = QtWidgets.QSpinBox(detector_ui)
+        self.expected_particles_default.setObjectName("expected_particles_default")
[…]
