+++ RodTracker/src/extensions/detection/requirements.txt
@@ -1,0 +1,4 @@
+opencv-python-headless
+pandas
+torch
+torchvision
