+++ RodTracker/pyproject.toml
@@ -111,7 +111,8 @@ line-length = 79
 target-version = ['py38', 'py39', 'py310', 'py311']
 extend-exclude = '''
 (
-    .*mainwindow_layout..*
+    .*mainwindow_layout..*|
+    .*detection_ui.py
 )
 '''
 
