+++ RodTracker/src/extensions/rods_3D/requirements.txt
@@ -1,0 +1,2 @@
+matplotlib
+PyQt3D
