+++ RodTracker/src/extensions/rods_reconstruction/reconstruction_ui.py
@@ -1,0 +1,184 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file '.\reconstruction_ui.ui'
+#
+# Created by: PyQt5 UI code generator 5.15.10
+#
+# WARNING: Any manual changes made to this file will be lost when pyuic5 is
+# run again.  Do not edit this file unless you know what you are doing.
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class Ui_reconstruction_ui(object):
+    def setupUi(self, reconstruction_ui):
+        reconstruction_ui.setObjectName("reconstruction_ui")
+        reconstruction_ui.resize(706, 661)
+        self.verticalLayout = QtWidgets.QVBoxLayout(reconstruction_ui)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.horizontalLayout_5 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
+        self.verticalLayout_11 = QtWidgets.QVBoxLayout()
+        self.verticalLayout_11.setObjectName("verticalLayout_11")
+        self.horizontalLayout_6 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_6.setObjectName("horizontalLayout_6")
+        self.lbl_start = QtWidgets.QLabel(reconstruction_ui)
+        self.lbl_start.setObjectName("lbl_start")
+        self.horizontalLayout_6.addWidget(self.lbl_start)
+        self.start_frame = QtWidgets.QSpinBox(reconstruction_ui)
+        self.start_frame.setObjectName("start_frame")
+        self.horizontalLayout_6.addWidget(self.start_frame)
+        self.verticalLayout_11.addLayout(self.horizontalLayout_6)
+        self.horizontalLayout_7 = QtWidgets.QHBoxLayout()
+        self.horizontalLayout_7.setObjectName("horizontalLayout_7")
+        self.lbl_end = QtWidgets.QLabel(reconstruction_ui)
+        self.lbl_end.setObjectName("lbl_end")
+        self.horizontalLayout_7.addWidget(self.lbl_end)
+        self.end_frame = QtWidgets.QSpinBox(reconstruction_ui)
+        self.end_frame.setObjectName("end_frame")
+        self.horizontalLayout_7.addWidget(self.end_frame)
+        self.verticalLayout_11.addLayout(self.horizontalLayout_7)
+        self.group_colors_reconstruction = QtWidgets.QGroupBox(reconstruction_ui)
+        self.group_colors_reconstruction.setObjectName("group_colors_reconstruction")
+        self.gridLayout_2 = QtWidgets.QGridLayout(self.group_colors_reconstruction)
+        self.gridLayout_2.setObjectName("gridLayout_2")
+        self.cb_purple = QtWidgets.QCheckBox(self.group_colors_reconstruction)
+        self.cb_purple.setObjectName("cb_purple")
+        self.gridLayout_2.addWidget(self.cb_purple, 0, 1, 1, 1)
+        self.cb_green = QtWidgets.QCheckBox(self.group_colors_reconstruction)
+        self.cb_green.setObjectName("cb_green")
+        self.gridLayout_2.addWidget(self.cb_green, 1, 1, 1, 1)
+        self.cb_red = QtWidgets.QCheckBox(self.group_colors_reconstruction)
+        self.cb_red.setObjectName("cb_red")
+        self.gridLayout_2.addWidget(self.cb_red, 1, 0, 1, 1)
+        self.cb_black = QtWidgets.QCheckBox(self.group_colors_reconstruction)
+        self.cb_black.setObjectName("cb_black")
[…]
