+++ RodTracker/src/extensions/rods_reconstruction/reconstruction.py
@@ -1,0 +1,888 @@
+# Copyright (c) 2023-24 Adrian Niemann, and others
+#
+# This file is part of RodTracker.
+# RodTracker is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# RodTracker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
+
+"""**TBD**"""
+
+import logging
+import sys
+import warnings
+from typing import List
+
+import numpy as np
+import pandas as pd
+from matplotlib.figure import Figure
+from ParticleDetection.reconstruct_3D import calibrate_cameras as cc
+from ParticleDetection.reconstruct_3D import matchND
+from ParticleDetection.reconstruct_3D import visualization as vis
+from ParticleDetection.reconstruct_3D.match2D import match_frame
+from ParticleDetection.utils import data_loading as dl
+from PyQt5 import QtCore
+from scipy.spatial.transform import Rotation as R
+
+from RodTracker.ui.tabs import UtilityTab
+
+from .reconstruction_ui import Ui_reconstruction_ui
+
+_logger = logging.getLogger(__name__)
+abort_reconstruction: bool = False
+lock = QtCore.QReadWriteLock(QtCore.QReadWriteLock.NonRecursive)
+
+
+class PlotterSignals(QtCore.QObject):
+    """Helper object to provide :class:`Plotter` access to ``pyqtSignal``."""
+
+    result_plot = QtCore.pyqtSignal(Figure, name="result_plot")
+    """pyqtSignal(Figure): Transfers a result ``Figure`` for display elsewhere.
+    """
+
+    error = QtCore.pyqtSignal(tuple, name="error")
+    """pyqtSignal(tuple) : Signal for propagating errors occuring in the
+    worker's thread.\n
+    | The transferred tuple should contain the following values:
+    | [0]: Exception type
+    | [1]: Exception value
+    | [2]: Exception traceback
+
+    See Also
+    --------
+    `sys.exc_info()`_
+
+    .. _sys.exc_info():
+        https://docs.python.org/3/library/sys.html#sys.exc_info
+    """
+
+
+class Plotter(QtCore.QRunnable):
+    """Object to run the generation of 3D reconstruction and tracking plots.
+
+    This object should run the generation of evaluation plots for the 3D
+    reconstruction of the whole or a portion of a dataset. It is capable of
+    running it in a thread different from the *main* thread to not keep the
+    application responsive.
+
+    Parameters
+    ----------
+    data : DataFrame
+        Slice of the *main* rod position dataset that shall be used for
+        plotting.
+
+    Attributes
+    ----------
+    data : DataFrame
+        Slice of the *main* rod position dataset that will be used for
+        plotting.
+    signals : PlotterSignals
+        Signals that can be emitted during the running of a :class:`Plotter`
+        object. Their purpose is to report errors and (intermediate) results.
+    kwargs : dict
+        Keyword arguments that will be used by the plotting functions. The
+        currrently recognized keywords are:\n
+        ``"colors"``, ``"start_frame"``, ``"end_frame"``,
+        ``"position_scaling"``, ``"cam_ids"``, ``"calibration"``, and
+        ``"transformation"``
+    """
+
+    def __init__(self, data: pd.DataFrame, **kwargs):
[…]
