+++ RodTracker/src/extensions/rods_reconstruction/requirements.txt
@@ -1,0 +1,3 @@
+matplotlib
+ParticleDetection
+scipy
