+++ RodTracker/setup.cfg
@@ -7,4 +7,5 @@ exclude =
     build,
     dist,
     *src/RodTracker/ui/mainwindow_layout.py,
-    *detection_ui.py
+    *detection_ui.py,
+    *reconstruction_ui.py

[…]
