### 2024-06-06 a643b0a
Add possibility to require other extensions
Extensions can now require other extensions to be available during runtime.
This also implies that those are loaded before the extension itself.
WARNING: Currently there is no mitigation for circular dependencies implemented.
