+++ RodTracker/src/RodTracker/main.py
@@ -19,11 +19,13 @@ import inspect
 import logging
 import sys
 from pathlib import Path
+from typing import Dict, List
 
 import importlib_resources
 from PyQt5 import QtCore, QtGui, QtWidgets
 
 import RodTracker
+from RodTracker import ExtensionState
 
 _logger = logging.getLogger(RodTracker.APPNAME)
 
@@ -71,34 +73,157 @@ def main():
     main_window = mw.RodTrackWindow()
 
     # Load extensions
+    discovered_exts: Dict[str, List[RodTracker.ExtensionState, dict, Path]] = (
+        {}
+    )
     extension_folder = Path(__file__).parent.parent / "extensions"
     if hasattr(sys, "_MEIPASS"):
         extension_folder = Path(__file__).parent / "extensions"
+
+    # Discover available extensions
     for entry in extension_folder.iterdir():
         if not entry.is_dir() or entry.stem == "__pycache__":
             continue
-        next_extension = entry.stem
+        ext = entry.stem
+        discovered_exts[ext] = [
+            ExtensionState.UNDEFINED,
+            {},
+            entry,
+        ]
         if list(entry.glob("DEACTIVATED")):
             # Skip loading if a file named 'DEACTIVATED' is present
-            _logger.info(f"Extension '{next_extension}' is deactived.")
+            _logger.info(f"Extension '{ext}' is deactived.")
+            discovered_exts[ext][0] = ExtensionState.DEACTIVATED
             continue
-        splash.showMessage(
-            f"Loading Extension: {next_extension}", align, color
-        )
+
+    def _try_loading(
+        ext: str, discovered_exts, circ_prevention_list: List[str]
+    ) -> ExtensionState:
+        nonlocal splash, align, color, main_window
+        if discovered_exts[ext][0] not in [
+            ExtensionState.UNDEFINED,
+            ExtensionState.DELAYED_LOADING,
+        ]:
+            return discovered_exts[ext][0]
         try:
             spec = importlib.util.spec_from_file_location(
-                next_extension, entry / "__init__.py"
+                ext, discovered_exts[ext][2] / "__init__.py"
             )
             module = importlib.util.module_from_spec(spec)
-            sys.modules[next_extension] = module
+            sys.modules[ext] = module
             spec.loader.exec_module(module)
+
+            # check extension dependencies
+            break_loading = False
+            for dep in module.REQUIRED_EXTENSIONS:
+                if dep not in discovered_exts.keys():
+                    break_loading = True
+                    discovered_exts[ext][1][dep] = ExtensionState.UNAVAILABLE
+                    discovered_exts[ext][0] = ExtensionState.MISSING_DEPENDENCY
+                    _logger.warning(
+                        f"Extension {ext} cannot be loaded due to missing "
+                        f"dependency {dep}."
+                    )
+                    continue
+
+                dep_state = discovered_exts[dep][0]
+                discovered_exts[ext][1][dep] = dep_state
+                if dep_state is ExtensionState.ACTIVE:
+                    # Dependency is fulfilled
+                    pass
+                elif dep_state is ExtensionState.DELAYED_LOADING:
+                    # WARNING: potential for circular dependency
+                    discovered_exts[ext][0] = ExtensionState.DELAYED_LOADING
+                    # TODO: check for circular dependency
+                    result = _try_loading(
+                        dep, discovered_exts, [*circ_prevention_list, ext]
+                    )
+                    if result is not ExtensionState.ACTIVE:
+                        break_loading = True
+                        discovered_exts[ext][
+                            0
+                        ] = ExtensionState.MISSING_DEPENDENCY
+                        _logger.warning(
+                            f"Extension {ext} cannot be loaded due to a "
[…]
