+++ RodTracker/src/extensions/rods_3D/__init__.py
@@ -22,6 +22,9 @@ from RodTracker.ui.mainwindow import RodTrackWindow
 from RodTracker.ui.settings import IntSetting
 
 _logger = logging.getLogger(__name__)
+REQUIRED_EXTENSIONS = [
+    "rods",
+]
 
 
 def setup(
