+++ RodTracker/src/extensions/detection/__init__.py
@@ -22,6 +22,9 @@ from PyQt5 import QtWidgets
 from RodTracker.ui.mainwindow import RodTrackWindow
 
 _logger = logging.getLogger(__name__)
+REQUIRED_EXTENSIONS = [
+    "rods",
+]
 
 
 def setup(
