### 2024-06-06 21cd15f
Add circular dependency detection to extension loading
