### 2024-06-06 8cece38
Add popup for extensions failed loading
