+++ RodTracker/src/extensions/detection/detection_ui.py
@@ -28,6 +28,13 @@ class Ui_detector_ui(object):
         self.tb_model = QtWidgets.QToolButton(detector_ui)
         self.tb_model.setObjectName("tb_model")
         self.horizontalLayout_13.addWidget(self.tb_model)
+        self.pb_use_example = QtWidgets.QPushButton(detector_ui)
+        font = QtGui.QFont()
+        font.setBold(True)
+        font.setWeight(75)
+        self.pb_use_example.setFont(font)
+        self.pb_use_example.setObjectName("pb_use_example")
+        self.horizontalLayout_13.addWidget(self.pb_use_example)
         self.verticalLayout.addLayout(self.horizontalLayout_13)
         self.horizontalLayout_12 = QtWidgets.QHBoxLayout()
         self.horizontalLayout_12.setObjectName("horizontalLayout_12")
@@ -136,13 +143,13 @@ class Ui_detector_ui(object):
         self.pb_detect.setSizePolicy(sizePolicy)
         self.pb_detect.setObjectName("pb_detect")
         self.verticalLayout.addWidget(self.pb_detect)
+        spacerItem2 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self.verticalLayout.addItem(spacerItem2)
         self.progress_detection = QtWidgets.QProgressBar(detector_ui)
         self.progress_detection.setProperty("value", 24)
         self.progress_detection.setInvertedAppearance(False)
         self.progress_detection.setObjectName("progress_detection")
         self.verticalLayout.addWidget(self.progress_detection)
-        spacerItem2 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
-        self.verticalLayout.addItem(spacerItem2)
 
         self.retranslateUi(detector_ui)
         QtCore.QMetaObject.connectSlotsByName(detector_ui)
@@ -152,6 +159,7 @@ class Ui_detector_ui(object):
         detector_ui.setWindowTitle(_translate("detector_ui", "Form"))
         self.lbl_model.setText(_translate("detector_ui", "Model: "))
         self.tb_model.setText(_translate("detector_ui", "..."))
+        self.pb_use_example.setText(_translate("detector_ui", "Use Example Model"))
         self.lbl_start_frame_detection.setText(_translate("detector_ui", "Start Frame:"))
         self.lbl_end_frame_detection.setText(_translate("detector_ui", "End Frame:"))
         self.lbl_expected_particles_default.setText(_translate("detector_ui", "Expected particles:"))
