### 2024-06-08 fc5c781
Add mechanism to install dependency libraries for extensions
