+++ RodTracker/src/extensions/detection/requirements.txt
@@ -1,4 +1,5 @@
 opencv-python-headless
 pandas
+ParticleDetection
 torch
 torchvision
