### 2024-06-12 #5986 @a-niem
Hi @jgostick. With @dmitrypuzyrev apparently to being busy, I'll attempt to answer here.
As mentioned to Kevin-Mattheus-Moerman above, I've fixed the issues opened during the review process. The changed version is available in the joss_paper branch.
I thought that now @mhubii was going to finish his review, and I'd fix any additional comments he has.
With that ongoing, I thought another reviewer had to be appointed (please let me know if I'm supposed to make suggestions again).
Fortunately, I have improved significantly in the meantime and there should be no periods of me being unavailable in the foreseeable future anymore.
I'm committed to get this process finished as soon as possible.
Please let me know about any tasks that I have overlooked that I'm supposed to finish here.
