### 2024-06-14 c8077e9
Update RodTracker's bundling
Adds installed extensions to the bundled application.
