### 2024-06-14 6ff60ee
Move PyInstaller dependency collection responsibility
Responsibility to ensure libraries are included during the bundling with PyInstaller is now moved to extensions.
These now should provide hooks for PyInstaller inside a __pyinstaller directory.
Collections performed in 'build_onedir.spec' have been moved to said hooks.
Bundling process now also uses RodTracker.APPNAME for file naming on all platforms.
