+++ RodTracker/src/extensions/rods_reconstruction/__pyinstaller/hook-extensions.rods_reconstruction.py
@@ -1,0 +1,28 @@
+# Copyright (c) 2023-24 Adrian Niemann, and others
+#
+# This file is part of RodTracker.
+# RodTracker is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# RodTracker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
+
+import PyInstaller.utils.hooks as hookutils
+
+datas = []
+binaries = []
+
+datas += hookutils.collect_data_files(
+    "extensions.rods_reconstruction",
+    include_py_files=True,
+    excludes=["__pyinstaller"],
+)
+datas += hookutils.collect_data_files("pulp", include_py_files=True)
+binaries += hookutils.collect_dynamic_libs("pulp")
