+++ RodTracker/src/extensions/rods_3D/__pyinstaller/hook-extensions.rods_3D.py
@@ -1,0 +1,21 @@
+# Copyright (c) 2023-24 Adrian Niemann, and others
+#
+# This file is part of RodTracker.
+# RodTracker is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# RodTracker is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
+
+import PyInstaller.utils.hooks as hookutils
+
+datas = hookutils.collect_data_files(
+    "extensions.rods_3D", include_py_files=True, excludes=["__pyinstaller"]
+)
