### 2024-06-19 #5986 @dmitrypuzyrev
Dear @jgostick,

I confirm that I have been busy last week and could not reply.
With @a-niem, we were expecting the second reviewer to be reappointed as @Kevin-Mattheus-Moerman mentioned and also waiting for further comments from @mhubii.
I am sorry that the review is taking so long, mainly due to temporary absense of the main developer.
However, @a-niem has already fixed multiple issues, the package is actively used at the moment and a paper was recently published in npj Microgravity Journal based on the result which were obtained with its help.
I believe that now when the main developer @a-niem is available, the review can be finished smoothly.
Please let us know what is required from our side.

Sincerely yours
Dr. Dmitry Puzyrev
