### 2024-06-20 #5986 @a-niem
Hi @mhubii. I'm sorry that you feel this way, and I'd like to apologize for any misunderstanding that I caused here. I thought I made it clear, that my absence was the reason for the delay.
Furthermore, I'd like you to know that I value your comments/issues you posted, especially as I had mentioned some of those months/years ago in the development process, but that got dismissed by my colleagues as 'unnecessary'.
I'll be looking through the list of potential reviewers and make suggestions here as asap so that the time you've sunken in this already does not go to waste in the end.
For any further comments, I'll try to be more mindful with my wording to not have any unwanted passive-aggressive undertone, as I believe was the case in my last comment.
