### 2024-06-26 #5986 @jgostick
Hi @merrygoat, when you are ready to begin your review you just create your own personal checklist with the command `editorialbot generate my checklist`, but put and `@` symbol before `editorialbot`.  (I didn't want to accidentally trigger a checklist)
