### 2024-06-28 ANP-Granular/ParticleTracking#92 @merrygoat
More than one error dialog can be shown at once
**Describe the bug**
An unhandled exception opens an error dialog. Because the dialog is not modal, you can continue interacting with the GUI and repeat the same operation causing a cascade of error dialog boxes to appear.

**Expected behaviour**
If the exception is non fatal and it is not required to inform the user it should be logged as a warning not with a dialog.

If it is required to inform the user then the error dialog should be modal to prevent further interaction with the GUI.
