### 2024-07-01 a2f8df5
Fix numpy version mismatch
Fix version mismatch of numpy with torchvision. Fixes #90.
