### 2024-07-01 af3ad9f
Skip GUI interaction tests with GitHub Actions
Skip any tests requiring GUI interactions when run with GitHub Actions because they don't run properly.
With this the automated testing workflow shall be enabled again to get early hints on issues like #90.
