+++ RodTracker/tests/test_userinteraction/test_usecases.py
@@ -15,6 +15,7 @@
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
 """Tests for typical usecases and cases where problems occurred in the past."""
+import os
 import pathlib
 from typing import List
 
@@ -28,7 +29,13 @@ from pytestqt.qtbot import QtBot
 import RodTracker.backend.logger as lg
 from RodTracker.ui.mainwindow import RodTrackWindow
 
-# pytestmark = pytest.mark.skip("Implementation changes needed.")
+if os.getenv("GITHUB_ACTIONS"):
+    pytestmark = pytest.mark.skip(
+        "Tests requiring interaction with GUI elements cannot be run properly "
+        "with GitHub Actions at the moment. Have a look this issue for "
+        "more information: "
+        "https://github.com/ANP-Granular/ParticleTracking/issues/87"
+    )
 
 
 def teardown_replacements(mp: MonkeyPatch):
