### 2024-07-01 fe6069c
Deactivate minimum test coverage
Temporarily deactivate minimum test coverage constraint until issue #87 is resolved.
