### 2024-07-03 #5986 @jgostick
Hi @dmitrypuzyrev
This submission is called ParticleTracking, yet the github repo has a downloadable GUI called RodTracker, while the ParticleTracker seems to be a python package that is not a gui.  Can we find a way to unify these names?
