### 2024-07-03 #5986 @jgostick
Also, I am trying to use this GUI with the default images and example files, but keep getting error messages.  I think @merrygoat is having the same issue.  I'm on windows installed the exe from the github releases page.
