### 2024-07-04 #5986 @dmitrypuzyrev
Dear @jgostick ,
I've just made a fresh install from .exe on Windows 11 (release 0.6.2 from releases page) and have no problems with example data.
I think we require more information, @merrygoat seemed to have a different problem with installation with `pip`.
Could you please submit an issue with the log file?

Also, we have a discussion with @a-niem today about the naming and other issues and then will get back to fixing it.
