### 2024-07-04 #5986 @a-niem
Hi @jgostick

> Hi @dmitrypuzyrev This submission is called ParticleTracking, yet the github repo has a downloadable GUI called RodTracker, while the ParticleTracker seems to be a python package that is not a gui. Can we find a way to unify these names?

I've just talked with @dmitrypuzyrev, and we were not 100% sure what you want us to rename. The simplest thing we thought of was to just rename the `RodTracker` GUI. The resulting names would be as shown below:

Repo-Name: `ParticleTracking`
Submission-Name: `ParticleTracking`
GUI-Name: `RodTracker` → `ParticleTracker` or `3D-ParticleTracker`
Library-Name: `ParticleDetection`

On the other hand, we thought that you might want us to clear up any naming conflicts with the [`ParticleTracker` software](https://github.com/MikeSmithLabTeam/particletracker) published in JOSS in 2021.

Please let me know if the proposed name changes are sufficient.

---

> Also, I am trying to use this GUI with the default images and example files, but keep getting error messages. I think @merrygoat is having the same issue. I'm on windows installed the exe from the github releases page.

It looks like the errors you are experiencing have been posted as bugs by @merrygoat already (see [#91](https://github.com/ANP-Granular/ParticleTracking/issues/91) and [#92](https://github.com/ANP-Granular/ParticleTracking/issues/92)). I am currently resolving them.
