### 2024-07-04 86dfc55
Add dialog before downloading example model
Adds a dialog explaining which model will be downloaded, where it will be saved and how large it is.
The user can then also decline the download.
Fixes #89.
