### 2024-07-05 117b13c
Fix numpy version mismatch
Fix version mismatch of numpy with torchvision. Fixes #90.
