+++ ParticleDetection/pyproject.toml
@@ -37,7 +37,7 @@ matplotlib = [
 ]
 networkx = ">=3.1"
 numpy = [
-    {version = ">=1.23.4", python = ">=3.9"},
+    {version = ">=1.23.4,<2", python = ">=3.9"},
     {version = ">=1.21,<=1.25", python = "<3.9"}
 ]
 opencv-python-headless = ">=4.8.1.78"
