### 2024-07-05 495743c
Avoid vanishing settings- and log-files during testing
