### 2024-07-05 06970c3
Make ErrorDialog modal
Change the dialog reporting errors to the user modal to avoid further interaction with the GUI prior to acknowledgement of the error.
