### 2024-07-05 923af87
Remove minimum test coverage for ParticleDetection
