### 2024-07-05 ANP-Granular/ParticleTracking#92 @a-niem
The error dialogue has been switched to be modal  (see  06970c3). It now prevents the user from repeating the same operation without having to acknowledge the error prior to that.

There is however still the possibility that multiple of these dialogues are shown. This might occur when an error occurs in different threads, e.g. during plotting of data for multiple particle classes for the `Reconstruction` tab.
