### 2024-07-09 ca301ed
Change file selection dialogs
Image file selection has been changed to selecting a directory instead of single file within a directory. The position data selection has been changed to show eligible during the folder selection.
See #88.
