### 2024-07-10 596af65
Avoid loss of dimension with only one rod
Avoid a loss of dimension during extraction of 3D rod representation with only one rod in the dataset by explicitly defining the axis of squeezing.
Fixes #91.
