### 2024-07-10 693f756
Fix folder selection for image and data opening
Fixes folder selection to allow correct data loading when the user did not navigate into the target folder but just selects it.
