### 2024-07-10 62bdb4e
Refactor folder opening
Refactor folder opening to make it testable again.
