### 2024-07-10 0309342
Refactor tests using image/data loading
Refactors the tests impacted by the changes in image & data loading.
