+++ RodTracker/tests/test_userinteraction/test_usecases.py
@@ -21,6 +21,7 @@ from typing import List
 
 import conftest
 import gui_actions as ga
+import importlib_resources
 import pytest
 from PyQt5 import QtCore, QtWidgets
 from pytest import MonkeyPatch
@@ -62,9 +63,17 @@ def test_typical(
     try:
         scenario = [
             ga.OpenData(conftest.csv_data),
-            ga.OpenImage(conftest.cam1_img1),
+            ga.OpenImage(
+                importlib_resources.files(
+                    "RodTracker.resources.example_data.images"
+                ).joinpath("gp3")
+            ),
             ga.SwitchCamera(),
-            ga.OpenImage(conftest.cam2_img1),
+            ga.OpenImage(
+                importlib_resources.files(
+                    "RodTracker.resources.example_data.images"
+                ).joinpath("gp4")
+            ),
             ga.SwitchRodNumber(12, 7, lg.NumberChangeActions.ALL_ONE_CAM),
             ga.SwitchCamera(),
             ga.SwitchRodNumber(7, 12, lg.NumberChangeActions.CURRENT),
