### 2024-07-10 4678a8a
Adjust tests to new example data scaling
