### 2024-07-11 40df4e7
Add mode to display all rods in a frame
Adds a display mode to show all available rods for a frame. In this mode the rods are read-only.
See #94.
