### 2024-07-11 9f700dd
Update tests
Update tests to changes made for the added display mode.
