### 2024-07-11 7d4cf2b
Use color names as display color
In the display mode showing all available rods, it is now attempted to use the color given by the rod class.
If matplotlib does not support a color name, a different color is chosen.
