### 2024-07-20 2539940
Add ability to delete data from the Particles tab
Adds the ability to delete data by selecting it in
the Particles tab and hitting the 'Delete' key.
Currently this action is not revertable.
See #93.
