### 2024-07-20 1eb0404
Make data deletions revertable
