### 2024-07-21 efdbe9e
Updated software paper file
Updated affiliations and included recently published research article which is using this software
