+++ paper/paper.bib
@@ -172,23 +172,24 @@
 }
 
 @article{Puzyrev2020,
-author={Puzyrev, Dmitry
-and Harth, Kirsten
-and Trittel, Torsten
-and Stannarius, Ralf},
-title={Machine Learning for 3D Particle Tracking in Granular Gases},
-journal={Microgravity Science and Technology},
-year={2020},
-month={Oct},
-day={01},
-volume={32},
-number={5},
-pages={897-906},
-abstract={Dilute ensembles of granular matter (so-called granular gases) are nonlinear systems which exhibit fascinating dynamical behavior far from equilibrium, including non-Gaussian distributions of velocities and rotational velocities, clustering, and violation of energy equipartition. In order to understand their dynamic properties, microgravity experiments were performed in suborbital flights and drop tower experiments. Up to now, the experimental images were evaluated mostly manually. Here, we introduce an approach for automatic 3D tracking of positions and orientations of rod-like particles in a dilute ensemble, based on two-view video data analysis. A two-dimensional (2D) localization of particles is performed using a Mask R-CNN neural network trained on a custom data set. The problem of 3D matching of the particles is solved by minimization of the total reprojection error, and finally, particle trajectories are tracked so that ensemble statistics are extracted. Depending on the required accuracy, the software can work fully self-sustainingly or serve as a base for subsequent manual corrections. The approach can be extended to other 3D and 2D particle tracking problems.},
-issn={1875-0494},
-doi={10.1007/s12217-020-09800-4},
-url={https://doi.org/10.1007/s12217-020-09800-4}
+  author={Puzyrev, Dmitry
+  and Harth, Kirsten
+  and Trittel, Torsten
+  and Stannarius, Ralf},
+  title={Machine Learning for 3D Particle Tracking in Granular Gases},
+  journal={Microgravity Science and Technology},
+  year={2020},
+  month={Oct},
+  day={01},
+  volume={32},
+  number={5},
+  pages={897-906},
+  abstract={Dilute ensembles of granular matter (so-called granular gases) are nonlinear systems which exhibit fascinating dynamical behavior far from equilibrium, including non-Gaussian distributions of velocities and rotational velocities, clustering, and violation of energy equipartition. In order to understand their dynamic properties, microgravity experiments were performed in suborbital flights and drop tower experiments. Up to now, the experimental images were evaluated mostly manually. Here, we introduce an approach for automatic 3D tracking of positions and orientations of rod-like particles in a dilute ensemble, based on two-view video data analysis. A two-dimensional (2D) localization of particles is performed using a Mask R-CNN neural network trained on a custom data set. The problem of 3D matching of the particles is solved by minimization of the total reprojection error, and finally, particle trajectories are tracked so that ensemble statistics are extracted. Depending on the required accuracy, the software can work fully self-sustainingly or serve as a base for subsequent manual corrections. The approach can be extended to other 3D and 2D particle tracking problems.},
+  issn={1875-0494},
+  doi={10.1007/s12217-020-09800-4},
+  url={https://doi.org/10.1007/s12217-020-09800-4}
 }
+
 @article{Pong2021,
   doi = {10.1051/epjconf/202124904003},
   url = {https://doi.org/10.1051/epjconf/202124904003},
@@ -200,4 +201,23 @@ url={https://doi.org/10.1007/s12217-020-09800-4}
   editor = {M.A. Aguirre and S. Luding and L.A. Pugnaloni and R. Soto},
   title = {Continuously heated granular gas of elongated particles},
   journal = {{EPJ} Web of Conferences}
+}
+
+@article{Puzyrev2024,
+  author={Puzyrev, Dmitry
+  and Trittel, Torsten
+  and Harth, Kirsten
+  and Stannarius, Ralf},
[…]
