### 2024-07-21 7a95563
Update paper.md
Additional update in acknowledgements
