### 2024-07-21 ANP-Granular/ParticleTracking#91 @merrygoat
Ah - that is what I did wrong. Even though 1 is probably not a sensible input value, that seems to be resolved now.
