### 2024-07-23 0273b46
Fix post-detection particle display
Fixes particles not being displayable post-detection.
See #94.
