### 2024-07-24 4b22975
Extend logic of Action reversal
Extends the logic of Action reversal, such that saving actions
are skipped and previous actions can be undone while
irreversable actions act as blockers.
