+++ RodTracker/src/RodTracker/ui/mainwindow_layout.py
@@ -14,7 +14,7 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
         MainWindow.setObjectName("MainWindow")
-        MainWindow.resize(1190, 607)
+        MainWindow.resize(1194, 608)
         icon = QtGui.QIcon()
         icon.addPixmap(QtGui.QPixmap("resources/icon_windows.ico"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
         MainWindow.setWindowIcon(icon)
@@ -30,7 +30,7 @@ class Ui_MainWindow(object):
         self.scrollArea.setWidgetResizable(True)
         self.scrollArea.setObjectName("scrollArea")
         self.scrollAreaWidgetContents_all = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_all.setGeometry(QtCore.QRect(0, 0, 1173, 631))
+        self.scrollAreaWidgetContents_all.setGeometry(QtCore.QRect(0, 0, 1194, 587))
         self.scrollAreaWidgetContents_all.setObjectName("scrollAreaWidgetContents_all")
         self.verticalLayout = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_all)
         self.verticalLayout.setContentsMargins(0, 5, 0, 0)
@@ -41,38 +41,6 @@ class Ui_MainWindow(object):
         self.hl_controlbar.setObjectName("hl_controlbar")
         self.grid_loads = QtWidgets.QGridLayout()
         self.grid_loads.setObjectName("grid_loads")
-        self.line_2 = QtWidgets.QFrame(self.scrollAreaWidgetContents_all)
-        self.line_2.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_2.setObjectName("line_2")
-        self.grid_loads.addWidget(self.line_2, 1, 0, 1, 2)
-        self.pb_load_images = QtWidgets.QPushButton(self.scrollAreaWidgetContents_all)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pb_load_images.sizePolicy().hasHeightForWidth())
-        self.pb_load_images.setSizePolicy(sizePolicy)
-        self.pb_load_images.setMaximumSize(QtCore.QSize(16777215, 16777215))
-        self.pb_load_images.setObjectName("pb_load_images")
-        self.grid_loads.addWidget(self.pb_load_images, 0, 0, 1, 1)
-        self.pb_load_rods = QtWidgets.QPushButton(self.scrollAreaWidgetContents_all)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.pb_load_rods.sizePolicy().hasHeightForWidth())
-        self.pb_load_rods.setSizePolicy(sizePolicy)
-        self.pb_load_rods.setObjectName("pb_load_rods")
-        self.grid_loads.addWidget(self.pb_load_rods, 2, 0, 1, 1)
-        self.le_image_dir = QtWidgets.QLineEdit(self.scrollAreaWidgetContents_all)
[…]
