### 2024-07-24 #5986 @a-niem
Hi everyone.
All bugs and enhancements posted by @merrygoat have been fixed/implemented. This is now available in the [`joss_paper`](https://github.com/ANP-Granular/ParticleTracking/tree/joss_paper) branch, as well as a new [release](https://github.com/ANP-Granular/ParticleTracking/releases/tag/v0.6.3) including updated installers for the `RodTracker`.

I did not change any names in this version yet, because there was no further feedback to my previous comment here. @jgostick, @merrygoat, and @mhubii could you let me know about your opinion on my suggestion there, such that I can change the naming, if necessary.
