### 2024-07-26 #5986 @jgostick
Hi Everyone. Thanks to @a-niem for addressing the comments of @merrygoat (and myself). I'm sorry I didn't reply with more input about the name change...I didn't feel right forcing you to change the name so I just kept quiet on that and left it up to you.

Anyway, to echo @mhubii's comments, we REALLY need to get the review moving. The review process at JOSS is pretty simple:  the authors need to convince the reviewers to 'tick all the checkboxes'.  As it stands now, all three reviewers have a lot of 'unticked' boxes.  I have played with the software myself and also had problems. I got the impressive that perhaps it is trying to do too much, and thus failing to deliver on it's core purpose (which TBH isn't really clear). On the other hand, the software is obviously very sophisticated and has many/most of the ingredients of a professional package in place.

So, based on the above, I am currently feeling like this submission might not ever get across the finish line. It's already dragged on so long that we've lost one reviewer, and I'm guessing we're getting perilously close to losing others.  Committing to do a peer-review is not supposed to be a months-long process after all.  So, I wonder if @mhubii and @merrygoat could please let me know how they feel about the prospects of this package being publication-ready within a 1-2 week span, either on this thread or via email if they feel more comfortable.
