### 2024-08-08 #5986 @dmitrypuzyrev
Dear @jgostick,
did you get the feedback from @mhubii and @merrygoat?
In any case, I went through the JOSS guidelines once again and I do not see objective reasons why the submission should be dropped. The Journal aims at ~3 months of the review cycle, but there should be some outliers in this case, given the sophisticated/narrowly focused nature of the software, as you mentioned. Moreover, while it does not directly adress the case of longer submissions, this paragraph (https://joss.readthedocs.io/en/latest/review_criteria.html#what-happens-if-the-software-i-m-reviewing-doesn-t-meet-the-joss-criteria) indicates that:

>  Unlike some journals we do not reject outright submissions requiring major revisions - we’re more than happy to give the author as long as they need to make these modifications/improvements.

I am once again sorry for one longer period when we did not fix the issues timely due to the heath reasons.
At the same time, while in a way we need to "convince" the reviewers to finish the review (although it is never put like this in the guidelines), as authors we have very limited power and authority to push the reviewers in this direction except for replying to their raised issues and fixing them. (Please don't take it at as criticizing the reviewers, I just indicate that the review process might not be as smooth as expected).
The current checklists of the reviewers do not seem to indicate critical problems, and it seems that at parts they were not updated with the fixes which were already provided.

I also strongly object your point that software is failing to deliver on its core purpose, since it is actively used to produce acknowledged scientific results and is currently irreplaceable for study of real 3D granular matter in microgravity.
I can understand the critique that "it is trying to do too much", since it includes some additional features which are more experimental, but this is indicated in documentation. If it is a major issue that prevents the publication, it should be indicated and can be fixed.

I would kindly request to re-evaluate and finish the review of this submission. From the next week (Aug 12), @a-niem will be available as well and we can immediately discuss and fix any issues.
