### 2024-08-09 #5986 @merrygoat
Apologies, I have been away for a couple of weeks.

My basic expectation for the package, is that I would be able to take the sample data, follow the documentation and produce some sort of simple physics - (I guess some sort of smooth curve of displacement). I would have expected it to be at this state at the time of submission. I think a lot of the issues that have been highlighted by the reviewers could have been fixed prior to submission by giving the software to some people locally who had not used it before and watching them try to use it.

I have come to look at it again this afternoon and it has improved a lot based on your recent changes. However, I have not yet been able to complete an analysis with the sample data.

I do not doubt that you have used the package to create good science, but for JOSS, the criteria are not just about science but about accessibility and usability. It is getting there, but there is still some way to go.

I have added another issue to the project repo and will add a pull request for suggested improvements to the documentation.
