### 2024-08-11 #5986 @dmitrypuzyrev
Dear @merrygoat,

I fully agree with the issue and your comment, and you are right about the displacement curve - it would roughly indicate the evolution of kinetic energy of the system which is one of the biggest research points.

Of cause, we were/are constantly using the package and observe the people using it)
The problem is that the sample data came from early development time, when the program was only intended to correct the 2D particle positions, and 3D matching and tracking was performed entirely outside of the GUI. In reality, we always worked with much longer image sequences from other experiments with smaller frame-to-frame displacements.

This is indeed a glaring problem and will be fixed ASAP.
