### 2024-08-16 e2cfc67
Add warnings to image folder selection
A warning is added to the image folder selection when the selected folder does not contain any valid image files. The loading of files does not break anymore when image files with invalid names or directories are present in the selected folder.
Fixes #99.
