### 2024-08-16 e712ecc
Update Changelog (ParticleDetection)
