+++ RodTracker/pyproject.toml
@@ -50,7 +50,7 @@ pandas= [
     {version = ">=2.1.1", python = ">=3.9"},
     {version = ">=1.2.5,<2.1", python = "<3.9"}
 ]
-ParticleDetection = ">=0.4.2.dev21"
+ParticleDetection = ">=0.4.2"
 platformdirs = ">=3.2.0"
 PyQt3D = ">=5.15.5"
 PyQt5 = ">=5.15.4"
