### 2024-08-16 bcb19f0
Fix broken tree-display functions
Fixes connections between tree-display and color selection.
Fixes insertion of previously unkown particles into the tree-display.
