### 2024-08-16 199f1bb
Update docstrings (RodTracker)
