### 2024-08-26 26ebdee
Update data before 3D-reconstruction/-tracking
Request (updated) position data immediately before attempting 3D-reconstruction/-tracking.
Fixes #100.
