+++ RodTracker/src/RodTracker/backend/reconstruction.py
@@ -793,6 +793,9 @@ class Tracker(Reconstructor):
                 df_out = pd.concat([df_out, tmp])
                 self.frames = self.frames[1:]
                 self.signals.progress.emit(1 / num_frames)
+            else:
+                # Set initial 3D data to previous frame
+                tmp=self.data[self.data.frame == self.frames[0] - 1]
 
             for i in range(len(self.frames)):
                 lock.lockForRead()
@@ -806,7 +809,7 @@ class Tracker(Reconstructor):
                 # Track particles
                 # fmt: off
                 tmp = matchND.match_frame(
-                    self.data, self.cams[0], self.cams[1], self.frames[i],
+                    self.data, tmp, self.cams[0], self.cams[1], self.frames[i],
                     self.color, self.calibration, P1, P2, rot,
                     trans, r1, r2, t1, t2
                 )[0]
