### 2024-09-09 ANP-Granular/ParticleTracking#101 @mhubii
JOSS: API
Documentation contains plenty `TBD`, is this intended?

https://github.com/ANP-Granular/ParticleTracking/blob/f06fa4b2e4da78d1b8e20e8417db8fb7910fb34d/RodTracker/src/RodTracker/backend/file_locations.py#L17

Refers to https://github.com/openjournals/joss-reviews/issues/5986
