### 2024-09-09 ANP-Granular/ParticleTracking#102 @mhubii
JOSS: GPU GUI inference and model selection
Is there the possibility to run this on GPU?

I just followed https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#automated-detection-of-particles

1. Would it be possible to add a tick box GPU? Inference takes a long time on CPU.
2. Furthermore, I don't see how to select models through the GUI, could this be achieved through the newly introduced https://github.com/ANP-Granular/ParticleTracking/issues/85

    As thus, please modify

    https://github.com/ANP-Granular/ParticleTracking/blob/f06fa4b2e4da78d1b8e20e8417db8fb7910fb34d/RodTracker/src/RodTracker/ui/detection.py#L255

    To pull via torch hub? Might simplify and generalize things to any available model...

3. Could you please just download to torch's default cache? Don't like this just downloads to `.config/RodTracker`, as this is really torch functionality

Refers to https://github.com/openjournals/joss-reviews/issues/5986
