### 2024-09-09 ANP-Granular/ParticleTracking#103 @mhubii
JOSS: Documentation
"registered command" here https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#starting-the-rodtracker is usually termed "entry point", which was confusing for me. Please advise if I am wrong..

Refers to https://github.com/openjournals/joss-reviews/issues/5986
