### 2024-09-09 ANP-Granular/ParticleTracking#104 @mhubii
JOSS: Automated detection of particles
Sorry as somewhat of a newbie here..

I ran `Detection` -> `Use Example Model` -> `Detect` with default settings. (https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#automated-detection-of-particles)

The result only detected sporadically, see below

https://github.com/user-attachments/assets/0554f618-4369-474e-8781-c7bac93bfff9

Any explanation for why the presented method does not auto-detect, as claimed? I assume with depth information and instance segmentation, this should be possible. Please advise.

Refers to https://github.com/openjournals/joss-reviews/issues/5986
