### 2024-09-09 #5986 @mhubii
okay so what leaves me a little confused is the following: You've got stereo images (right?) and very well separable rods (even by means of color). I am reviewing software, not the methodology, but how come such a sophisticated approach fails to autodetect (following https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#automated-detection-of-particles). Installation-wise, I am now happy with the adjustments made.
