### 2024-09-09 ANP-Granular/ParticleTracking#105 @mhubii
JOSS: typo
https://particletracking.readthedocs.io/en/latest/ParticleDetection/ParticleDetection.html#model-training

"of particles as model" -> "of particles a model" ?

Refers to https://github.com/openjournals/joss-reviews/issues/5986
