### 2024-09-09 ANP-Granular/ParticleTracking#104 @dmitrypuzyrev
Hello @mhubii,
can you please describe what steps you've made here exactly?
It seems that you've run 2D detection for one camera view for blue color, and then you are plotting only rod number 0.
In each frame on your video this corresponds to one of the detected objects, but it jumps from object to object because detected rods are not yet sorted (tracked).
Because tracking/view matching is performed in 3D, to get reasonable tracking results you also need to run detection for the second view, load the example stereo calibration data and then run detection on selected frame:
https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#rod-tracking-and-3d-coordinate-reconstruction
Did you perform those steps as well?
