### 2024-09-09 #5986 @dmitrypuzyrev
Dear @mhubii,
thank you for resuming the review process!)
I have replied to you in the issue thread (https://github.com/ANP-Granular/ParticleTracking/issues/104)
I think there might be some confusion between 2D detection (which sees almost all of the rods but reassigns their numbers quasi-randomly between frames) and 3D tracking process.
