### 2024-09-09 ANP-Granular/ParticleTracking#104 @dmitrypuzyrev
So if you indeed used only one camera data, this behavior is expected, because no 2D tracking was performed at all.
2D tracking would be indeed possible too, but then there would be problems with tracking rods during overlapping and complications with 3D matching too.

By the way, to test the 3D tracking, please use the latest updates on `develop` branch, there was an issue that prevented proper tracking (Issue #96) which was revently fixed. We will merge the fix to the other branches as well soon.
