### 2024-09-09 ANP-Granular/ParticleTracking#104 @mhubii
I mean the tutorial says "Automated detection of particle", so tracking is the expected outcome.

Sorry, but I won't use the `develop` branch. As per JOSS, this package is expected to work out of the box.
