### 2024-09-11 #5986 @dmitrypuzyrev
> well sure, I was doing 2D detection following the demo, but it didn't detect. Please fix. I think a good start for you would be to get [ANP-Granular/ParticleTracking#102](https://github.com/ANP-Granular/ParticleTracking/issues/102) sorted

Agreed, we first fix the technical/documentation issues and update the `JOSS` branch. We can then discuss the detection/tracking question.
