### 2024-09-19 #5986 @dmitrypuzyrev
Dear @jgostick, dear @mhubii, dear @merrygoat,
I was notified on Sunday that the main developer, @a-niem, will not be able to actively work on the project and can only provide some consultation, at least for a while. This is totally understandable, since this year he did a lot of work after his contract with the university ended.
I am focusing my time on fixing the remaining issues but it will take me some time to fully take over all the development. It's is a little complicated by the fact that we expect a baby next month with my wife) Still, the delay should not be substantial. I hope for your understanding.
