### 2024-09-19 #5986 @jgostick
Hmmm.  The following statement gives me pause:

> "...we expect a baby next month with my wife) Still, the delay should not be substantial."

These seems like contradictory facts.

Are you sure that pursuing the JOSS paper at this time is the right approach? The software still exists and can be used and cited by end-users. The only advantage of a JOSS paper is to appear as an official publication on your CV, but spending time with your family is definitely a higher priority.
