### 2024-09-20 #5986 @dmitrypuzyrev
Dear @jgostick,

you are partly right, for sure, the CV of me and even more so of my colleague is an important factor. It cannot be the only advantage of JOSS paper though, since JOSS publication dramatically improves the quality of scientific software - this we learned first-hand - and overall dissemination of research methods and results.

As to delay, I don't plan to take any more than few days holiday at the moment and still working my working hours, which I will now focus more on finishing this review process. Just needed to inform you of the current situation.

**Anyway, I suggest that I try to deal with open issues in the next 10 working days (until 04.10) and inform you on the progress done. If it is not satisfactory, we can put the review process to rest.**

If it is OK for all of you, I would also kindly ask @merrygoat to update the review checklist, if not done yet - I think it is not fully updated, or some issues have not been submitted. The checklist of @mhubii seems to be up-to-date, there are only three unchecked marks related to functionality.
