### 2024-09-20 70dcbe1
Change python to 3.12 for GitHub actions
