+++ .github/workflows/code_style.yml
@@ -16,7 +16,7 @@ jobs:
     continue-on-error: true
     strategy:
       matrix:
-        python_version: ['3.11']
+        python_version: ['3.12']
     steps:
         - uses: actions/checkout@v4
         - name: Set up Python ${{ matrix.python_version }}
