### 2024-09-24 f837798
Publish model to torch.hub
