+++ RodTracker/src/RodTracker/ui/check_hub_model.py
@@ -1,0 +1,8 @@
+import torch
+
+if __name__ == "__main__":
+
+    model = torch.hub.load(
+        "ANP-Granular/ParticleTracking:develop", "model", pretrained=True
+    )
+    print(model)
