+++ RodTracker/src/RodTracker/ui/check_hub_model.py
@@ -3,6 +3,8 @@ import torch
 if __name__ == "__main__":
 
     model = torch.hub.load(
-        "ANP-Granular/ParticleTracking:develop", "model", pretrained=True
+        "ANP-Granular/ParticleTracking:develop",
+        "rods_example_model",
+        pretrained=True,
     )
     print(model)
