### 2024-09-25 4d09cc3
Loading model in GUI from torch.hub
