### 2024-09-30 803b397
Added CUDA example model to hub, updates to documentation
