### 2024-10-01 0c0b39b
Now default CUDA model is used when GPU PyTorch version is present, plus minor fix in test_parallelism file name
