+++ RodTracker/src/RodTracker/ui/check_hub_model.py
@@ -1,10 +1,0 @@
-import torch
-
-if __name__ == "__main__":
-
-    model = torch.hub.load(
-        "ANP-Granular/ParticleTracking:develop",
-        "rods_example_model",
-        pretrained=True,
-    )
-    print(model)
