+++ RodTracker/CHANGELOG.md
@@ -1,5 +1,18 @@
 ## [Unreleased]
 
+### Added
+- missing sections in documentation were added (i.e. world coordinate transformation) 
+
+### Changed
+- default rod detection models (CPU and CUDA) are now published and downloaded via PyTorch Hub ([#102](https://github.com/ANP-Granular/ParticleTracking/issues/102))
+- default number of detected particles changed to 25 for each color (corresponding to default sample data)
+
+### Fixed
+- tracking now works as intended, before it required to press the "solve" button frame by frame ([#96](https://github.com/ANP-Granular/ParticleTracking/issues/96))
+- solved issue when changes in data were not saved and reverted after pressing "solve" button 
+- tracking/matching tests were updated according to the fixed tracking issue 
+
+
 ## [v0.6.3]
 ### Added
 - support for Python 3.12
