### 2024-10-03 1440dce
Updated documentation, requirement for ParticleDetection package
