+++ RodTracker/src/RodTracker/backend/logger.py
@@ -14,7 +14,12 @@
 # You should have received a copy of the GNU General Public License
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
-"""**TBD**"""
+"""
+Class and methods used in RodTracked GUI for manipulation with particles coordinate data and logging the resulting actions.
+
+**Author:**     Adrian Niemann (adrian.niemann@ovgu.de)\n
+**Date:**       2022-2024
+"""
 
 import logging
 import pathlib
