+++ RodTracker/src/RodTracker/backend/rod_data.py
@@ -14,7 +14,11 @@
 # You should have received a copy of the GNU General Public License
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
-""" **TBD**
+"""
+Main classes and methods used in RodTracked for manipulation with particles coordinate data.
+
+**Author:**     Adrian Niemann (adrian.niemann@ovgu.de)\n
+**Date:**       2022-2024
 
 Attributes
 ----------
@@ -1193,7 +1197,7 @@ class RodData(QtCore.QObject):
             classes are deleted if ``None`` is given.
             By default ``None``.
         particle : Union[int, None], optional
-            TBD
+            TBD (meant for deletion of particular particles - not used at the moment)
             By default ``None``.
         all : bool, optional
             Flag whether to delete all currently loaded data.
