+++ RodTracker/src/RodTracker/ui/rodnumberwidget.py
@@ -14,7 +14,13 @@
 # You should have received a copy of the GNU General Public License
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
-"""**TBD**"""
+"""
+Includes classes which define the widget used to add/manipulate the particle numbers
+to the image overlay in RodTracker GUI.
+
+**Author:**     Adrian Niemann (adrian.niemann@ovgu.de)\n
+**Date:**       2022-2024
+"""
 
 from enum import Enum
 from typing import List
