+++ RodTracker/src/RodTracker/ui/rodnumberwidget.py
@@ -15,8 +15,8 @@
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
 """
-Includes classes which define the widget used to add/manipulate the particle numbers
-to the image overlay in RodTracker GUI.
+Includes classes which define the widget used to add/manipulate
+the particle numbers to the image overlay in RodTracker GUI.
 
 **Author:**     Adrian Niemann (adrian.niemann@ovgu.de)\n
 **Date:**       2022-2024
