+++ RodTracker/src/RodTracker/backend/logger.py
@@ -15,7 +15,8 @@
 # along with RodTracker. If not, see <http://www.gnu.org/licenses/>.
 
 """
-Class and methods used in RodTracked GUI for manipulation with particles coordinate data and logging the resulting actions.
+Class and methods used in RodTracked GUI for manipulation with particles
+coordinate data and logging the resulting actions.
 
 **Author:**     Adrian Niemann (adrian.niemann@ovgu.de)\n
 **Date:**       2022-2024
