+++ RodTracker/CHANGELOG.md
@@ -1,4 +1,4 @@
-## [Unreleased]
+## [v0.6.4]
 
 ### Added
 - missing sections in documentation were added (i.e. world coordinate transformation)
@@ -8,7 +8,7 @@
 - default rod detection models (CPU and CUDA) are now published and downloaded via PyTorch Hub ([#102](https://github.com/ANP-Granular/ParticleTracking/issues/102))
 - default number of detected particles changed to 25 for each color (corresponding to default sample data)
 - ParticleDetection dependency updated (ParticleDetection v0.4.3 now required)
-- reverted to original dialogue window for stereo calibration/world transformation (for consistency with image loading)
+- reverted to original dialogue window for stereo calibration/world transformation (for consistency with image loading dialogue)
 
 ### Fixed
 - tracking now works as intended, previously it required to press the "solve" button frame by frame ([#96](https://github.com/ANP-Granular/ParticleTracking/issues/96))
