### 2024-10-04 ANP-Granular/ParticleTracking#103 @dmitrypuzyrev
Dear @mhubii,
I am not really from programming background myself, but I have never seen term "entry point" used for a console command. As I understand, "entry point" usually refers to the main function in the source code (which is indeed called by entering the registered console command).
I found that ENTRYPOINT might mean something a little different in Docker containers, which I am not familiar with. Maybe this is what you refer to?
