### 2024-10-04 ANP-Granular/ParticleTracking#102 @dmitrypuzyrev
Hi @mhubii,
this issue is adressed in the new release
https://github.com/ANP-Granular/ParticleTracking/releases/tag/v0.6.4
I could not realize the 'tick' solution yet mainly because of the issues with automatic PyTorch/CUDA installation on different systems.
But now, depending on the installed PyTorch version (GPU or CPU), the program downloads or loads from cashe the corresponding (faster) GPU or (slower) CPU model.
By default, on Linux the CUDA version is installed. On Windows, enabling the GPU version is explained here:
https://particletracking.readthedocs.io/en/latest/installation/particledetection.html#behavior-on-windows
Alternatively, the installer for Windows GPU version of the program is attached to the release. It is considerably larger (because it includes GPU PyTorch) and requires the CUDA 12.4 to be installed beforehand.

The models are now published and downloaded via PyTorch hub, as suggested.

I did not experiment yet with possibility to choose other existing models published on PyTorch hub, at the moment I don't know  how it will work (considering that the models can have different architecture).
