### 2024-10-04 ANP-Granular/ParticleTracking#104 @dmitrypuzyrev
Hi @mhubii,
now the default number of detected particles is increased to 25 for each color (according to the experiment shown on example images).
Most of the particles are automatically detected on the frames.
Please check that it works as expected for you!
