### 2024-10-04 #5986 @dmitrypuzyrev
Dear @jgostick, dear @mhubii, dear @merrygoat,
The new release of the package has been prepared:
https://github.com/ANP-Granular/ParticleTracking/releases/tag/v0.6.4

The existing issues are adressed, I also commented on all of them sepately.
Please check out the results.

I also kindly ask @merrygoat to resume the review and update the checklist, there was no feedback from him recently.
Thank you in advance!
