### 2024-10-04 f61ac87
Updated upload-artifact version for joss paper pdf
