### 2025-01-06 #5986 @jgostick
Hi All, happy new year to everyone. It's just a day on the calendar, but somehow feels like a new start.

I will start by apologizing for my lack of engagement on this submission over the last couple of months. I was absolutely slammed by my teaching load this fall, and simply could not find the time/energy to move this submission along.  Anyway, new year, new teaching schedule, and renewed energy.

So, firstly, @mhubii, did you manage to find the time to play around with this package again over the break?

@merrygoat, how about you?  The authors have been busy working on this over the fall, so hopefully you can take a quick look to see if it meets all your standards?

I know this submission is probably the longest running in JOSSs history, but I am feeling hopeful that it's almost over.

PS - If either @mhubii or @merrygoat are no longer able to commit their time to this review, please let me know so I can search for replacements.

Thanks,
Jeff
