### 2025-01-13 ANP-Granular/ParticleTracking#103 @mhubii
hi @dmitrypuzyrev. Please refer here: https://packaging.python.org/en/latest/specifications/entry-points/

think this is quite an easy fix
