### 2025-01-13 ANP-Granular/ParticleTracking#106 @mhubii
JOSS: Version?
Could it be that this version is invalid

https://github.com/ANP-Granular/ParticleTracking/blob/5412b2539d6c435459c43f82adf56d78fb184c96/RodTracker/pyproject.toml#L9

Same here

https://github.com/ANP-Granular/ParticleTracking/blob/5412b2539d6c435459c43f82adf56d78fb184c96/ParticleDetection/pyproject.toml#L7

Refers to https://github.com/openjournals/joss-reviews/issues/5986
