### 2025-01-13 ANP-Granular/ParticleTracking#107 @mhubii
JOSS: No valid image data
`Load Images` fails to load data?

![Screenshot from 2025-01-13 11-42-10](https://github.com/user-attachments/assets/14d07450-4855-43a4-b3cd-314da9e04a77)

Maybe this could be documented better?

It appears `example_data/images/gp3` works

Refers to https://github.com/openjournals/joss-reviews/issues/5986
