### 2025-01-13 ANP-Granular/ParticleTracking#108 @mhubii
JOSS: Batched detection
When running `Detection`  -> `Detect`, would it be possible to support batched inference? Detection each image individually seems rather slow. Let me know

Refers https://github.com/openjournals/joss-reviews/issues/5986
