### 2025-01-13 ANP-Granular/ParticleTracking#109 @mhubii
JOSS: 3D reconstruction
I have successfully detected all particles. Then imported:

- Camera Calibration: `/tmp/ParticleTracking/RodTracker/src/RodTracker/resources/example_data/calibrations/gp34.json`
- World Transformation: `/tmp/ParticleTracking/RodTracker/src/RodTracker/resources/example_data/calibrations/transformation.json`

I then hit `Solve`, but nothing happens? Could you explain a little better how to reconstruct. What am I doing wrong?

When `Update Plots`, this shows up

![Screenshot from 2025-01-13 12-18-00](https://github.com/user-attachments/assets/01b5c53f-54dd-43eb-86ba-25041bf0e17a)

This seems also related to this issue: https://github.com/ANP-Granular/ParticleTracking/issues/96

Refers to https://github.com/openjournals/joss-reviews/issues/5986
