### 2025-01-13 ANP-Granular/ParticleTracking#110 @mhubii
JOSS: Fail to load rod data
I am failing to load rod data, could you please advise?

![Screenshot from 2025-01-13 12-12-33](https://github.com/user-attachments/assets/9a434228-eb76-4318-8b4e-56947921422b)

Refers to https://github.com/openjournals/joss-reviews/issues/5986
