### 2025-01-13 ANP-Granular/ParticleTracking#111 @mhubii
JOSS: No more than 2 digits
Under 3D-Reconstruct, I cannot type in start / end frame with more than 2 digits, is this a general issue?

![Screenshot from 2025-01-13 13-53-05](https://github.com/user-attachments/assets/2f6d50b7-57fd-467a-8c40-3ea5b950d7bf)

This issue is also apparent for all other inputs. Somehow the GUI doesn't support it. Please fix.

Refers to https://github.com/openjournals/joss-reviews/issues/5986
