### 2025-01-21 ANP-Granular/ParticleTracking#103 @dmitrypuzyrev
Hi @mhubii,
thank you for the reference!
Following it, will it be correct to write
**"Run the GUI script entry point"** instead of "Use the registered command"?
I am replacing it now in the documentation.
