### 2025-01-21 ANP-Granular/ParticleTracking#106 @dmitrypuzyrev
Hi @mhubii,
this is a requirement of the dynamic versioning plugin, please see:
https://pypi.org/project/poetry-dynamic-versioning/
