### 2025-01-21 ANP-Granular/ParticleTracking#107 @dmitrypuzyrev
Hi @mhubii,

you are correct, this is a problem in the documentation. The data loading process and data structure should be explained clearer, it will be fixed.
