### 2025-01-21 ANP-Granular/ParticleTracking#110 @dmitrypuzyrev
Hi @mhubii,

what is in the "test_Data" folder, is it in the right format?
(see https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#loading-rod-position-data)

The example data in "resources/example_data/csv" loads without problems.
