### 2025-01-25 ANP-Granular/ParticleTracking#109 @mhubii
shouldn't the 3D reconstruction work regardless of detecting all particles?
