### 2025-01-29 ANP-Granular/ParticleTracking#109 @dmitrypuzyrev
Hi @mhubii,
yes, it should work. The results won't be very accurate, but it should not give an error.

I am trying to reproduce your problem but cannot so far.
I have loaded images from `gp3` folder to Front view, then `gp4` to Top view, then ran detection for all particle colors without any changes or corrections.
I have then loaded same stereo calibration and coordinate transformation files.
Is it exactly what you do here?

For me it works as then intended if I press `solve`, both with "Tracking" checkbox marked and unmarked.

Please check the attached screenshot, the detected rod data looks identical to yours, but I cannot compare the cropped part of your image:

![Image](https://github.com/user-attachments/assets/ad3fec53-6214-45a2-bcf8-1239e975bfd9)
