### 2025-01-29 ANP-Granular/ParticleTracking#109 @dmitrypuzyrev
Hi again @mhubii,

I think I got it now, you try to do 3D reconstruction after loading images from only one camera, it is impossible in this setup)
So it is also related to issue https://github.com/ANP-Granular/ParticleTracking/issues/110.

I guess it's a problem of documentation. I will update it ASAP and also add corresponding warning messages to Solve button.
