### 2025-01-30 ANP-Granular/ParticleTracking#109 @mhubii
hm I see, yes I probably only loaded gp3. Given the segmentations and rods of known dimension, you should be able to solve pnp regardless. But I suppose this would be a separate issue
