### 2025-01-30 ANP-Granular/ParticleTracking#109 @dmitrypuzyrev
Hi @mhubii,
It is another issue indeed.
We don't think it would be realistic to have enough information for a proper 3D reconstruction from one camera view in case of rods. The main problem is that one cannot really infer how far the rod is from camera based on its visible length, since it depends on the unknown angle between the line from the camera to the rod and the rod main axis; and the visible width of the rods do not change significantly enough to help to infer this distance.
Single view-based reconstructions were used in some experiments with the spherical particles, I guess in case of thicker cylindrical objects it might be possible to a degree, but not in case of relatively thin rods, unfortunately.
