### 2025-01-30 84dbbab
Added warning for data loading for a single camera, fixes to documentation
