### 2025-01-30 ANP-Granular/ParticleTracking#109 @mhubii
I mean you've got a rod of known length so classifying 4 points on it would allow you to solve PnP (I might be wrong, but am somewhat confident this should work). Please refer e.g. to OpenCV https://docs.opencv.org/4.x/d5/d1f/calib3d_solvePnP.html
