### 2025-01-31 ANP-Granular/ParticleTracking#109 @dmitrypuzyrev
Could it be that you were initially confusing the PnP problem with its inverse problem, triangulation/reconstruction?
In our case the camera position is known, and we need to transform coordinates from camera frame to world frame, not the other way around. Generally, this problem cannot be solved with one camera view (unless one has a camera with depth information, like LIDAR), one either needs a stereo camera or moving (scanning) camera.

In some cases it can be solved with a single view. See, for example, here:
https://cvg.cit.tum.de/research/image-based_3d_reconstruction/singleviewreconstruction
Unfortunately, in case of thin rods it does not seem applicable.
