### 2025-01-31 ANP-Granular/ParticleTracking#109 @mhubii
well inverse or not, kind of the same problem. You can estimate poses simply using AprilTags or Aruco markers using PnP, no need for depth, just an object of known shape and correspondences in the image space. You can even go down the alley of differentiable rendering or what not in your case. But I think this is a little beyond.
