### 2025-02-06 ANP-Granular/ParticleTracking#107 @dmitrypuzyrev
Hi @mhubii,

I have updated the documentation with a clear path to example folder with images, see: https://particletracking.readthedocs.io/en/latest/RodTracker/RodTracker.html#loading-image-data

Other updates to documentation for dataset structure and loading process were also added.
