### 2025-02-06 ANP-Granular/ParticleTracking#109 @dmitrypuzyrev
Hi @mhubii,

you are completely right that AprilTags or Aruco markers are a very good idea. For example, it would help with long-time tracking of particles or detection of some rotational motions which are otherwise hard to access. We thought about it for possible experiments with larger particles, in this case, however, the size constrains do not allow it (the rods are actually less than 1mm thick).
