### 2025-02-06 ANP-Granular/ParticleTracking#103 @dmitrypuzyrev
Hi @mhubii,

I have changed it in documentation, please see:
https://particletracking.readthedocs.io/en/stable/RodTracker/RodTracker.html#starting-the-rodtracker
