### 2025-02-06 ANP-Granular/ParticleTracking#110 @dmitrypuzyrev
Hi @mhubii,

Please check in the new release:
now the warning message will appear when loading data from only one camera view.
Similar warnings will appear when attempting to press `Solve` or update plots (#109)

![Image](https://github.com/user-attachments/assets/8400741f-86dc-4ab4-b599-fbc2d34fe446)
