+++ RodTracker/src/RodTracker/ui/mainwindow_layout.py
@@ -16,7 +16,11 @@ class Ui_MainWindow(object):
         MainWindow.setObjectName("MainWindow")
         MainWindow.resize(1194, 608)
         icon = QtGui.QIcon()
-        icon.addPixmap(QtGui.QPixmap("resources/icon_windows.ico"), QtGui.QIcon.Normal, QtGui.QIcon.Off)
+        icon.addPixmap(
+            QtGui.QPixmap("resources/icon_windows.ico"),
+            QtGui.QIcon.Normal,
+            QtGui.QIcon.Off,
+        )
         MainWindow.setWindowIcon(icon)
         self.centralized = QtWidgets.QWidget(MainWindow)
         self.centralized.setObjectName("centralized")
@@ -30,9 +34,15 @@ class Ui_MainWindow(object):
         self.scrollArea.setWidgetResizable(True)
         self.scrollArea.setObjectName("scrollArea")
         self.scrollAreaWidgetContents_all = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_all.setGeometry(QtCore.QRect(0, 0, 1194, 587))
-        self.scrollAreaWidgetContents_all.setObjectName("scrollAreaWidgetContents_all")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_all)
+        self.scrollAreaWidgetContents_all.setGeometry(
+            QtCore.QRect(0, 0, 1194, 587)
+        )
+        self.scrollAreaWidgetContents_all.setObjectName(
+            "scrollAreaWidgetContents_all"
+        )
+        self.verticalLayout = QtWidgets.QVBoxLayout(
+            self.scrollAreaWidgetContents_all
+        )
         self.verticalLayout.setContentsMargins(0, 5, 0, 0)
         self.verticalLayout.setSpacing(0)
         self.verticalLayout.setObjectName("verticalLayout")
@@ -41,11 +51,17 @@ class Ui_MainWindow(object):
         self.hl_controlbar.setObjectName("hl_controlbar")
         self.grid_loads = QtWidgets.QGridLayout()
         self.grid_loads.setObjectName("grid_loads")
-        self.le_rod_dir = QtWidgets.QLineEdit(self.scrollAreaWidgetContents_all)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.MinimumExpanding, QtWidgets.QSizePolicy.Fixed)
+        self.le_rod_dir = QtWidgets.QLineEdit(
+            self.scrollAreaWidgetContents_all
+        )
+        sizePolicy = QtWidgets.QSizePolicy(
+            QtWidgets.QSizePolicy.MinimumExpanding, QtWidgets.QSizePolicy.Fixed
+        )
         sizePolicy.setHorizontalStretch(0)
         sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.le_rod_dir.sizePolicy().hasHeightForWidth())
+        sizePolicy.setHeightForWidth(
+            self.le_rod_dir.sizePolicy().hasHeightForWidth()
+        )
         self.le_rod_dir.setSizePolicy(sizePolicy)
         self.le_rod_dir.setMinimumSize(QtCore.QSize(100, 0))
         self.le_rod_dir.setMaximumSize(QtCore.QSize(400, 16777215))
@@ -56,48 +72,76 @@ class Ui_MainWindow(object):
         self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
         self.line_2.setObjectName("line_2")
         self.grid_loads.addWidget(self.line_2, 1, 0, 1, 2)
[…]
