### 2025-02-06 ANP-Granular/ParticleTracking#111 @dmitrypuzyrev
Hi @mhubii ,

I have checked the code and it is the intended behavior.
The inputs here are realized by `QSpinBox` widgets. Until no image data are loaded, they have default range `0..99`.
As soon as image frames are loaded, they change to a proper range `min_frame..max_frame`.
This seems alright to me, do you think it is a problem? The only possible change I can think of is to completely disable them somehow until no images are loaded, but I don't think it is a really meaningful improvement.
