### 2025-02-07 ANP-Granular/ParticleTracking#111 @mhubii
Well when I loaded data they were ranging from 500 to 520 or so. I wasn't able to select images
