### 2025-02-08 ANP-Granular/ParticleTracking#111 @dmitrypuzyrev
Hi @mhubii,
I've just checked, the fields in `3D-Reconstruct` tab change to proper limits (in the case of example data, for the range 500 to 519) as soon as detection is performed or coordinate data is loaded.
One cannot perform 3D reconstruction until it is done, so it is the same principle, before coordinate data are available, the Solve button is inactive and the `QSpinBox` fields are in their default state (as on your screenshot).
