### 2025-02-18 #5986 @jgostick
Hi @dmitrypuzyrev
Ok, this is good to hear.  Perhaps we can then wrap this up.  Since the last issue from @mhubii is a feature request, I am sure we can call that optional.  I just glanced through the issue tracker and see that have addressed the bugs but did not close the issues.  I don't mind...I will just note here that all of the linked issues above have been addressed (expect for the feature request about batch processing).
