### 2025-02-18 #5986 @dmitrypuzyrev
Sounds good! I also planned to look into possibility of this enchancement, but you're right that it is optional (in this kind of tasks the detection speed is not a bottleneck and especially with the GPU option it is more than sufficient in the current version).
