### 2025-02-23 ANP-Granular/ParticleTracking#108 @dmitrypuzyrev
I have investigated this possibility. Unfortunately it does not look easy to do, as our Detectron2 models for detection are exported via tracing (https://detectron2.readthedocs.io/en/latest/tutorials/deployment.html), which does not support dynamic batch size.

Changing this would require more work but could be one of the improvements for the future versions. At the moment, the detection speed is not a bottleneck of the 3D tracking tasks which are performed by the package.
