### 2025-02-23 #5986 @dmitrypuzyrev
Hi @jgostick,
I have looked into the possibility to implement the enchancement (batch detection), but it does not seem to be a simple task.
I've added a comment in the issue thread (https://github.com/ANP-Granular/ParticleTracking/issues/108#issuecomment-2677074269)

There have been no further response in the other issue threads which were addressed.
Do you think we can proceed (wrap it up) with the review?)
