### 2025-02-26 ANP-Granular/ParticleTracking#114 @mhubii
JOSS: Tracking continuity
So basically I managed to 3D reconstruct. I did not assign labels and used the `Tracking` checkbox under `3D-Reconstruct`. The results do appear rather inconsistent. Is this something that can be fixed @dmitrypuzyrev ?

https://github.com/user-attachments/assets/fb773971-89be-4fbb-a1cd-e99bb2a4bea7

 Refers to https://github.com/openjournals/joss-reviews/issues/5986#issuecomment-2666473776
