### 2025-02-26 #5986 @mhubii
so it appears most issues were addressed, with one minor remaining issue under https://github.com/ANP-Granular/ParticleTracking/issues/107. Using the 3D-reconstruction, however, it appears that results are rather inconsistent (https://github.com/ANP-Granular/ParticleTracking/issues/114). Whether this is operator mistake is up to you guys to decide. I haven't had time to review the `ParticleDetection` package, other than that, looks good to me now.
