### 2025-02-26 ANP-Granular/ParticleTracking#114 @dmitrypuzyrev
Hi @mhubii,

the tracking would improve as soon as you start correcting the 2D coordinates (you don't need to renumber/relabel particles at this step). It can be seen on your video that on some frames one or a few rods of each color are not detected automatically and instead assigned the position in the left upper corner of the image. When they are manually marked close to their actual positions, the tracking and 3D-reconstruction becomes much more consistent.

This is a known limitation of the current matching/tracking algorithm, unfortunately it is quite rigid. We have tried different approaches and this one worked the best so far for corrected data.
We were recently looking into more intelligent methods which can better deal with outliers and gaps in detection, but it is the work for the future.
