### 2025-03-04 #5986 @jgostick
Hi @dmitrypuzyrev
I got the software running and got pretty far in the tutorial.  There are a few things that puzzle me:

- What is the point of the pip install?  I had to eventually clone the git repo and navigate down several levels to find the `main.py` file to get the gui to run, which would be rather difficult to do if I had to navigate to the folder where uv installed the package.
- The 3D reconstruction kinda worked, but the instructions mentioned that I needed to load the images from both cameras.  I don't think I was able to do that, and the 3D view did not show me anything other than the static image.
- I loaded the saved particle locations from the csv data provided in the examples folder, but I was not clear on how I might save this data myself.  The 'save' button at the top opened a navigation window and let me select a folder, but not save a file.
-
