### 2025-03-04 #5986 @mhubii
> * What is the point of the pip install?  I had to eventually clone the git repo and navigate down several levels to find the `main.py` file to get the gui to run, which would be rather difficult to do if I had to navigate to the folder where uv installed the package.

This relates to https://github.com/ANP-Granular/ParticleTracking/issues/103. There is an entry point for running it.

> * The 3D reconstruction kinda worked, but the instructions mentioned that I needed to load the images from both cameras.  I don't think I was able to do that, and the 3D view did not show me anything other than the static image.

This relates to https://github.com/ANP-Granular/ParticleTracking/issues/114. Tracking should work little more out of the box imho.
