+++ paper/paper.bib
@@ -221,3 +221,108 @@
   doi={10.1038/s41526-024-00369-5},
   url={https://doi.org/10.1038/s41526-024-00369-5}
 }
+
+@article{Schindelin2012,
+  author={Schindelin, Johannes
+  and Arganda-Carreras, Ignacio
+  and Frise, Erwin
+  and Kaynig, Verena
+  and Longair, Mark
+  and Pietzsch, Tobias
+  and Preibisch, Stephan
+  and Rueden, Curtis
+  and Saalfeld, Stephan
+  and Schmid, Benjamin
+  and Tinevez, Jean-Yves
+  and White, Daniel James
+  and Hartenstein, Volker
+  and Eliceiri, Kevin
+  and Tomancak, Pavel
+  and Cardona, Albert},
+  title={Fiji: an open-source platform for biological-image analysis},
+  journal={Nature Methods},
+  year={2012},
+  month={Jul},
+  day={01},
+  volume={9},
+  number={7},
+  pages={676-682},
+  abstract={Presented is an overview of the image-analysis software platform Fiji, a distribution of ImageJ that updates the underlying ImageJ architecture and adds modern software design elements to expand the capabilities of the platform and facilitate collaboration between biologists and computer scientists.},
+  issn={1548-7105},
+  doi={10.1038/nmeth.2019},
+  url={https://doi.org/10.1038/nmeth.2019}
+}
+
+@article{Ershov2022,
+  author={Ershov, Dmitry
+  and Phan, Minh-Son
+  and Pylv{\"a}n{\"a}inen, Joanna W.
+  and Rigaud, St{\'e}phane U.
+  and Le Blanc, Laure
+  and Charles-Orszag, Arthur
+  and Conway, James R. W.
+  and Laine, Romain F.
+  and Roy, Nathan H.
+  and Bonazzi, Daria
+  and Dum{\'e}nil, Guillaume
+  and Jacquemet, Guillaume
+  and Tinevez, Jean-Yves},
+  title={TrackMate 7: integrating state-of-the-art segmentation algorithms into tracking pipelines},
+  journal={Nature Methods},
+  year={2022},
+  month={Jul},
+  day={01},
+  volume={19},
+  number={7},
+  pages={829-832},
+  abstract={TrackMate is an automated tracking software used to analyze bioimages and is distributed as a Fiji plugin. Here, we introduce a new version of TrackMate. TrackMate 7 is built to address the broad spectrum of modern challenges researchers face by integrating state-of-the-art segmentation algorithms into tracking pipelines. We illustrate qualitatively and quantitatively that these new capabilities function effectively across a wide range of bio-imaging experiments.},
+  issn={1548-7105},
+  doi={10.1038/s41592-022-01507-1},
+  url={https://doi.org/10.1038/s41592-022-01507-1}
+}
+
+@article{berg2019,
+	title = {ilastik: interactive machine learning for (bio)image analysis},
+	issn = {1548-7105},
+	url = {https://doi.org/10.1038/s41592-019-0582-9},
+	doi = {10.1038/s41592-019-0582-9},
[…]
