### 2025-03-24 #5986 @jgostick
Hi @dmitrypuzyrev, thanks for this...one little nit:  the point of listing the other software options is to highlight the need/niche that your package fills which are not handled by the other packages.  I think you're list is good, but just needs more context.  Also, probably it should go at the beginning of the "Statement of Need" section, then once you have setup why your package is "needed" you can elaborate on it's special features.

Sorry to be a pain, but once this is fixed up I promise it will be ready to move on to the acceptance phase.
