+++ paper/paper.bib
@@ -21,6 +21,7 @@
   adsurl  = {https://rdcu.be/b08Wh},
   doi     = {10.1038/s41592-019-0686-2},
 }
+
 @Article{         harris2020array,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
@@ -43,9 +44,10 @@
  publisher     = {Springer Science and Business Media {LLC}},
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
+
 @software{the_pandas_development_team_2022_7344967,
   author       = {The pandas development team},
-  title        = {pandas-dev/pandas: Pandas},
+  title        = {{pandas-dev/pandas: Pandas}},
   month        = nov,
   year         = 2022,
   note         = {If you use this software, please cite it as below.},
@@ -54,6 +56,13 @@
   doi          = {10.5281/zenodo.7344967},
   url          = {https://doi.org/10.5281/zenodo.7344967}
 }
+
+@software{pulp
+  author       = {Roy, J.S and Mitchell, Stuart A and Duquesne, Christophe-Marie and Peschiera, Franco and Phillips, Antony},
+  title        = {{PuLP: A python Linear Programming API}}      
+  url          = {https://coin-or.github.io/pulp/}
+}
+
 @misc{wu2019detectron2,
   author =       {Yuxin Wu and Alexander Kirillov and Francisco Massa and
                   Wan-Yen Lo and Ross Girshick},
@@ -154,6 +163,18 @@
   Volume = {124},
   Pages = {208007},
   doi = {10.1103/PhysRevLett.124.208007},
+}
+
+@article{Schroeter2022, 
+  author={Schroeter, Matthias and Lyv, Chen and Huang, Jiayun and Huang, Kai}, 
+  title={Challenges of ’imaging’ particulate materials in three dimensions},
+  journal={Papers in Physics},  
+  volume={14}, 
+  url={https://www.papersinphysics.org/papersinphysics/article/view/816}, 
+  DOI={10.4279/pip.140015},  journal={Papers in Physics}, 
+  year={2022}, 
+  month={Dec}, 
+  pages={140015} 
 }
 
 @Article{Pitikaris2022,
