+++ paper/paper.md
@@ -59,22 +59,28 @@ While most of 2D experiments can be performed in normal gravity, 3D experiments 
 
 In typical microgravity experiments, ensembles of particles are placed in the container, excited mechanically or magnetically and observed with a stereo-camera setup. Many experiments were performed in the VIP-Gran instrument by the Space Grains ESA Topical team (spacegrains.org) during parabolic flight campaigns. In the majority of VIP-Gran experiments, particle density does not allow for tracking individual grains.
 
-Another possibility is to perform the experiment with dilute ensembles, where most particles can be directly observed on video footage [@Harth2018; @Puzyrev2020]. In this case, the focus has been on experiments with elongated particles, due to the fact that collision rates for such particles are much higher than for spheres for the same packing fraction. Thus, even if particles overlap on the camera views, usually their endpoints still can be observed and their 3D positions and orientations can be reconstructed. In addition, study of elongated particles allows to observe the evolution of their orientations and to find the kinetic energy associated with the rotational degrees of freedom. Experiments with other particle types are planned as well.
+Another possibility is to perform the experiment with dilute ensembles, where most particles can be directly observed on video footage [@Harth2018; @Puzyrev2020]. In this case, the focus has been on experiments with elongated particles, due to the fact that collision rates for such particles are much higher than for spheres for the same packing fraction. Thus, even if particles overlap on the camera views, usually their endpoints still can be observed and their 3D positions and orientations can be reconstructed. In addition, study of elongated particles allows to observe the evolution of their orientations and to find the kinetic energy associated with the rotational degrees of freedom. Experiments with more complex particle types, such as spatial crosses (hexapods) were recently performed as well.
 
 For the study of such systems, it is beneficial to know the 3D positions and orientations over time for as many particles as possible. To achieve statistically meaningful results, the tracking of many tens to hundreds of particles is usually required. With that information, a reliable statistical analysis of the ensemble properties and their evolution over time can be achieved.
-Due to the large number of simultaneously tracked objects and their relatively high velocity, accurate experimental data analysis requires high frame rates. In one drop tower experimental run, around 9 seconds of 100 to 240 fps video footage must be analyzed. This makes manual data analysis exceptionally time-consuming. Due to the large number of overlapping particles, conventional particle detection methods based on color separation, morphological operations, and Hough transform have proven to be unstable.
-For this reason, an AI-assisted approach based on Matterport Mask R-CNN implementation [@MaskRCNN2017; @Abdulla2017] has been successfully employed [@Puzyrev2020] in extraction and processing of data from the raw stereo camera images.
+Due to the large number of simultaneously tracked objects and their relatively high velocity, accurate experimental data analysis requires high frame rates. In one drop tower experimental run, around 9 seconds of 100 to 240 fps video footage must be analyzed. This makes manual data analysis exceptionally time-consuming.
+
+An overview of challenges and methods of visual analysis of 3D granular systems is provided in [@Schroeter2022]. In case of relatively simple particles shapes, for automatic 2D detection and tracking tasks the common image processing tools can be used, including Fiji [@Schindelin2012] with TrackMate plugin [@Ershov2022] or ilastik [@berg2019].
+
+In case of more complex particle shapes, i.e. rod-like particles, more elaborate custom workflows are employed. Due to the large number of overlapping particles, conventional particle detection methods based on color separation, morphological operations, and Hough transform have proven to be unstable.
+
[…]
