### 2025-03-30 fcdcccc
bibliography fix
