+++ paper/paper.bib
@@ -57,7 +57,7 @@
   url          = {https://doi.org/10.5281/zenodo.7344967}
 }
 
-@software{pulp
+@software{pulp,
   author       = {Roy, J.S and Mitchell, Stuart A and Duquesne, Christophe-Marie and Peschiera, Franco and Phillips, Antony},
   title        = {{PuLP: A python Linear Programming API}}      
   url          = {https://coin-or.github.io/pulp/}
