### 2025-03-30 837c42e
another minor bibliography fix
