+++ paper/paper.bib
@@ -59,7 +59,7 @@
 
 @software{pulp,
   author       = {Roy, J.S and Mitchell, Stuart A and Duquesne, Christophe-Marie and Peschiera, Franco and Phillips, Antony},
-  title        = {{PuLP: A python Linear Programming API}}      
+  title        = {{PuLP: A python Linear Programming API}},      
   url          = {https://coin-or.github.io/pulp/}
 }
 
