### 2025-03-30 cc6904c
Fixes to pandas and pulp citations
