+++ paper/paper.bib
@@ -46,7 +46,7 @@
 }
 
 @software{the_pandas_development_team_2022_7344967,
-  author       = {The pandas development team},
+  author       = {{The pandas development team}},
   title        = {{pandas-dev/pandas: Pandas}},
   month        = nov,
   year         = 2022,
@@ -58,8 +58,11 @@
 }
 
 @software{pulp,
-  author       = {Roy, J.S and Mitchell, Stuart A and Duquesne, Christophe-Marie and Peschiera, Franco and Phillips, Antony},
-  title        = {{PuLP: A python Linear Programming API}},      
+  author       = {Roy, J.S. and Mitchell, Stuart A. and Duquesne, Christophe-Marie and Peschiera, Franco and Phillips, Antony},
+  title        = {{PuLP: A python Linear Programming API}},
+  publisher    = {{COIN-OR project}},
+  version      = {v2.8.0},
+  year         = 2024,      
   url          = {https://coin-or.github.io/pulp/}
 }
 
