### 2025-04-04 265994c
Small stylistic fix to paper
