+++ paper/paper.md
@@ -73,7 +73,7 @@ This approach still suffered from long manual data processing times, due to the 
 
 The `ParticleTracking` software is an evolution of the AI-assisted framework for the analysis of dilute granular ensembles, improved by the transition to the Detectron2 platform, inclusion of the `RodTracker` GUI, and a documented and extensible codebase. Our aim is to provide a solution which allows to perform several interconnected analysis steps, namely 2D particle detection, data labeling/correction, and 3D reconstruction/tracking in one software interface.
 
-As an alternative, individual data analysis steps can be performed with alternative methods and integrated into `ParticleTracking` workflow, either with data import/export or source code modification.
+As an alternative, individual data analysis steps can be performed with the other methods and integrated into `ParticleTracking` workflow, either with data import/export or source code modification.
 For example, experimentation with intelligent segmentation and tracking solutions such as SAM 2/ SAMURAI [@ravi2024sam2; @yang2024samurai] might provide useful results.
 Data correction and labeling can be performed in various image annotation tools, see, for example, the list here: [https://github.com/HumanSignal/awesome-data-labeling](https://github.com/HumanSignal/awesome-data-labeling). Note that for further processing in `RodTracker`, one should account for the necessity of labeling both endpoints in case of elongated particles and export the data in correct format as explained in `ParticleTracking` documentation.
 The 3D tracking task can be performed by implementing custom user scripts using MATLAB (one can start with [@Himpel2011] and following articles) or OpenCV functions (as in `ParticleDetection` package). 
