### 2025-04-14 #5986 @dmitrypuzyrev
Dear @jgostick ,
I am preparing the release version with final changes etc. to upload on Zenodo, I will also add the installers missing in the last release version.

We have also double-checked checked the authors list. I need to add one affiliation for myself in the paper (in the second half of the last year I was working in a different department in the University and now am affiliated with it), can you confirm I can do it at this stage?
