### 2025-04-14 fcb3210
temp fixes to new flake8 warnings
