### 2025-04-14 #5986 @jgostick
Yes, go ahead and change things as needed.  This is the time to tighten things up, such as affiliations etc
