### 2025-04-18 2faa6e9
changelog updated for version 0.6.7 release
