### 2025-04-22 #5986 @dmitrypuzyrev
Dear @jgostick ,
I am sorry for an additional delay. I tried to make a new Zenodo release remotely from holidays and had some technical difficulties (mostly with creation of installers). I will be back to my main PC on 29th of April and wrap up the post-release tasks in the end of the next week.
