### 2025-05-05 #5986 @dmitrypuzyrev
Dear @jgostick,
please check the post-review task list:

1. Version **v0.6.7** was released (https://github.com/ANP-Granular/ParticleTracking/releases/tag/v0.6.7), intended for JOSS publication
2. This version was published on Zenodo: https://zenodo.org/records/15341822
3. DOI for the Zenodo publication of this version: https://doi.org/10.5281/zenodo.15341822
4. The authors, their ORCIDs, affiliations and software license were checked and updated in the paper on `joss_paper` branch and in the Zenodo release
6. In the paper file on `joss_paper` branch, the `date` field was updated to the published release date (today), before it was the submission date. I am not completely sure this is correct, I could not find the corresponding info in the guidelines. Also, so far this date does not appear in the resulting paper .pdf, where "submitted/published" dates are both placeholders.

Please check it and let me know if anything else should be corrected/added/re-checked.
Thank you a lot once again!
