### 2025-05-16 #5986 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
