### 2025-05-17 #5986 @Kevin-Mattheus-Moerman
👉 The archive title should be edited to match the paper title: `ParticleTracking: A GUI and library for particle tracking on stereo camera images`. You should be able to manually edit the current archive.
